### 2023-01-12 #5070 @editorialbot
[PRE REVIEW]: ProgPy: Python Packages for Prognostics and Health Management of Engineering Systems
**Submitting author:** @teubert (<a href="http://orcid.org/0000-0001-6788-4507">Chris Teubert</a>)
**Repository:** https://github.com/nasa/prog_models
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.4.5
