### 2023-01-12 #5070 @teubert
Hello @kellyrowland: Thank you for agreeing to be the editor for our subscription. Let me know if you need anything from us.

In the meantime, can I suggest Matthew Daigle (github username: matthewjdaigle) from Novity as a reviewer? He's had a lot of experience developing software in this field, so he would be able to provide that perspective to the review process.
