### 2023-01-13 1f93613
Final edits to trajectory generation
