### 2023-01-18 caaacf8
Initial edits from PR
Editing imports for readability, other readability changes, adjusting outputs, adding __init__
