+++ src/prog_models/models/uav_model/uav_model.py
@@ -1,17 +1,15 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
-from prog_models.prognostics_model import PrognosticsModel
-import prog_models.models.uav_model.trajectory.route as route 
-import prog_models.models.uav_model.trajectory.trajectory as trajectory
-from prog_models.models.uav_model.vehicles import AircraftModels
-from prog_models.models.uav_model.utilities import geometry
-
 import numpy as np
 import datetime
 from copy import deepcopy
-import prog_models.models.uav_model.utilities.geometry as geom
 from warnings import warn
+
+from prog_models.prognostics_model import PrognosticsModel
+from .trajectory import route, trajectory
+from .vehicles import AircraftModels
+from .utilities import geometry as geom
 from prog_models.exceptions import ProgModelInputException
 
 class UAVGen(PrognosticsModel):
@@ -20,10 +18,12 @@ class UAVGen(PrognosticsModel):
     
     Model generates cartesian positions and velocities, pitch, roll, and yaw, and angular velocities throughout time to satisfy some user-define waypoints. 
 
+    See [0]_ for modeling details. 
+
     :term:`Events<event>`: (1)
-        TrajectoryComplete: percentage of waypoints remaining for trajectory to pass through
+        TrajectoryComplete: All waypoints are completed 
     
-    :term:`Inputs/Loading<input>`: ()
+    :term:`Inputs/Loading<input>`: (0)
         | User-defined inputs: waypoints plus ETAs or speeds
         | Model-defined inputs: 
             | T: thrust
@@ -47,9 +47,9 @@ class UAVGen(PrognosticsModel):
         | t: time 
 
     :term:`Outputs<output>`: (12)
-        | x: cartesian position
-        | y: cartesian position
-        | z: cartesian position
+        | x: cartesian position, pointing East, with reference frame centered on the first waypoint
+        | y: cartesian position, pointing North, with reference frame centered on the first waypoint 
+        | z: cartesian position, pointing Up, with reference frame centered on the first waypoint 
         | phi: pitch
         | theta: roll
         | psi: yaw
@@ -59,7 +59,6 @@ class UAVGen(PrognosticsModel):
         | p: angular velocity 
         | q: angular velocity 
         | r: angular velocity 
-        | t: time 
 
     Keyword Args
     ------------
@@ -92,9 +91,9 @@ class UAVGen(PrognosticsModel):
           Optional string to identify flight plan.
         aircraft_name : Optional, str
           Optional string to identify aircraft. 
-        dt : float
+        dt : Optional, float
           Time step in seconds for trajectory generation
-        gravity : float
+        gravity : Optional, float
           m/s^2, gravity magnitude
         cruise_speed : float
           m/s, avg speed in-between way-points
@@ -104,34 +103,34 @@ class UAVGen(PrognosticsModel):
           m/s, vertical speed (down)
         landing_speed : float
           m/s, landing speed when altitude < 10m
-        hovering_time : float
+        hovering_time : Optional, float
           s, time to hover between waypoints
-        takeoff_time : float
+        takeoff_time : Optional, float
           s, additional takeoff time 
-        landing_time: float
+        landing_time: Optional, float
           s, additional landing time 
-        waypoint_weights: 
+        waypoint_weights: Optional, float
           weights of the waypoints in nurbs calculation 
-        adjust_eta: dict 
+        adjust_eta: Optional, dict 
           Dictionary with keys ['hours', 'seconds'], to adjust route time
-        nurbs_basis_length: float
+        nurbs_basis_length: Optional, float
[…]
