### 2023-01-18 7f676be
Edits on uav_model.py: line 237, computing waypoints in xyz from route lat-lon-alt in one shot instead of for loop. output container generated with a for loop instead of explicitly calling all keys. However, this might need to be changed in the future when only a subset of the state will become an output. future_loading_new compressed. there's still the issue of having vehicle_model.state AND x as both states.
