+++ examples/uav_dynamics_model.py
@@ -7,6 +7,7 @@ Example of generating a trajectory for a rotorcraft through a set of coarse wayp
 from prog_models.models.uav_model.uav_model import UAVGen 
 
 import numpy as np
+import matplotlib.pyplot as plt
 
 def run_example(): 
 
@@ -15,9 +16,9 @@ def run_example():
     # Define coarse waypoints: waypoints must be defined with a dictionary of numpy arrays or as columns in a text file 
     # See documentation for specific information on inputting waypoints 
     waypoints = {}
-    waypoints['lat_deg'] = np.array([37.09776, 37.09776, 37.09776, 37.09798, 37.09748, 37.09665, 37.09703, 37.09719, 37.09719, 37.09719, 37.09719, 37.09748, 37.09798, 37.09776, 37.09776])
-    waypoints['lon_deg'] = np.array([-76.38631, -76.38629, -76.38629, -76.38589, -76.3848, -76.38569, -76.38658, -76.38628, -76.38628, -76.38628, -76.38628, -76.3848, -76.38589, -76.38629, -76.38629])
-    waypoints['alt_ft'] = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
+    waypoints['lat_deg']   = np.array([37.09776, 37.09776, 37.09776, 37.09798, 37.09748, 37.09665, 37.09703, 37.09719, 37.09719, 37.09719, 37.09719, 37.09748, 37.09798, 37.09776, 37.09776])
+    waypoints['lon_deg']   = np.array([-76.38631, -76.38629, -76.38629, -76.38589, -76.3848, -76.38569, -76.38658, -76.38628, -76.38628, -76.38628, -76.38628, -76.3848, -76.38589, -76.38629, -76.38629])
+    waypoints['alt_ft']    = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
     waypoints['time_unix'] = np.array([1544188336, 1544188358, 1544188360, 1544188377, 1544188394, 1544188411, 1544188428, 1544188496, 1544188539, 1544188584, 1544188601, 1544188635, 1544188652, 1544188672, 1544188692])
 
     # Define model parameters:
@@ -30,7 +31,8 @@ def run_example():
     # Create a model object, define noise
     uav_1 = UAVGen(**params_1)
     uav_1.parameters['process_noise'] = 0
-
+    uav_1.initialize()
+    
     # Define future loading function to return empty InputContainer, since there is no user-defined loading in trajectory generation 
     def future_loading_1(t, x=None):
         return uav_1.InputContainer({}) 
@@ -49,9 +51,9 @@ def run_example():
     uav_1.visualize_traj(traj_results_1)
     
     # Plot internal states 
-    traj_results_1.outputs.plot(keys = ['x', 'y', 'z'], ylabel = 'Cartesian position',title='Example 1 Predicted Outputs')
-    traj_results_1.outputs.plot(keys = ['phi', 'theta', 'psi'], ylabel = 'Pitch, roll, and yaw',title='Example 1 Predicted Outputs')
-    traj_results_1.outputs.plot(keys = ['vx', 'vy', 'vz'], ylabel = 'Velocities',title='Example 1 Predicted Outputs')
+    traj_results_1.outputs.plot(keys = ['x', 'y', 'z'], ylabel = 'Cartesian position', title='Example 1 Predicted Outputs')
[…]
