+++ src/prog_models/models/uav_model/uav_model.py
@@ -16,7 +16,21 @@ from prog_models.exceptions import ProgModelInputException
 
 class UAVGen(PrognosticsModel):
     """
-    Vectorized prognostics :term:`model` to generate a predicted trajectory for a UAV. 
+    Vectorized prognostics :term:`model` to generate a predicted trajectory for a UAV using a n=6 degrees-of-freedom dynamic model
+    with feedback control loop. The model follows the form:
+    
+    u     = h(x, x_{ref})
+    dx/dt = f(x, \theta, u)
+    
+    where:
+      x is a 2n state vector containing position, attitude and corresponding derivatives
+      \theta is a vector of model parameters including UAV mass, inertia moment, aerodynamic coefficients, etc
+      u is the input vector: thrust along the body vertical axis, and three moments along the UAV body axis to follow the desired trajectory.
+      x_{ref} is the desired state vector at that specific time step, with dimension 2n
+      f(.) is growth rate function of all vechile state
+      h(.) is the feedback-loop control function that returns the necessary thrust and moments (u vector) to cover the error between desired state x_{ref} and current state x
+      dx/dt is the state-increment per unit time.
+
     
     Model generates cartesian positions and velocities, pitch, roll, and yaw, and angular velocities throughout time to satisfy some user-define waypoints. 
 
@@ -32,33 +46,33 @@ class UAVGen(PrognosticsModel):
             | mz: moment in z
 
     :term:`States<state>`: (13)
-        | x: cartesian position
-        | y: cartesian position
-        | z: cartesian position
-        | phi: pitch
-        | theta: roll
-        | psi: yaw
-        | vx: velocity 
-        | vy: velocity 
-        | vz: velocity 
-        | p: angular velocity 
-        | q: angular velocity 
-        | r: angular velocity 
+        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame
+        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame 
+        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame
+        | phi: Euler's first attitude angle
+        | theta: Euler's second attitude angle
+        | psi: Euler's third attitude angle
+        | vx: velocity along x-axis, i.e., velocity along East in fixed inertia frame
+        | vy: velocity along y-axis, i.e., velocity along North in fixed inertia frame
+        | vz: velocity along z-axis, i.e., velocity Up in fixed inertia frame
+        | p: angular velocity around UAV body x-axis
+        | q: angular velocity around UAV body y-axis 
+        | r: angular velocity around UAV body z-axis 
         | t: time 
 
     :term:`Outputs<output>`: (12)
-        | x: cartesian position
-        | y: cartesian position
-        | z: cartesian position
-        | phi: pitch
-        | theta: roll
-        | psi: yaw
-        | vx: velocity 
-        | vy: velocity 
-        | vz: velocity 
-        | p: angular velocity 
-        | q: angular velocity 
-        | r: angular velocity 
+        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame
[…]
