### 2023-01-18 89594f5
fixing a couple of errors that made the new code break. now it runs.
