### 2023-01-18 99d1fd2
adding some comments to functions
