+++ src/prog_models/models/uav_model/uav_model.py
@@ -47,9 +47,9 @@ class UAVGen(PrognosticsModel):
             | mz: moment in z
 
     :term:`States<state>`: (13)
-        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame
-        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame 
-        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame
+        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame, center is at first waypoint
+        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame, center is at first waypoint
+        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame, center is at first waypoint
         | phi: Euler's first attitude angle
         | theta: Euler's second attitude angle
         | psi: Euler's third attitude angle
@@ -62,10 +62,9 @@ class UAVGen(PrognosticsModel):
         | t: time 
 
     :term:`Outputs<output>`: (12)
-<<<<<<< HEAD
-        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame
-        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame 
-        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame
+        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame, center is at first waypoint
+        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame, center is at first waypoint 
+        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame, center is at first waypoint
         | phi: Euler's first attitude angle
         | theta: Euler's second attitude angle
         | psi: Euler's third attitude angle
@@ -76,20 +75,6 @@ class UAVGen(PrognosticsModel):
         | q: angular velocity around UAV body y-axis 
         | r: angular velocity around UAV body z-axis 
         | t: time 
-=======
-        | x: cartesian position, pointing East, with reference frame centered on the first waypoint
-        | y: cartesian position, pointing North, with reference frame centered on the first waypoint 
-        | z: cartesian position, pointing Up, with reference frame centered on the first waypoint 
-        | phi: pitch
-        | theta: roll
-        | psi: yaw
-        | vx: velocity 
-        | vy: velocity 
-        | vz: velocity 
-        | p: angular velocity 
-        | q: angular velocity 
-        | r: angular velocity 
->>>>>>> 557553c7c48b59f1c965e30dadec4506f84a8cab
 
     Keyword Args
     ------------
@@ -202,6 +187,9 @@ class UAVGen(PrognosticsModel):
 
       super().__init__(**kwargs)
       
+      # Get Flight Plan
+      # ================
+      # Option 1: fligh_plan, in form of dict of numpy arrays with way-points and time/speed, is passed, while there's no file to load the flight plan
       if self.parameters['flight_plan'] is not None and self.parameters['flight_file'] == None:
         # Check for appropriate input:
         if not isinstance(self.parameters['flight_plan'], dict):
@@ -209,26 +197,39 @@ class UAVGen(PrognosticsModel):
[…]
