### 2023-01-19 a73dcd8
Second round of PR edits
Imports clean up, deleting unnecessary pass, clean up example
