+++ src/prog_models/models/uav_model/trajectory/trajectory.py
@@ -8,13 +8,9 @@ Auxiliary functions for trajectories and aircraft routes
 import numpy as np
 import scipy.interpolate as interp
 
-import os
-import sys
-sys.path.insert(1, os.path.join(os.path.dirname(__file__), '../utilities/'))
-
-import geometry as geom
+from ..utilities import geometry as geom
 from .nurbs import generate_3dnurbs, generate_intermediate_points, evaluate, knot_vector
-import prog_models.models.uav_model.trajectory.load_trajectories as load 
+import prog_models.models.uav_model.trajectory.load_trajectories as load
 
 
 def angular_vel_from_attitude(phi, theta, psi, delta_t=1):
@@ -167,4 +163,3 @@ class Trajectory():
         self.attitude         = traj['attitude']
         self.angular_velocity = traj['angVel']
         self.time             = traj['time']
-        pass 
