### 2023-01-24 #5070 @matthewjdaigle
Yes, I'm available. I'll complete the review this week.

Matt

On Fri, Jan 13, 2023, 11:23 AM Kelly L. Rowland ***@***.***>
wrote:

> Hi @matthewjdaigle <https://github.com/matthewjdaigle> 👋 would you be
> available for and interested in reviewing this JOSS submission?
>
