### 2023-01-26 #5099 @matthewjdaigle
@kellyrowland Regarding COI policy, I'm a former colleague with the submitting author (5+ years ago), however, within the repository's readme I am credited as an author, since it is based on my previous work (which is also cited in the paper). So there is definitely a perception of a COI here. Let me know how to proceed.
