### 2023-01-26 dbf74c3
Added is_continuous/discrete
