+++ .github/workflows/print-benchmarking.yml
@@ -18,7 +18,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt
           python -m pip install -e .
       - name: benchmark
         id: bench
