+++ src/prog_models/composite_model.py
@@ -43,6 +43,7 @@ class CompositeModel(PrognosticsModel):
         self.states = set()
         self.outputs = set()
         self.events = set()
+        self.performance_metric_keys = set()
         self.model_names = set()
         duplicate_names = {}
         kwargs['models'] = []
@@ -76,6 +77,7 @@ class CompositeModel(PrognosticsModel):
             self.states |= set([name + DIVIDER + x for x in m.states])
             self.outputs |= set([name + DIVIDER + z for z in m.outputs])
             self.events |= set([name + DIVIDER + e for e in m.events])
+            self.performance_metric_keys |= set([name + DIVIDER + p for p in m.performance_metric_keys])
         
         # Handle outputs
         if 'outputs' in kwargs:
@@ -205,6 +207,20 @@ class CompositeModel(PrognosticsModel):
             for key, value in z_i.items():
                 z[name + '.' + key] = value
         return self.OutputContainer(z)
+
+    def performance_metrics(self, x):
+        pf = {}
+        for (name, m) in self.parameters['models']:
+            # Prepare state
+            x_i = m.StateContainer({key: x[name + '.' + key] for key in m.states})
+
+            # Get outputs
+            pf_i = m.performance_metrics(x_i)
+
+            # Save to super outputs
+            for key, value in pf_i.items():
+                pf[name + '.' + key] = value
+        return pf
 
     def event_state(self, x):
         e = {}
