+++ src/prog_models/models/test_models/linear_models.py
@@ -67,6 +67,16 @@ class OneInputOneOutputNoEventLM(LinearModel):
     }
 
 
+class OneInputOneOutputNoEventLMPM(OneInputOneOutputNoEventLM):
+    """
+    Same as OneInputOneOutputNoEventLM, but with performance metrics defined as a function. Has a single performance metric that is always the state, plus 1
+    """
+    performance_metric_keys = ['x1+1']
+
+    def performance_metrics(self, x) -> dict:
+        return {'x1+1': x['x1']+1}
+
+
 class OneInputNoOutputOneEventLM(OneInputNoOutputNoEventLM):
     """
     Simple model that increases state by u1 every step. Event occurs when state == 10
