+++ tests/test_powertrain.py
@@ -4,7 +4,7 @@ from io import StringIO
 import sys
 import unittest
 
-from prog_models.models import ESC, DCMotor, Powertrain
+from prog_models.models import ESC, DCMotor, Powertrain, PropellerLoad
 
 
 class TestPowertrain(unittest.TestCase):
@@ -14,7 +14,43 @@ class TestPowertrain(unittest.TestCase):
 
     def tearDown(self):
         sys.stdout = sys.__stdout__
-    
+        
+    def test_propeller_load(self):
+        m = PropellerLoad()
+
+        # initial state
+        x0 = m.initialize()
+        self.assertSetEqual(set(x0.keys()), {'t_l'})
+        self.assertEqual(x0['t_l'], 0)
+
+        # Check state transition
+        x = m.next_state(x0, m.InputContainer({'v_rot': 2}), None)
+        self.assertSetEqual(set(x.keys()), {'t_l'})
+        self.assertEqual(x['t_l'], m.parameters['t_l'] * 4)
+
+        x = m.next_state(x0, m.InputContainer({'v_rot': -2}), None)
+        self.assertEqual(x['t_l'], m.parameters['t_l'] * 4)
+
+        # Check output
+        z = m.output(x)
+        self.assertSetEqual(set(z.keys()), {'t_l'})
+        self.assertEqual(z['t_l'], x['t_l'])
+
+        # Events
+        self.assertSetEqual(set(m.events), {})
+
+        # State limits
+        x_limited = m.apply_limits(m.StateContainer({'t_l': -1}))
+        self.assertEqual(x_limited['t_l'], 0)
+
+        # Updating derived parameters 
+        c_q = m.parameters['C_q']
+        m.parameters['c_q'] *= 2
+        self.assertEqual(m.parameters['C_q'], c_q * 2)
+
+        m.parameters['D'] = 1
+        self.assertEqual(m.parameters['C_q'], m.parameters['c_q'] * m.parameters['rho'])
+
     def test_powertrain(self):
         esc = ESC()
         motor = DCMotor()
