+++ src/prog_models/models/propeller_load.py
@@ -1,6 +1,8 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
+import numpy as np
+
 from .. import PrognosticsModel
 
 def update_Cq(params):
@@ -33,6 +35,10 @@ class PropellerLoad(PrognosticsModel):
         'x0': {
             't_l': 0,
         }
+    }
+
+    state_limits = {
+        't_l': (0, np.inf),
     }
 
     def next_state(self, x, u, dt):
