### 2023-01-31 a628105
test linear model created
Moved linear model tests from test_base_models to separate file.
