+++ tests/test_linear_model.py
@@ -1,0 +1,333 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the National Aeronautics and Space Administration.  All Rights Reserved.
+
+import numpy as np
+import unittest
+
+from prog_models import *
+from prog_models.models.test_models.linear_models import FNoneNoEventStateLM
+
+class LinearThrownObject(LinearModel):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 40,  # m/s
+        'g': -9.81  # Acceleration due to gravity in m/s^2
+    }
+
+    def initialize(self, u=None, z=None):
+        return self.StateContainer({
+            'x': self.parameters['thrower_height'],  # Thrown, so initial altitude is height of thrower
+            'v': self.parameters['throwing_speed']  # Velocity at which the ball is thrown - this guy is a professional baseball pitcher
+            })
+    
+    def threshold_met(self, x):
+        return {
+            'falling': x['v'] < 0,
+            'impact': x['x'] <= 0
+        }
+
+    def event_state(self, x): 
+        x_max = x['x'] + np.square(x['v'])/(-self.parameters['g']*2) # Use speed and position to estimate maximum height
+        return {
+            'falling': np.maximum(x['v']/self.parameters['throwing_speed'],0),  # Throwing speed is max speed
+            'impact': np.maximum(x['x']/x_max,0) if x['v'] < 0 else 1  # 1 until falling begins, then it's fraction of height
+        }
+
+
+class TestLinearModel(unittest.TestCase):
+    def test_linear_model(self):
+        m = LinearThrownObject()
+        m.simulate_to_threshold(lambda t, x = None: m.InputContainer({}))
+        # len() = events states inputs outputs
+        #         1      2      0      1
+        # Matrix overwrite type checking (Can't set attributes for B, D, G; not overwritten)
+        # when matrix is not of type NumPy ndarray or standard list
+        # @A
+        with self.assertRaises(TypeError):
+            m.A = "[[0, 1], [0, 0]]" # string
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.A = None # None
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.A = 0 # int
+            m.matrixCheck()
+        with self.assertRaises(TypeError):
+            m.A = 3.14 # float
+            m.matrixCheck()
+        with self.assertRaises(TypeError):
+            m.A = {} # dict
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.A = () # tuple
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.A = set() # set
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.A = True # boolean
+            m.matrixCheck()
+        # @C
+        with self.assertRaises(TypeError):
+            m.C = "[[0, 1], [0, 0]]" # string
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.C = None # None
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
+            m.C = 0 # int
+            m.matrixCheck()
+        with self.assertRaises(TypeError):
+            m.C = 3.14 # float
+            m.matrixCheck()
+        with self.assertRaises(TypeError):
+            m.C = {} # dict
+            m.matrixCheck() 
[…]
