### 2023-02-01 4656ca9
Cleaning up LinearThrownObject
Removed LinearThrownObject from test_linear_model class and directly imported the LinearThrownObject from test_base_models.
