+++ tests/test_linear_model.py
@@ -5,43 +5,7 @@ import unittest
 
 from prog_models import *
 from prog_models.models.test_models.linear_models import FNoneNoEventStateLM
-
-class LinearThrownObject(LinearModel):
-    inputs = [] 
-    states = ['x', 'v']
-    outputs = ['x']
-    events = ['impact']
-
-    A = np.array([[0, 1], [0, 0]])
-    E = np.array([[0], [-9.81]])
-    C = np.array([[1, 0]])
-    F = None # Will override method
-
-    default_parameters = {
-        'thrower_height': 1.83,  # m
-        'throwing_speed': 40,  # m/s
-        'g': -9.81  # Acceleration due to gravity in m/s^2
-    }
-
-    def initialize(self, u=None, z=None):
-        return self.StateContainer({
-            'x': self.parameters['thrower_height'],  # Thrown, so initial altitude is height of thrower
-            'v': self.parameters['throwing_speed']  # Velocity at which the ball is thrown - this guy is a professional baseball pitcher
-            })
-    
-    def threshold_met(self, x):
-        return {
-            'falling': x['v'] < 0,
-            'impact': x['x'] <= 0
-        }
-
-    def event_state(self, x): 
-        x_max = x['x'] + np.square(x['v'])/(-self.parameters['g']*2) # Use speed and position to estimate maximum height
-        return {
-            'falling': np.maximum(x['v']/self.parameters['throwing_speed'],0),  # Throwing speed is max speed
-            'impact': np.maximum(x['x']/x_max,0) if x['v'] < 0 else 1  # 1 until falling begins, then it's fraction of height
-        }
-
+from test_base_models import LinearThrownObject
 
 class TestLinearModel(unittest.TestCase):
     def test_linear_model(self):
@@ -269,7 +233,7 @@ class TestLinearModel(unittest.TestCase):
         with self.assertRaises(AttributeError): 
             m.G = np.array([[]]) # less row
             m.matrixCheck()
-            
+
     def test_F_property_not_none(self):
         class ThrownObject(LinearThrownObject):
             F = np.array([[1, 0]]) # Will override method
@@ -314,7 +278,7 @@ class TestLinearModel(unittest.TestCase):
                     'falling': x['v'] < 0,
                     'impact': x['x'] <= 0
                 }
-        # test coverage needs testing of event_state not overridden
+        # Needs more development; test coverage needs testing of event_state not overridden
 
 # This allows the module to be executed directly
 def run_tests():
