### 2023-02-02 54d27b1
Testing when LinearModel is initiated when no inputs are given but matrix B is defined.
