### 2023-02-02 ff66afb
Fixed Pull Request Comments 2/2
Move LinearThrownObject to to fix github specific import issues.
