+++ tests/test_linear_model.py
@@ -5,7 +5,8 @@ import unittest
 
 from prog_models import *
 from prog_models.models.test_models.linear_models import FNoneNoEventStateLM
-from test_base_models import LinearThrownObject
+from prog_models.models.test_models.linear_thrown_object import LinearThrownObject
+
 
 class TestLinearModel(unittest.TestCase):
     def test_linear_model(self):
