### 2023-02-07 457e1b5
changed some code for clarification and added comments
