+++ src/prog_models/utils/containers.py
@@ -19,7 +19,9 @@ class DictLikeMatrixWrapper():
         The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
     def __init__(self, keys : list, data : Union[dict, np.array]):
-        self._keys = keys.copy()
+        if not isinstance(keys, list):  # if keys is not a list- try to convert it to one
+            keys = list(keys)
+        self._keys = keys.copy()    # saves list of keys in object, ids for data
         if isinstance(data, np.matrix):
             self.matrix = np.array(data, dtype=np.float64)
         elif isinstance(data, np.ndarray):
@@ -32,60 +34,90 @@ class DictLikeMatrixWrapper():
                     [data[key]] if key in data else [None] for key in keys
                 ], dtype=np.float64)
         else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")     
+            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
     def __reduce__(self):
         return (DictLikeMatrixWrapper, (self._keys, self.matrix))
 
+    # get all values associated with a key, ex: all values of 'i'
     def __getitem__(self, key : str) -> int:
-        row = self.matrix[self._keys.index(key)]
-        if len(row) == 1:
-            return self.matrix[self._keys.index(key)][0]
-        return self.matrix[self._keys.index(key)]
+        row = self.matrix[self._keys.index(key)]    # creates list from a row of matrix
+        if len(row) == 1:   # if row list contains 1 value, returns that value (non-vectorized)
+            return row[0]
+        return row   # else returns entire row/list (vectorized case)
 
+    # sets a row at the key given
     def __setitem__(self, key : str, value : int) -> None:
-        self.matrix[self._keys.index(key)] = np.atleast_1d(value)
+        index = self._keys.index(key)   # the int value index for the key given
+        self.matrix[index] = np.atleast_1d(value)
 
+    # removes row associated with key
     def __delitem__(self, key : str) -> None:
         self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
         self._keys.remove(key)
 
+    # add another matrix to the existing matrix
     def __add__(self, other : "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
         return DictLikeMatrixWrapper(self._keys, self.matrix + other.matrix)
 
+    # creates iterator object for the list of keys
     def __iter__(self):
         return iter(self._keys)
 
+    # returns the length of key list
     def __len__(self) -> int:
         return len(self._keys)
 
+    # checks if the other is a dictionary
     def __eq__(self, other : "DictLikeMatrixWrapper") -> bool:
         if isinstance(other, dict):
-            return list(self.keys()) == list(other.keys()) and (self.matrix == np.array([[other[key]] for key in self._keys])).all()
-        return self.keys() == other.keys() and (self.matrix == other.matrix).all()
+            # checks that the list of keys for each matrix match
+            list_key_check = (list(self.keys()) == list(other.keys()))
+            # checks to see that each row matches
+            matrix_check = (self.matrix == np.array([[other[key]] for key in self._keys])).all()
+            return list_key_check and matrix_check
+        list_key_check = self.keys() == other.keys()
+        matrix_check = (self.matrix == other.matrix).all()
+        return list_key_check and matrix_check
 
+    # returns hash value sum for keys and matrix
     def __hash__(self):
         return hash(self.keys) + hash(self.matrix)
-    
+
+    # (?) why
     def __str__(self) -> str:
         return self.__repr__()
 
+    # gets the list of values associated with the key given
+    def get(self, key, default=None):
+        if key in self._keys:
[…]
