### 2023-02-08 52ab518
changed some code for clarification and added comments
