+++ src/prog_models/utils/containers.py
@@ -10,7 +10,7 @@ from ..exceptions import ProgModelTypeError
 class DictLikeMatrixWrapper():
     """
     A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-    
+
     Arguments
     ---------
     keys: list
@@ -18,11 +18,11 @@ class DictLikeMatrixWrapper():
     data: dict or numpy array
         The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
-    def __init__(self, keys : list, data : Union[dict, np.array]):
-        if not isinstance(keys, list):
-            keys = list(keys)
 
-        self._keys = keys.copy()
+    def __init__(self, keys: list, data: Union[dict, np.array]):
+        if not isinstance(keys, list):  # if keys is not a list- try to convert it to one
+            keys = list(keys)
+        self._keys = keys.copy()  # saves list of keys in object, ids for data
         if isinstance(data, np.matrix):
             self.matrix = np.array(data, dtype=np.float64)
         elif isinstance(data, np.ndarray):
@@ -35,66 +35,91 @@ class DictLikeMatrixWrapper():
                     [data[key]] if key in data else [None] for key in keys
                 ], dtype=np.float64)
         else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")     
+            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
     def __reduce__(self):
         return (DictLikeMatrixWrapper, (self._keys, self.matrix))
 
-    def __getitem__(self, key : str) -> int:
-        row = self.matrix[self._keys.index(key)]
-        if len(row) == 1:
-            return self.matrix[self._keys.index(key)][0]
-        return self.matrix[self._keys.index(key)]
+    # get all values associated with a key, ex: all values of 'i'
+    def __getitem__(self, key: str) -> int:
+        row = self.matrix[self._keys.index(key)]  # creates list from a row of matrix
+        if len(row) == 1:  # if row list contains 1 value, returns that value (non-vectorized)
+            return row[0]
+        return row  # else returns entire row/list (vectorized case)
 
-    def __setitem__(self, key : str, value : int) -> None:
-        self.matrix[self._keys.index(key)] = np.atleast_1d(value)
+    # sets a row at the key given
+    def __setitem__(self, key: str, value: int) -> None:
+        index = self._keys.index(key)  # the int value index for the key given
+        self.matrix[index] = np.atleast_1d(value)
 
-    def __delitem__(self, key : str) -> None:
+    # removes row associated with key
+    def __delitem__(self, key: str) -> None:
         self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
         self._keys.remove(key)
 
-    def __add__(self, other : "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
+    # add another matrix to the existing matrix
+    def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
         return DictLikeMatrixWrapper(self._keys, self.matrix + other.matrix)
 
+    # creates iterator object for the list of keys
     def __iter__(self):
         return iter(self._keys)
 
+    # returns the length of key list
     def __len__(self) -> int:
         return len(self._keys)
 
-    def __eq__(self, other : "DictLikeMatrixWrapper") -> bool:
+    # checks if the other is a dictionary
+    def __eq__(self, other: "DictLikeMatrixWrapper") -> bool:
         if isinstance(other, dict):
-            return list(self.keys()) == list(other.keys()) and (self.matrix == np.array([[other[key]] for key in self._keys])).all()
[…]
