### 2023-02-10 6d33c32
Traj gen modularity change - new models, moving files, new example
