+++ src/prog_models/aux_fcns/traj_gen_utils/load_trajectories.py
@@ -8,7 +8,8 @@ Functions to extract user-defined waypoint information and convert to appropriat
 import numpy as np
 import datetime as dt
 
-from ..utilities import loadsavewrite_utils as loadsave
+# from ..utilities import loadsavewrite_utils as loadsave
+from prog_models.aux_fcns.traj_gen_utils import loadsave_utils as loadsave
 from prog_models.exceptions import ProgModelInputException
 
 DEG2RAD = np.pi/180.0
