+++ src/prog_models/models/uav_model/uav_model.py
@@ -7,9 +7,8 @@ from warnings import warn
 from typing import Callable
 
 from prog_models.prognostics_model import PrognosticsModel
-from .trajectory import route, trajectory
 from .vehicles import AircraftModels
-from .utilities import geometry as geom
+from prog_models.aux_fcns.traj_gen_utils import geometry as geom
 from prog_models.exceptions import ProgModelInputException
 
 class UAVGen(PrognosticsModel):
@@ -146,7 +145,7 @@ class UAVGen(PrognosticsModel):
     -------------
      .. [0] M. Corbetta et al., "Real-time UAV trajectory prediction for safely monitoring in low-altitude airspace," AIAA Aviation 2019 Forum,  2019. https://arc.aiaa.org/doi/pdf/10.2514/6.2019-3514
     """
-    events = ['TrajectoryComplete']
+    events = [] # ['TrajectoryComplete']
     inputs = ['T','mx','my','mz']
     states = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r','t']
     outputs = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r']
@@ -154,8 +153,8 @@ class UAVGen(PrognosticsModel):
 
     default_parameters = {  # Set to defaults
         # Flight information
-        'flight_file': None, 
-        'flight_name': 'flight-1', 
+        # 'flight_file': None, 
+        # 'flight_name': 'flight-1', 
         'aircraft_name': 'aircraft-1', 
         'flight_plan': None,
 
@@ -169,10 +168,10 @@ class UAVGen(PrognosticsModel):
         'hovering_time': 0.0,
         'takeoff_time': 0.0, 
         'landing_time': 0.0, 
-        'waypoint_weights': 10.0, 
-        'adjust_eta': None, 
-        'nurbs_basis_length': 2000, 
-        'nurbs_order': 4, 
+        # 'waypoint_weights': 10.0, 
+        # 'adjust_eta': None, 
+        # 'nurbs_basis_length': 2000, 
+        # 'nurbs_order': 4, 
         'final_time_buffer_sec': 30, 
         'final_space_buffer_m': 2, 
 
@@ -185,38 +184,6 @@ class UAVGen(PrognosticsModel):
 
       super().__init__(**kwargs)
       
-      # Get Flight Plan
-      # ================
-      # Option 1: fligh_plan, in form of dict of numpy arrays with way-points and time, is passed, while there's no file to load the flight plan
-      if self.parameters['flight_plan'] is not None and self.parameters['flight_file'] == None:
-        # Check for appropriate input:
-        if not isinstance(self.parameters['flight_plan'], dict):
-          raise ProgModelInputException("'flight_plan' must be a dictionary. Type {} was given".format(type(self.parameters['flight_plan'])))
-        for flight_plan_element in self.parameters['flight_plan'].values():
-          if not isinstance(flight_plan_element, np.ndarray):
-            raise ProgModelInputException("'flight_plan' entries must be numpy arrays specifying waypoint information. Type {} was given".format(type(flight_plan_element)))
-        
-        # Extract data from flight plan: latitude, longitude, altitude, time stamps
-        flightplan = trajectory.load.convert_dict_inputs(self.parameters['flight_plan'])
-        lat = flightplan['lat_rad']
-        lon = flightplan['lon_rad']
-        alt = flightplan['alt_m']
-        tstamps = flightplan['timestamp']
-      
-      # Option 2: a file with flight plan is passed
-      elif self.parameters['flight_file'] != None and self.parameters['flight_plan'] == None:
[…]
