### 2023-02-11 4cf0c40
created a version of sim_result for dataframe data management and created some tester files.
