### 2023-02-11 f4faae4
Getters & Setters, default params, fix & add tests
