+++ tests/test_linear_model.py
@@ -3,8 +3,8 @@
 import numpy as np
 import unittest
 
-from prog_models.models.test_models.linear_models import FNoneNoEventStateLM
-from prog_models.models.test_models.linear_thrown_object import LinearThrownObject, LinearThrownObject_WrongB
+from prog_models.models.test_models.linear_models import *
+from prog_models.models.thrown_object import LinearThrownObject, LinearThrownObject_WrongB
 
 
 # TODO: IF no inputs, parameters should be not be instantied at all. e.g) Having 0 inputs means B will be a 2x0, which is not possible.
@@ -13,9 +13,18 @@ from prog_models.models.test_models.linear_thrown_object import LinearThrownObje
 
 # Figure out why its an attribute error and not a typeerror FOR paramter B. line 55.
 
+# 
+
 class TestLinearModel(unittest.TestCase):
     def test_linear_model(self):
         m = LinearThrownObject()
+
+        # Parameters for reference
+
+        # A = np.array([[0, 1], [0, 0]])
+        # E = np.array([[0], [-9.81]])
+        # C = np.array([[1, 0]])
+
         m.simulate_to_threshold(lambda t, x = None: m.InputContainer({}))
         # len() = events states inputs outputs
         #         1      2      0      1
@@ -46,33 +55,61 @@ class TestLinearModel(unittest.TestCase):
         with self.assertRaises(TypeError):
             m.A = True # boolean
             m.matrixCheck()
-        # @B
-        # PART 2: Add a test that sees the new runtime check works from linear_model line 73
-        # Tests created for @B mainly because we are exploring when optional paramters are inputted.
+        # Matrix Dimension Checking
+        # when matrix is not proper dimensional (1-D array = C, D, G; 2-D array = A,B,E; None = F;)
+        # @A 2x2
         with self.assertRaises(AttributeError):
-            m.B = "[[0, 1], [0, 0]]" # string
-            m.matrixCheck() 
-        with self.assertRaises(AttributeError):
-            m.B = None # None
+            m.A = np.array([[0, 1]]) # 1-D array
             m.matrixCheck()
         with self.assertRaises(AttributeError):
+            m.A = np.array([[0, 1], [0, 0], [1, 0]]) # 3-D array
+            m.matrixCheck()
+        # When Matrix is improperly formed
+        with self.assertRaises(AttributeError):
+            m.A = np.array([[0, 1, 2, 3], [0, 0, 1, 2]]) # extra column values per row
+            m.matrixCheck()
+        with self.assertRaises(AttributeError):
+            m.A = np.array([[0], [0]]) # less column values per row
+            m.matrixCheck()
+        with self.assertRaises(AttributeError): 
+            m.A = np.array([[0, 1], [0, 0], [2, 2]]) # extra row
+            m.matrixCheck()
+        with self.assertRaises(AttributeError): 
+            m.A = np.array([[0, 1]]) # less row
+            m.matrixCheck()
+        # Reset Process
+        m.A = np.array([[0, 1], [0, 2]])
+        m.matrixCheck()
+
+        # @B
+        with self.assertRaises(TypeError):
+            m.B = "[[0, 1], [0, 0]]"
+            m.matrixCheck()
+        with self.assertRaises(TypeError):
+            m.B = None # None
+            m.matrixCheck() 
+        with self.assertRaises(TypeError):
             m.B = 0 # int
             m.matrixCheck()
-        with self.assertRaises(AttributeError):
+        with self.assertRaises(TypeError):
             m.B = 3.14 # float
             m.matrixCheck()
-        with self.assertRaises(AttributeError):
+        with self.assertRaises(TypeError):
             m.B = {} # dict
[…]
