+++ src/prog_models/models/test_models/linear_models.py
@@ -141,3 +141,11 @@ class TwoInputNoOutputTwoEventLM(LinearModel):
             'x1': 0
         }
     }
+
+class ThreeStatesTwoB(LinearModel):
+    inputs = ['u1']
+    states = ['x1']
+    A = np.array([[0]])
+    B = np.array([[1, 2]])
+    C = np.empty((0,1))
+    F = None

[…]
