+++ tests/test_base_models.py
@@ -15,7 +15,7 @@ from prog_models import *
 from prog_models.models import *
 from prog_models.models.test_models.linear_models import (
     OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM)
-from prog_models.models.test_models.linear_thrown_object import LinearThrownObject
+from prog_models.models.thrown_object import LinearThrownObject
 
 class MockModel():
     states = ['a', 'b', 'c', 't']
