### 2023-02-14 f8e0fc9
Update battery_circuit.py
