+++ src/prog_models/prognostics_model_df.py
@@ -1,1463 +1,0 @@
-# Copyright © 2021 United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration.  All Rights Reserved.
-
-from abc import ABC
-from collections import abc, namedtuple
-from copy import deepcopy
-import itertools
-import json
-from numbers import Number
-import numpy as np
-from typing import Callable, Iterable, List
-from warnings import warn
-import pandas as pd
-
-from .exceptions import ProgModelInputException, ProgModelTypeError, ProgModelException, ProgModelStateLimitWarning
-from .sim_result import SimResult, LazySimResult
-from .utils import ProgressBar
-from .utils.containers import DictLikeMatrixWrapper
-from .utils.parameters import PrognosticsModelParameters
-from .utils.serialization import *
-
-
-class PrognosticsModelDf(ABC):
-    """
-    A general time-variant state space :term:`model` of system degradation behavior.
-
-    The PrognosticsModel class is a wrapper around a mathematical model of a system as represented by a state, output, input, event_state and threshold equation.
-
-    A Model also has a parameters structure, which contains fields for various model parameters.
-
-    Keyword Args
-    ------------
-        process_noise : Optional, float or dict[str, float]
-          :term:`Process noise<process noise>` (applied at dx/next_state).
-          Can be number (e.g., .2) applied to every state, a dictionary of values for each
-          state (e.g., {'x1': 0.2, 'x2': 0.3}), or a function (x) -> x
-        process_noise_dist : Optional, str
-          distribution for :term:`process noise` (e.g., normal, uniform, triangular)
-        measurement_noise : Optional, float or dict[str, float]
-          :term:`Measurement noise<measurement noise>` (applied in output eqn).
-          Can be number (e.g., .2) applied to every output, a dictionary of values for each
-          output (e.g., {'z1': 0.2, 'z2': 0.3}), or a function (z) -> z
-        measurement_noise_dist : Optional, str
-          distribution for :term:`measurement noise` (e.g., normal, uniform, triangular)
-
-    Additional parameters specific to the model
-
-    Raises
-    ------
-        ProgModelTypeError, ProgModelInputException, ProgModelException
-
-    Example
-    -------
-        m = PrognosticsModel(process_noise = 3.2)
-
-    Attributes
-    ----------
-        is_vectorized : bool, optional
-            True if the model is vectorized, False otherwise. Default is False
-        default_parameters : dict[str, float], optional
-            Default parameters for the model class
-        parameters : dict[str, float]
-            Parameters for the specific model object. This is created automatically from the default_parameters and kwargs
-        state_limits: dict[str, tuple[float, float]], optional
-            Limits on the state variables format {'state_name': (lower_limit, upper_limit)}
-        param_callbacks : dict[str, list[function]], optional
-            Callbacks for derived parameters
-        inputs: list[str]
-            Identifiers for each :term:`input`
-        states: list[str]
-            Identifiers for each :term:`state`
-        outputs: list[str]
-            Identifiers for each :term:`output`
-        performance_metric_keys: list[str], optional
-            Identifiers for each performance metric
-        events: list[str], optional
-            Identifiers for each :term:`event` predicted
-        StateContainer : DataFrameShortcuts
-            Class for state container - used for representing :term:`state`
-        OutputContainer : DataFrameShortcuts
-            Class for output container - used for representing :term:`output`
-        InputContainer : DataFrameShortcuts
-            Class for input container - used for representing :term:`input`
-    """
-    is_vectorized = False
-
[…]
