+++ src/prog_models/utils/containers.py
@@ -10,7 +10,7 @@ from ..exceptions import ProgModelTypeError
 class DictLikeMatrixWrapper():
     """
     A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-    
+
     Arguments
     ---------
     keys: list
@@ -18,10 +18,15 @@ class DictLikeMatrixWrapper():
     data: dict or numpy array
         The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
-    def __init__(self, keys : list, data : Union[dict, np.array]):
+
+    """
+    if keys is not a list- try to convert it to one
+        saves list of keys in object, ids for data
+    """
+
+    def __init__(self, keys: list, data: Union[dict, np.array]):
         if not isinstance(keys, list):
             keys = list(keys)
-
         self._keys = keys.copy()
         if isinstance(data, np.matrix):
             self.matrix = np.array(data, dtype=np.float64)
@@ -35,66 +40,135 @@ class DictLikeMatrixWrapper():
                     [data[key]] if key in data else [None] for key in keys
                 ], dtype=np.float64)
         else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")     
+            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
+    """
+        reduce is overridden for pickles
+    """
     def __reduce__(self):
         return (DictLikeMatrixWrapper, (self._keys, self.matrix))
 
-    def __getitem__(self, key : str) -> int:
+    """
+        get all values associated with a key, ex: all values of 'i'
+            row: creates list from a row of matrix
+        if: row list contains 1 value, returns that value (non-vectorized)
+        else: returns entire row/list (vectorized case)
+    """
+    def __getitem__(self, key: str) -> int:
         row = self.matrix[self._keys.index(key)]
         if len(row) == 1:
-            return self.matrix[self._keys.index(key)][0]
-        return self.matrix[self._keys.index(key)]
+            return row[0]
+        return row
 
-    def __setitem__(self, key : str, value : int) -> None:
-        self.matrix[self._keys.index(key)] = np.atleast_1d(value)
+    """
+        sets a row at the key given
+            the int value index for the key given
+    """
+    def __setitem__(self, key: str, value: int) -> None:
+        index = self._keys.index(key)
+        self.matrix[index] = np.atleast_1d(value)
 
-    def __delitem__(self, key : str) -> None:
+    """
+        removes row associated with key
+    """
+    def __delitem__(self, key: str) -> None:
         self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
         self._keys.remove(key)
 
-    def __add__(self, other : "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
+    """
+        add another matrix to the existing matrix
+    """
+    def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
         return DictLikeMatrixWrapper(self._keys, self.matrix + other.matrix)
 
+    """
+        creates iterator object for the list of keys
+    """
     def __iter__(self):
         return iter(self._keys)
 
+    """
+        returns the length of key list
+    """
     def __len__(self) -> int:
         return len(self._keys)
 
-    def __eq__(self, other : "DictLikeMatrixWrapper") -> bool:
+    """
+        Compares two DictLikeMatrixWrappers (i.e. *Containers) or a DictLikeMatrixWrapper and a dictionary
+            if: checks that the list of keys for each matrix match
+            bools: 
+                list_key_check: checks that the list of keys for each matrix are equal
+                matrix_check: checks to see that each row matches
[…]
