+++ src/prog_models/utils/containers.py
@@ -19,12 +19,13 @@ class DictLikeMatrixWrapper():
         The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
 
-    """
-    if keys is not a list- try to convert it to one
-        saves list of keys in object, ids for data
-    """
-
     def __init__(self, keys: list, data: Union[dict, np.array]):
+        """
+        Initializes the container
+
+        if keys is not a list- try to convert it to one
+            saves list of keys in object, ids for data
+        """
         if not isinstance(keys, list):
             keys = list(keys)
         self._keys = keys.copy()
@@ -42,66 +43,67 @@ class DictLikeMatrixWrapper():
         else:
             raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
-    """
-        reduce is overridden for pickles
-    """
     def __reduce__(self):
+        """
+        reduce is overridden for pickles
+        """
         return (DictLikeMatrixWrapper, (self._keys, self.matrix))
 
-    """
+    def __getitem__(self, key: str) -> int:
+        """
         get all values associated with a key, ex: all values of 'i'
             row: creates list from a row of matrix
         if: row list contains 1 value, returns that value (non-vectorized)
         else: returns entire row/list (vectorized case)
-    """
-    def __getitem__(self, key: str) -> int:
+        """
         row = self.matrix[self._keys.index(key)]
         if len(row) == 1:
             return row[0]
         return row
 
-    """
+    def __setitem__(self, key: str, value: int) -> None:
+        """
         sets a row at the key given
             the int value index for the key given
-    """
-    def __setitem__(self, key: str, value: int) -> None:
+        """
         index = self._keys.index(key)
         self.matrix[index] = np.atleast_1d(value)
 
-    """
-        removes row associated with key
-    """
+    
     def __delitem__(self, key: str) -> None:
+        """
+        removes row associated with key
+        """
         self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
         self._keys.remove(key)
 
-    """
-        add another matrix to the existing matrix
-    """
     def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
+        """
+        add another matrix to the existing matrix
+        """
         return DictLikeMatrixWrapper(self._keys, self.matrix + other.matrix)
 
-    """
-        creates iterator object for the list of keys
-    """
     def __iter__(self):
+        """
+        creates iterator object for the list of keys
+        """
         return iter(self._keys)
 
-    """
-        returns the length of key list
-    """
     def __len__(self) -> int:
+        """
+        returns the length of key list
+        """
         return len(self._keys)
 
-    """
+    def __eq__(self, other: "DictLikeMatrixWrapper") -> bool:
+        """
         Compares two DictLikeMatrixWrappers (i.e. *Containers) or a DictLikeMatrixWrapper and a dictionary
             if: checks that the list of keys for each matrix match
             bools: 
                 list_key_check: checks that the list of keys for each matrix are equal
                 matrix_check: checks to see that each row matches
             returns bool value
-    """
-    def __eq__(self, other: "DictLikeMatrixWrapper") -> bool:
+        """
         if isinstance(other, dict):
             list_key_check = (list(self.keys()) == list(other.keys()))
             matrix_check = (self.matrix == np.array([[other[key]] for key in self._keys])).all()
@@ -110,65 +112,65 @@ class DictLikeMatrixWrapper():
         matrix_check = (self.matrix == other.matrix).all()
         return list_key_check and matrix_check
 
-    """
-        returns hash value sum for keys and matrix
-    """
     def __hash__(self):
+        """
[…]
