### 2023-02-15 052414e
Removing aircraft_name and flight_name
