+++ tests/test_composite.py
@@ -1,0 +1,20 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the National Aeronautics and Space Administration.  All Rights Reserved.
+
+from copy import deepcopy
+import io
+import numpy as np
+from os.path import dirname, join
+import pickle
+import sys
+import unittest
+
+# This ensures that the directory containing ProgModelTemplate is in the python search directory
+sys.path.append(join(dirname(__file__), ".."))
+
+from prog_models import *
+from prog_models.models import *
+from prog_models.models.test_models.linear_models import (
+    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM)
+from prog_models.models.thrown_object import LinearThrownObject
+
+
