### 2023-02-19 #5099 @samim91
Hi Kelly,

Yes, would be happy to. Can you include my stanford address in copy? ***@***.***

Best,
Samim

Sent from Outlook for iOS<https://aka.ms/o0ukef>
