+++ model_save_json.txt
@@ -1,0 +1,1 @@
+{"process_noise": {"tb": 0.0, "Vo": 0.0, "Vsn": 0.0, "Vsp": 0.0, "qnB": 0.0, "qnS": 0.0, "qpB": 0.0, "qpS": 0.0, "_original_type": "DictLikeMatrixWrapper"}, "measurement_noise": {"t": 0.0, "v": 0.0, "_original_type": "DictLikeMatrixWrapper"}, "qMobile": 7600, "xnMax": 0.6, "xnMin": 0, "xpMax": 1.0, "xpMin": 0.4, "Ro": 0.117215, "alpha": 0.5, "Sn": 0.000437545, "Sp": 0.00030962, "kn": 2120.96, "kp": 248898, "Vol": 2e-05, "VolSFraction": 0.1, "tDiffusion": 7000000.0, "to": 6.08671, "tsn": 1001.38, "tsp": 46.4311, "U0p": 4.03, "Ap": [-31593.7, 0.106747, 24606.4, -78561.9, 13317.9, 307387, 84916.1, -1074690.0, 2285.04, 990894, 283920, -161513, -469218], "U0n": 0.01, "An": [86.19, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], "x0": {"Vo": 0, "Vsn": 0, "Vsp": 0, "tb": 292.1, "qpS": 506.6666666666668, "qpB": 4560.000000000001, "qnS": 760.0, "qnB": 6840.0}, "VEOD": 3.0, "VDropoff": 0.1, "qMax": 12666.666666666668, "qpMin": 5066.666666666668, "qpMax": 12666.666666666668, "qpSMin": 506.6666666666668, "qpBMin": 4560.000000000001, "qpSMax": 1266.666666666667, "qpBMax": 11400.000000000002, "qnMin": 0.0, "qnMax": 7600.0, "qSMax": 1266.666666666667, "qBMax": 11400.000000000002, "VolS": 2.0000000000000003e-06, "VolB": 1.8e-05, "qnSMin": 0.0, "qnBMin": 0.0, "qnSMax": 760.0, "qnBMax": 6840.0}

[…]
