### 2023-02-23 6ffae34
eq added, basic copy tests added, need to be more thorough
