+++ src/prog_models/models/thrown_object.py
@@ -174,8 +174,66 @@ class LinearThrownObject_WrongB(LinearThrownObject):
     C = np.array([[1, 0]])
     F = None # Will override method
 
+class LinearThrownObject2(LinearThrownObject):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 20,  # m/s
+        'g': -9.81  # Acceleration due to gravity in m/s^2
+    }
+
+# Wrong x statecontainer parameter. Has Throwing_speed when it should be thrower_height
+class LinearThrownObject3(LinearThrownObject):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 40,  # m/s
+        'g': -9.81  # Acceleration due to gravity in m/s^2
+    }
+
+    def initialize(self, u=None, z=None):
+        return self.StateContainer({
+            'x': self.parameters['throwing_speed'],  # Thrown, so initial altitude is height of thrower
+            })
+
+class LinearThrownObjectDiffKey(LinearThrownObject):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 40,  # m/s
+        'g': -9.81,  # Acceleration due to gravity in m/s^2
+        'x': 1111
+    }
+
 #future implementation
-class LinearThrownOverrideClassLevel(LinearModel):
+class LinearThrownOverrideClassLevel(LinearThrownObject):
     def __init__ (self):
 
         self.A = np.array([[0, 1], [0, 0]])
