+++ tests/test_base_models.py
@@ -13,9 +13,8 @@ sys.path.append(join(dirname(__file__), ".."))
 
 from prog_models import *
 from prog_models.models import *
-from prog_models.models.test_models.linear_models import (
-    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM)
-from prog_models.models.thrown_object import LinearThrownObject, LinearThrownObject2, LinearThrownObjectDiffKey, LinearThrownObject3
+from prog_models.models.test_models.linear_models import (OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
+
 
 class MockModel():
     states = ['a', 'b', 'c', 't']
@@ -68,6 +67,42 @@ def derived_callback3(config):
     return {  # Testing 2nd chained update 
         'p4': -2 * config['p2'], 
     }
+
+class LinearThrownObject(LinearModel):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 40,  # m/s
+        'g': -9.81  # Acceleration due to gravity in m/s^2
+    }
+
+    def initialize(self, u=None, z=None):
+        return self.StateContainer({
+            'x': self.parameters['thrower_height'],  # Thrown, so initial altitude is height of thrower
+            'v': self.parameters['throwing_speed']  # Velocity at which the ball is thrown - this guy is a professional baseball pitcher
+            })
+    
+    def threshold_met(self, x):
+        return {
+            'falling': x['v'] < 0,
+            'impact': x['x'] <= 0
+        }
+
+    def event_state(self, x): 
+        x_max = x['x'] + np.square(x['v'])/(-self.parameters['g']*2) # Use speed and position to estimate maximum height
+        return {
+            'falling': np.maximum(x['v']/self.parameters['throwing_speed'],0),  # Throwing speed is max speed
+            'impact': np.maximum(x['x']/x_max,0) if x['v'] < 0 else 1  # 1 until falling begins, then it's fraction of height
+        }
 
 class MockModelWithDerived(MockProgModel):
     param_callbacks = {
@@ -143,39 +178,8 @@ class TestModels(unittest.TestCase):
         self.assertAlmostEqual(m.parameters['p3'], 5, 5)
         self.assertAlmostEqual(m.parameters['p4'], -10, 5)
 
-
-
-    #Move this to the bottom after done testing
-    def test_parameter_equality(self):
-        # Tests to include:
-            # A key exists in one and not in another, for both directions.
-            # Wrong parameters
-
-        m1 = LinearThrownObject()
-        m2 = LinearThrownObject()
-
-        self.assertTrue(m1.parameters == m2.parameters) #Checking too see if the parameters are equal
-        self.assertTrue(m2.parameters == m1.parameters) #Parameters should be equal
-
-        m3 = LinearThrownObject2() # A model with a different throwing speed
-
-        self.assertFalse(m1.parameters == m3.parameters)
-        self.assertFalse(m3.parameters == m1.parameters) # Checking both directions 
-
-        m4 = LinearThrownObjectDiffKey() # Model with an extra default parameter.
-
[…]
