+++ tests/test_base_models.py
@@ -8,13 +8,14 @@ import pickle
 import sys
 import unittest
 
+
 # This ensures that the directory containing ProgModelTemplate is in the python search directory
 sys.path.append(join(dirname(__file__), ".."))
 
 from prog_models import *
 from prog_models.models import *
 from prog_models.models.test_models.linear_models import (OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
-
+from prog_models.models.thrown_object import *
 
 class MockModel():
     states = ['a', 'b', 'c', 't']
