+++ tests/test_composite.py
@@ -14,7 +14,215 @@ sys.path.append(join(dirname(__file__), ".."))
 from prog_models import *
 from prog_models.models import *
 from prog_models.models.test_models.linear_models import (
-    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM)
+    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
 from prog_models.models.thrown_object import LinearThrownObject
 
+class TestCompositeModel(unittest.TestCase):
+    def test_composite_broken(self):
+        m1 = OneInputOneOutputNoEventLM()
 
+        # Insufficient number of models
+        with self.assertRaises(ValueError):
+            CompositeModel([])
+        with self.assertRaises(ValueError):
+            CompositeModel([m1])
+        
+        # Wrong type
+        with self.assertRaises(ValueError):
+            CompositeModel([m1, m1, 'abc'])
+
+        # Incorrect named format
+        with self.assertRaises(ValueError):
+            # Too many elements
+            CompositeModel([('a', m1, 'Something else'), ('b', m1)])
+        with self.assertRaises(ValueError):
+            # Not a string
+            CompositeModel([(m1, m1)])
+        with self.assertRaises(ValueError):
+            # Not a model
+            CompositeModel([('a', 'b')])
+        with self.assertRaises(ValueError):
+            # Too few elements
+            CompositeModel([(m1, )])
+
+        # Incorrect connections
+        with self.assertRaises(ValueError):
+            # without model name
+            CompositeModel([m1, m1], connections=[('z1', 'u1')])
+        with self.assertRaises(ValueError):
+            # broken in
+            CompositeModel([m1, m1], connections=[('z1', 'OneInputOneOutputNoEventLM.u1')])
+        with self.assertRaises(ValueError):
+            # broken out
+            CompositeModel([m1, m1], connections=[('OneInputOneOutputNoEventLM.z1', 'u1')])
+        with self.assertRaises(ValueError):
+            # Switched
+            CompositeModel([m1, m1], connections=[('OneInputOneOutputNoEventLM.u1', 'OneInputOneOutputNoEventLM_2.z1')])
+        with self.assertRaises(ValueError):
+            # Improper format - too long
+            CompositeModel([m1, m1], connections=[('OneInputOneOutputNoEventLM.z1', 'OneInputOneOutputNoEventLM.u1', 'Something else')])
+        with self.assertRaises(ValueError):
+            # Improper format - not a string
+            CompositeModel([m1, m1], connections=[(m1, m1)])
+        with self.assertRaises(ValueError):
+            # Improper format - too short
+            CompositeModel([m1, m1], connections=[('OneInputOneOutputNoEventLM.z1', )])
+        with self.assertRaises(ValueError):
+            # Improper format - not a tuple
+            CompositeModel([m1, m1], connections=['m1'])
+
+        # Incorrect outputs
+        with self.assertRaises(ValueError):
+            # without model name
[…]
