+++ src/prog_models/utils/containers.py
@@ -4,30 +4,23 @@
 import numpy as np
 from typing import Union
 
-from ..exceptions import ProgModelTypeError
+from prog_models.exceptions import ProgModelTypeError
 
 
 class DictLikeMatrixWrapper():
     """
     A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-
-    Arguments
-    ---------
-    keys: list
-        The keys of the dictionary. e.g., model.states or model.inputs
-    data: dict or numpy array
-        The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
 
     def __init__(self, keys: list, data: Union[dict, np.array]):
-        """
-        Initializes the container
+        """ Initializes the container
 
-        if keys is not a list- try to convert it to one
-            saves list of keys in object, ids for data
+        Arguments:
+        keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
+        data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
         """
         if not isinstance(keys, list):
-            keys = list(keys)
+            keys = list(keys)   # creates list with keys
         self._keys = keys.copy()
         if isinstance(data, np.matrix):
             self.matrix = np.array(data, dtype=np.float64)
@@ -52,21 +45,17 @@ class DictLikeMatrixWrapper():
     def __getitem__(self, key: str) -> int:
         """
         get all values associated with a key, ex: all values of 'i'
-            row: creates list from a row of matrix
-        if: row list contains 1 value, returns that value (non-vectorized)
-        else: returns entire row/list (vectorized case)
         """
-        row = self.matrix[self._keys.index(key)]
-        if len(row) == 1:
+        row = self.matrix[self._keys.index(key)]    #creates list from a row of matrix
+        if len(row) == 1:   #list contains 1 value, returns that value (non-vectorized)
             return row[0]
-        return row
+        return rows #returns entire row/list (vectorized case)
 
     def __setitem__(self, key: str, value: int) -> None:
         """
         sets a row at the key given
-            the int value index for the key given
         """
-        index = self._keys.index(key)
+        index = self._keys.index(key)   #the int value index for the key given
         self.matrix[index] = np.atleast_1d(value)
 
     
@@ -98,15 +87,10 @@ class DictLikeMatrixWrapper():
     def __eq__(self, other: "DictLikeMatrixWrapper") -> bool:
         """
         Compares two DictLikeMatrixWrappers (i.e. *Containers) or a DictLikeMatrixWrapper and a dictionary
-            if: checks that the list of keys for each matrix match
-            bools: 
-                list_key_check: checks that the list of keys for each matrix are equal
-                matrix_check: checks to see that each row matches
-            returns bool value
         """
-        if isinstance(other, dict):
-            list_key_check = (list(self.keys()) == list(other.keys()))
-            matrix_check = (self.matrix == np.array([[other[key]] for key in self._keys])).all()
+        if isinstance(other, dict): #checks that the list of keys for each matrix match
+            list_key_check = (list(self.keys()) == list(other.keys()))  #checks that the list of keys for each matrix are equal
+            matrix_check = (self.matrix == np.array([[other[key]] for key in self._keys])).all()    #checks to see that each row matches
             return list_key_check and matrix_check
         list_key_check = self.keys() == other.keys()
         matrix_check = (self.matrix == other.matrix).all()
@@ -147,35 +131,28 @@ class DictLikeMatrixWrapper():
     def values(self) -> np.array:
         """
         returns array of matrix values
[…]
