### 2023-02-24 c54ae24
Cleaning up TrajGen new format for Matteo to work on controller
