+++ src/prog_models/aux_fcns/traj_gen.py
@@ -9,7 +9,7 @@ from .traj_gen_utils import route, trajectory
 from prog_models.exceptions import ProgModelInputException
 
 
-def trajectory_gen_fcn(waypoints_info, **params):
+def trajectory_gen_fcn(waypoints_info, vehicle, **params):
 
     parameters = {  # Set to defaults
         # Flight information
@@ -17,21 +17,23 @@ def trajectory_gen_fcn(waypoints_info, **params):
         'flight_plan': None,
 
         # Simulation parameters:
-        'dt': 0.1, 
-        'gravity': 9.81,
-        'cruise_speed': None, 
-        'ascent_speed': None, 
-        'descent_speed': None,  
-        'landing_speed': None, 
-        'hovering_time': 0.0,
-        'takeoff_time': 0.0, 
-        'landing_time': 0.0, 
+        'dt': vehicle.parameters['dt'], 
+        'gravity': vehicle.parameters['gravity'],
+        'cruise_speed': vehicle.parameters['cruise_speed'],
+        'ascent_speed': vehicle.parameters['ascent_speed'],
+        'descent_speed': vehicle.parameters['descent_speed'],
+        'landing_speed': vehicle.parameters['landing_speed'],
+        'hovering_time': vehicle.parameters['hovering_time'],
+        'takeoff_time': vehicle.parameters['takeoff_time'], 
+        'landing_time': vehicle.parameters['landing_time'],
         'waypoint_weights': 10.0, 
         'adjust_eta': None, 
         'nurbs_basis_length': 2000, 
         'nurbs_order': 4, 
 
-        # Vehicle parameters:
+        # Vehicle parameters: 
+        # TODO: These should be consistent with vehicle model, so should come from vehicle_info.parameters, 
+        # but these values aren't currently in parameters for the vehicle model 
         'vehicle_max_speed': 15.0, #### NEED TO FIX THIS 
         'vehicle_max_roll': 0.7853981633974483, ### NEED TO FIX 
         'vehicle_max_pitch': 0.7853981633974483, #### NEED TO FIX 
@@ -39,7 +41,12 @@ def trajectory_gen_fcn(waypoints_info, **params):
         # 'vehicle_payload': 0.0,
     }
 
+    # Update parameters with any user-defined parameters 
+        # TODO: This is necessary so the user can configure nurbs_order, etc. 
+        # But this will override any vehicle_info parameters, if the user defines them 
+        # Either need to include warnings if the user overrides them, or need to do this differently 
     parameters.update(params)
+
     if isinstance(waypoints_info, dict):
         parameters['flight_plan'] = waypoints_info
     elif isinstance(waypoints_info, str):
@@ -132,5 +139,4 @@ def trajectory_gen_fcn(waypoints_info, **params):
     else: 
         raise ProgModelInputException("Reference trajectory format is not yet configured for this vehicle type.")
 
-    debug = 1
     return x_ref
