### 2023-02-25 b80468f
cleanup: test_sim_result.py
