+++ tests/test_sim_result.py
@@ -12,17 +12,82 @@ from prog_models.utils.containers import DictLikeMatrixWrapper
 
 
 class TestSimResult(unittest.TestCase):
+    # unit test variables
+    time_data = [0, 1, 2, 3, 4]
+    time_data2 = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
+    time_data_ext = [0, 1, 2, 3, 4, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
+    result_data = [
+        {'a': 0.0, 'b': 0.0},
+        {'a': 2.5, 'b': 2.5},
+        {'a': 5.0, 'b': 5.0},
+        {'a': 7.5, 'b': 7.5},
+        {'a': 10.0, 'b': 10.0}
+    ]
+    result2_data = [
+        {'a': 0, 'b': 0},
+        {'a': 5, 'b': 5},
+        {'a': 10, 'b': 10},
+        {'a': 15, 'b': 15},
+        {'a': 20, 'b': 20},
+        {'a': 25, 'b': 25},
+        {'a': 30, 'b': 30},
+        {'a': 35, 'b': 35},
+        {'a': 40, 'b': 40},
+        {'a': 45, 'b': 45}
+    ]
+    result_data_ext = [
+        {'a': 0.0, 'b': 0.0},
+        {'a': 2.5, 'b': 2.5},
+        {'a': 5.0, 'b': 5.0},
+        {'a': 7.5, 'b': 7.5},
+        {'a': 10.0, 'b': 10.0},
+        {'a': 0, 'b': 0},
+        {'a': 5, 'b': 5},
+        {'a': 10, 'b': 10},
+        {'a': 15, 'b': 15},
+        {'a': 20, 'b': 20},
+        {'a': 25, 'b': 25},
+        {'a': 30, 'b': 30},
+        {'a': 35, 'b': 35},
+        {'a': 40, 'b': 40},
+        {'a': 45, 'b': 45}
+    ]
+
+    test_rm_clr_data = [
+        {'a': 0.0, 'b': 0},
+        {'a': 2.5, 'b': 5},
+        {'a': 5, 'b': 10},
+        {'a': 7.5, 'b': 15},
+        {'a': 10.0, 'b': 20}
+    ]
+
+    test_rm_clr_data2 = [
+        {'a': 0.0, 'b': 0},
+        {'a': 2.5, 'b': 5},
+        {'a': 7.5, 'b': 15}
+    ]
+
+    test_lazy_fcn = [
+        {'a': 0.0, 'b': 0},
+        {'a': 5.0, 'b': 10},
+        {'a': 10.0, 'b': 20},
+        {'a': 15.0, 'b': 30},
+        {'a': 20.0, 'b': 40}
+    ]
+
+    num_elem_5 = 5
+    num_elem_10 = 10
+
     def setUp(self):
         # set stdout (so it wont print)
         sys.stdout = StringIO()
 
     def tearDown(self):
         sys.stdout = sys.__stdout__
-    
+
     def test_sim_result(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
+        time = list(range(self.num_elem_5))
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(self.num_elem_5)]
         result = SimResult(time, state)
         self.assertListEqual(list(result), state)
         self.assertListEqual(result.times, time)
@@ -31,76 +96,106 @@ class TestSimResult(unittest.TestCase):
             self.assertEqual(result[i], state[i])
 
         try:
-            tmp = result[NUM_ELEMENTS]
+            tmp = result[self.num_elem_5]
[…]
