+++ tests/test_sim_result.py
@@ -12,7 +12,7 @@ from prog_models.utils.containers import DictLikeMatrixWrapper
 
 
 class TestSimResult(unittest.TestCase):
-    # unit test variables
+    # UNIT TEST VARIABLES
     time_data = [0, 1, 2, 3, 4]
     time_data2 = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     time_data_ext = [0, 1, 2, 3, 4, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
@@ -52,7 +52,6 @@ class TestSimResult(unittest.TestCase):
         {'a': 40, 'b': 40},
         {'a': 45, 'b': 45}
     ]
-
     test_rm_clr_data = [
         {'a': 0.0, 'b': 0},
         {'a': 2.5, 'b': 5},
@@ -60,13 +59,11 @@ class TestSimResult(unittest.TestCase):
         {'a': 7.5, 'b': 15},
         {'a': 10.0, 'b': 20}
     ]
-
     test_rm_clr_data2 = [
         {'a': 0.0, 'b': 0},
         {'a': 2.5, 'b': 5},
         {'a': 7.5, 'b': 15}
     ]
-
     test_lazy_fcn = [
         {'a': 0.0, 'b': 0},
         {'a': 5.0, 'b': 10},
@@ -75,8 +72,12 @@ class TestSimResult(unittest.TestCase):
         {'a': 20.0, 'b': 40}
     ]
 
+    # number of elements
     num_elem_5 = 5
     num_elem_10 = 10
+    # lists of time data
+    time_ne5 = list(range(num_elem_5))
+    time_ne10 = list(range(num_elem_10))
 
     def setUp(self):
         # set stdout (so it wont print)
@@ -86,13 +87,15 @@ class TestSimResult(unittest.TestCase):
         sys.stdout = sys.__stdout__
 
     def test_sim_result(self):
-        time = list(range(self.num_elem_5))
+        """
+            tests SimResult object
+        """
         state = [{'a': i * 2.5, 'b': i * 5} for i in range(self.num_elem_5)]
-        result = SimResult(time, state)
+        result = SimResult(self.time_ne5, state)
         self.assertListEqual(list(result), state)
-        self.assertListEqual(result.times, time)
+        self.assertListEqual(result.times, self.time_ne5)
         for i in range(5):
-            self.assertEqual(result.time(i), time[i])
+            self.assertEqual(result.time(i), self.time_ne5[i])
             self.assertEqual(result[i], state[i])
 
         try:
@@ -108,21 +111,26 @@ class TestSimResult(unittest.TestCase):
             pass
 
     def test_pickle(self):
-        time = list(range(self.num_elem_5))
+        """
+            tests SimResult being pickled
+        """
         state = [{'a': i * 2.5, 'b': i * 5} for i in range(self.num_elem_5)]
-        result = SimResult(time, state)
+        result = SimResult(self.time_ne5, state)
         pickle.dump(result, open('model_test.pkl', 'wb'))
         result2 = pickle.load(open('model_test.pkl', 'rb'))
         self.assertEqual(result, result2)
 
     def test_extend(self):
-        # Creating two result objects
-        time = list(range(self.num_elem_5))
+        """
+            tests extend in sim_result
+        """
+        # Creating result objects
         state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(self.num_elem_5)]
-        result = SimResult(time, state)
-        time = list(range(self.num_elem_10))
-        state = [{'a': i * 5, 'b': i * 5} for i in range(self.num_elem_10)]
-        result2 = SimResult(time, state)
[…]
