+++ src/prog_models/models/test_models/linear_models.py
@@ -15,6 +15,7 @@ class FNoneNoEventStateLM(LinearModel):
     A = np.array([[0]])
     B = np.array([[1]])
     C = np.empty((0,1))
+    D = np.empty((0,1))
     F = None
 
     default_parameters = {
@@ -35,6 +36,7 @@ class OneInputNoOutputNoEventLM(LinearModel):
     A = np.array([[0]])
     B = np.array([[1]])
     C = np.empty((0,1))
+    D = np.empty((0,1))
     F = np.empty((0,1))
 
     default_parameters = {
@@ -56,6 +58,7 @@ class OneInputOneOutputNoEventLM(LinearModel):
     A = np.array([[0]])
     B = np.array([[1]])
     C = np.array([[1]])
+    D = np.array([[1]])
     F = np.empty((0,1))
 
     default_parameters = {
@@ -97,6 +100,7 @@ class OneInputNoOutputTwoEventLM(LinearModel):
     A = np.array([[0]])
     B = np.array([[1, 0.5]])
     C = np.empty((0,1))
+    D = np.empty((0,1))
     F = np.array([[-0.1], [-0.2]])
     G = np.array([[1], [1]])
 
