+++ src/prog_models/models/test_models/linear_models.py
@@ -152,13 +152,3 @@ class TwoInputNoOutputTwoEventLM(LinearModel):
             'x1': 0
         }
     }
-
-class ThreeStatesTwoB(LinearModel):
-    inputs = ['u1']
-    states = ['x1']
-    A = np.array([[0]])
-    B = np.array([[1, 2]])
-    C = np.empty((0,1))
-    D = np.empty((0,1))
-
-    F = None
