+++ tests/test_base_models.py
@@ -68,42 +68,6 @@ def derived_callback3(config):
         'p4': -2 * config['p2'], 
     }
 
-class LinearThrownObject(LinearModel):
-    inputs = [] 
-    states = ['x', 'v']
-    outputs = ['x']
-    events = ['impact']
-
-    A = np.array([[0, 1], [0, 0]])
-    E = np.array([[0], [-9.81]])
-    C = np.array([[1, 0]])
-    F = None # Will override method
-
-    default_parameters = {
-        'thrower_height': 1.83,  # m
-        'throwing_speed': 40,  # m/s
-        'g': -9.81  # Acceleration due to gravity in m/s^2
-    }
-
-    def initialize(self, u=None, z=None):
-        return self.StateContainer({
-            'x': self.parameters['thrower_height'],  # Thrown, so initial altitude is height of thrower
-            'v': self.parameters['throwing_speed']  # Velocity at which the ball is thrown - this guy is a professional baseball pitcher
-            })
-    
-    def threshold_met(self, x):
-        return {
-            'falling': x['v'] < 0,
-            'impact': x['x'] <= 0
-        }
-
-    def event_state(self, x): 
-        x_max = x['x'] + np.square(x['v'])/(-self.parameters['g']*2) # Use speed and position to estimate maximum height
-        return {
-            'falling': np.maximum(x['v']/self.parameters['throwing_speed'],0),  # Throwing speed is max speed
-            'impact': np.maximum(x['x']/x_max,0) if x['v'] < 0 else 1  # 1 until falling begins, then it's fraction of height
-        }
-
 class MockModelWithDerived(MockProgModel):
     param_callbacks = {
             'p1': [derived_callback],
@@ -571,11 +535,11 @@ class TestModels(unittest.TestCase):
             return {'i1': 1, 'i2': 2.1}
         (times, inputs, states, outputs, event_states) = m.simulate_to(6, load, {'o1': 0.8}, **{'dt': 0.5, 'save_freq': 1.0})
         named_results = m.simulate_to(6, load, {'o1': 0.8}, **{'dt': 0.5, 'save_freq': 1.0})
-        self.assertEquals(times, named_results.times)
-        self.assertEquals(inputs, named_results.inputs)
-        self.assertEquals(states, named_results.states)
-        self.assertEquals(outputs, named_results.outputs)
-        self.assertEquals(event_states, named_results.event_states)
+        self.assertEqual(times, named_results.times)
+        self.assertEqual(inputs, named_results.inputs)
+        self.assertEqual(states, named_results.states)
+        self.assertEqual(outputs, named_results.outputs)
+        self.assertEqual(event_states, named_results.event_states)
         
     def test_next_time_fcn(self):
         m = MockProgModel(process_noise = 0.0)
