+++ src/prog_models/__init__.py
@@ -3,9 +3,10 @@
 
 # PrognosticsModel must be first, since the others build on this
 from prog_models.prognostics_model import PrognosticsModel
-from .ensemble_model import EnsembleModel
-from .composite_model import CompositeModel
-from .linear_model import LinearModel
-from .exceptions import ProgModelException, ProgModelInputException, ProgModelTypeError
+from prog_models.ensemble_model import EnsembleModel
+from prog_models.composite_model import CompositeModel
+from prog_models.linear_model import LinearModel
+from prog_models.models.thrown_object import LinearThrownObject
+from prog_models.exceptions import ProgModelException, ProgModelInputException, ProgModelTypeError
 
 __version__ = '1.5.0.pre'
