### 2023-03-01 2b40a5f
deepcopy not working as intended.
