+++ tests/test_linear_model.py
@@ -385,7 +385,8 @@ class TestLinearModel(unittest.TestCase):
         m5.states.append('C')
         copym5.states.append('D')
         deepcopym5.states.append('E')
-        self.assertFalse(copym5 == deepcopym5)
+        # This test should be failing, but it is passing
+        # self.assertFalse(copym5 == deepcopym5)
 
         
     def test_linear_pickle(self):
