+++ src/prog_models/prognostics_model.py
@@ -133,7 +133,7 @@ class PrognosticsModel(ABC):
 
     def __setstate__(self, params : dict) -> None:
         # This method is called when depickling and in construction. It builds the model from the parameters
-
+        
         if not hasattr(self, 'inputs'):
             self.inputs = []
         self.n_inputs = len(self.inputs)
@@ -162,8 +162,21 @@ class PrognosticsModel(ABC):
             self.performance_metric_keys = []
         self.n_performance = len(self.performance_metric_keys)
 
+        # Make class attributes instance attributes to ensure copy and deepcopy function as intended.
+        # Without this, deepcopy would not create a new copy these attributes.
+
+        self.states = self.states.copy()
+        self.inputs = self.inputs.copy()
+        self.outputs = self.outputs.copy()
+        self.events = self.events.copy()
+        self.performance_metric_keys = self.performance_metric_keys.copy()
+        self.param_callbacks = self.param_callbacks.copy()
+        self.state_limits = self.state_limits.copy()
+        self.default_parameters = self.default_parameters.copy()
+
         # Setup Containers
         # These containers should be used instead of dictionaries for models that use the internal matrix state
+
         states = self.states
 
         class StateContainer(DictLikeMatrixWrapper):
