+++ tests/__main__.py
@@ -14,7 +14,7 @@ from tests.test_surrogates import main as surrogates_main
 from tests.test_data_model import main as lstm_main
 from tests.test_direct import main as direct_main
 from tests.test_linear_model import main as linear_main
-from tests.test_composite import main as composte_main
+from tests.test_composite import main as composite_main
 
 from io import StringIO
 import matplotlib.pyplot as plt
@@ -119,10 +119,10 @@ if __name__ == '__main__':
     except Exception:
         was_successful = False
     
-    # try:
-    #     composite_main()
-    # except Exception:
-    #     was_successful = False
+    try:
+        composite_main()
+    except Exception:
+        was_successful = False
 
     if not was_successful:
         raise Exception("Failed test")
