### 2023-03-02 2c53ca9
removed .copy() from prognostic_models
