+++ src/prog_models/composite_model.py
@@ -8,7 +8,6 @@ from . import PrognosticsModel
 DIVIDER = '.'
 
 
-#Can connect states and outputs. Circuilt Connections (output -> Input, State of the Second -> Event of the first)
 class CompositeModel(PrognosticsModel):
     """
     A CompositeModel is a PrognosticsModel that is composed of multiple PrognosticsModels. This is a tool for modeling system-of-systems. I.e., interconnected systems, where the behavior and state of one system effects the state of another system. The composite prognostics models are connected using defined connections between the output or state of one model, and the input of another model. The resulting CompositeModel behaves as a single model. 
