+++ tests/test_base_models.py
@@ -14,7 +14,7 @@ sys.path.append(join(dirname(__file__), ".."))
 from prog_models import ProgModelTypeError, ProgModelInputException, ProgModelException, PrognosticsModel, CompositeModel, LinearModel
 from prog_models.models import ThrownObject, BatteryElectroChemEOD
 from prog_models.models.test_models.linear_models import (OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
-from prog_models.models.test_models.linear_thrown_object import (LinearThrownObject, LinearThrownDiffThrowingSpeed, LinearThrownObjectUpdatedInitalizedMethod, LinearThrownObjectDiffDefaultParameters)
+from prog_models.models.test_models.linear_thrown_object import (LinearThrownObject, LinearThrownDiffThrowingSpeed, LinearThrownObjectUpdatedInitalizedMethod, LinThrwnObjectDiffDefaultParams)
 
 class MockModel():
     states = ['a', 'b', 'c', 't']
@@ -1322,7 +1322,7 @@ class TestModels(unittest.TestCase):
         self.assertFalse(m1.parameters == m3.parameters)
         self.assertFalse(m3.parameters == m1.parameters) # Checking both directions 
 
-        m4 = LinearThrownObjectDiffDefaultParameters() # Model with an extra default parameter.
+        m4 = LinThrwnObjectDiffDefaultParams() # Model with an extra default parameter.
 
         self.assertFalse(m1.parameters == m4.parameters)
         self.assertFalse(m4.parameters == m1.parameters) # checking both directions
@@ -1333,7 +1333,7 @@ class TestModels(unittest.TestCase):
         self.assertFalse(m5.parameters == m1.parameters) 
 
         self.assertTrue(m1.parameters == m2.parameters) # Checking to see previous equal statements stay the same
-        self.assertTrue(m2.parameters == m1.parameters) 
+        self.assertTrue(m2.parameters == m1.parameters)
 
 # This allows the module to be executed directly
 def run_tests():
