+++ src/prog_models/models/test_models/linear_thrown_object.py
@@ -12,7 +12,6 @@ class LinearThrownObject(LinearModel):
     outputs = ['x']
     events = ['impact']
 
-    A = np.array([[0, 1], [0, 0]])
     C = np.array([[1, 0]])
     E = np.array([[0], [-9.81]])
     F = None # Will override method
@@ -101,7 +100,7 @@ class LinearThrownObjectUpdatedInitalizedMethod(LinearThrownObject):
             'x': self.parameters['throwing_speed'],  # Thrown, so initial altitude is height of thrower
             })
 
-class LinearThrownObjectDiffDefaultParameters(LinearThrownObject):
+class LinearThrownObjectDiffDefaultParams(LinearThrownObject):
     inputs = [] 
     states = ['x', 'v']
     outputs = ['x']
@@ -119,3 +118,13 @@ class LinearThrownObjectDiffDefaultParameters(LinearThrownObject):
         'g': -9.81,  # Acceleration due to gravity in m/s^2
         'x': 1111
     }
+
+class LinearThrownObjectFourStates(LinearThrownObject):
+    inputs = []
+    states = ['x', 'v', 'y' ,'z']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1, 2, 3], [0, 1, 2, 3], [0, 1, 2, 3], [0, 1, 2, 3]])
+    C = np.array([[0, 1, 2, 3]])
+    E = np.array([[0], [1], [2], [3]])
