+++ tests/test_linear_model.py
@@ -150,10 +150,10 @@ class TestLinearModel(unittest.TestCase):
         with self.assertRaises(AttributeError):
             m.C = np.array([[1]]) # less column values per row
             m.matrixCheck()
-        with self.assertRaises(AttributeError):
+        with self.assertRaises(AttributeError): 
             m.C = np.array([[0, 0], [1, 1], [2, 2]]) # extra row
             m.matrixCheck()
-        with self.assertRaises(AttributeError):
+        with self.assertRaises(AttributeError): 
             m.C = np.array([[]]) # less row 
             m.matrixCheck()
         m.C = np.array([[1, 0]])
@@ -345,6 +345,12 @@ class TestLinearModel(unittest.TestCase):
         m.matrixCheck()
         m.G = None # sets to Default Value
         m.matrixCheck()
+
+        # Error Demonstration
+        mTest = LinearThrownObjectFourStates()
+        with self.assertRaises(AttributeError):
+            mTest.B = np.array([[0], [1], [2], [3]])
+
 
         # Error Demonstration
         mTest = LinearThrownObjectFourStates()
