### 2023-03-13 7fdc547
Adding serialization tests to __main__ for tests
