+++ src/prog_models/models/battery_circuit_df.py
@@ -1,0 +1,208 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the
+# National Aeronautics and Space Administration.  All Rights Reserved.
+
+from math import inf
+import numpy as np
+
+from .. import PrognosticsModel
+
+
+class BatteryCircuit(PrognosticsModel):
+    """
+    Vectorized prognostics :term:`model` for a battery, represented by an equivilant circuit model as described in the following paper:
+    `M. Daigle and S. Sankararaman, "Advanced Methods for Determining Prediction Uncertainty in Model-Based Prognostics with Application to Planetary Rovers," Annual Conference of the Prognostics and Health Management Society 2013, pp. 262-274, New Orleans, LA, October 2013. https://papers.phmsociety.org/index.php/phmconf/article/view/2253`
+    
+    :term:`Events<event>`: (1)
+        EOD: End of Discharge
+    
+    :term:`Inputs/Loading<input>`: (1)
+        i: Current draw on the battery
+
+    :term:`States<state>`: (4)
+        | tb : Battery Temperature (K)
+        | qb : Charge stored in Capacitor Cb of the equivalent circuit model
+        | qcp : Charge stored in Capacitor Ccp of the equivalent circuit model
+        | qcs : Charge stored in Capacitor Ccs of the equivalent circuit model
+
+    :term:`Outputs<output>`: (2)
+        | t: Temperature of battery (K)
+        | v: Voltage supplied by battery
+
+    Keyword Args
+    ------------
+        process_noise : Optional, float or dict[str, float]
+          :term:`Process noise<process noise>` (applied at dx/next_state). 
+          Can be number (e.g., .2) applied to every state, a dictionary of values for each 
+          state (e.g., {'x1': 0.2, 'x2': 0.3}), or a function (x) -> x
+        process_noise_dist : Optional, str
+          distribution for :term:`process noise` (e.g., normal, uniform, triangular)
+        measurement_noise : Optional, float or dict[str, float]
+          :term:`Measurement noise<measurement noise>` (applied in output eqn).
+          Can be number (e.g., .2) applied to every output, a dictionary of values for each
+          output (e.g., {'z1': 0.2, 'z2': 0.3}), or a function (z) -> z
+        measurement_noise_dist : Optional, str
+          distribution for :term:`measurement noise` (e.g., normal, uniform, triangular)
+        V0 : float
+          Nominal Battery Voltage
+        Rp : float
+          Battery Parasitic Resistance
+        qMax : float
+          Maximum Charge
+        CMax : float
+          Maximum Capacity
+        VEOD : float
+          End of Discharge Voltage Threshold
+        Cb0 : float 
+          Battery Capacity Parameter
+        Cbp0 : float
+          Battery Capacity Parameter
+        Cbp1 : float
+          Battery Capacity Parameter
+        Cbp2 : float
+          Battery Capacity Parameter
+        Cbp3 : float
+          Battery Capacity Parameter
+        Rs : float
+          R-C Pair Parameter
+        Cs : float
+          R-C Pair Parameter
+        Rcp0 : float
+          R-C Pair Parameter
+        Rcp1 : float
+          R-C Pair Parameter
+        Rcp2 : float
+          R-C Pair Parameter
+        Ccp : float
+          R-C Pair Parameter
[…]
