### 2023-03-13 06b4420
Moved tests, created thrownobjectmodels class, improved testing.
moving laptops
