+++ tests/test_estimate_params.py
@@ -1,0 +1,307 @@
+from copy import deepcopy
+import io
+import numpy as np
+from os.path import dirname, join
+import pickle
+import sys
+import unittest
+
+# This ensures that the directory containing ProgModelTemplate is in the python search directory
+sys.path.append(join(dirname(__file__), ".."))
+
+from prog_models import *
+from prog_models.models import *
+from prog_models.models.test_models.linear_models import (
+    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
+from prog_models.models.thrown_object import LinearThrownObject
+from prog_models.models.test_models.thrown_object_models import defaultParams
+
+class TestEstimateParams(unittest.TestCase):
+    def test_estimate_params_works(self):
+        m = ThrownObject()
+        results = m.simulate_to_threshold(save_freq=0.5)
+        data = [(results.times, results.inputs, results.outputs)]
+        gt = m.parameters.copy()
+
+        self.assertTrue(m.parameters == gt)
+        self.assertEqual(m.parameters, gt)
+
+        # Now lets reset some parameters
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        keys = ['thrower_height', 'throwing_speed', 'g']
+        m.estimate_params(data, keys)
+        for key in keys: # using assert not equal also works.
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+        
+        # self.assertTrue(m.parameters == gt)
+
+
+    def test_estimate_params(self):
+        # Some things to consider, dictionary vs tuple. The function is accepting both it would seem.
+        # Personal opinion on which data type to use. Tuples make the most amount of sense.
+
+        # gt copy works as intended. Going to a different object location
+
+        # Potentially include something that focuses on specifically the key value pairs between gt and m.parameters? mayb not
+
+        #MAKE SURE IT WORKS WITH AN ARRAY AND A NUMPY ARRAY.
+
+        # Create a model that will easily raise an Exception with calc_error.
+
+        # Write a test that fails with the examples bug.
+        m = ThrownObject()
+        results = m.simulate_to_threshold(save_freq=0.5)
+        data = [(results.times, results.inputs, results.outputs)]
+        gt = m.parameters.copy()
+
+
+        # Now lets reset some parameters
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        keys = ['thrower_height', 'throwing_speed', 'g']
+        m.estimate_params(data, keys)
+
+        for key in keys: # using assert not equal also works.
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+
+        # Now with limits that dont include the true values
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
+        # Not the most accurate way of seeing when these values are not accurate.
+        for key in keys:
+            self.assertNotAlmostEqual(m.parameters[key], gt[key], 1, "Limits of the Bounds do not include the true values")
+
+        # Now with limits that do include the true values
+        m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
+        for key in keys:
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2, "Limits of the Bounds do not include the true values")
+        
+        # Lower Bound Greater than Upper Bound
[…]

[…]
