### 2023-03-14 35163ef
Creating a new battery model for testing.
