### 2023-03-15 #5099 @tbsexton
Definitely, just had some obligations come up recently but I should get through a good amount this week
