+++ tests/test_composite.py
@@ -90,7 +90,7 @@ class TestCompositeModel(unittest.TestCase):
         self.assertSetEqual(m_composite.inputs, {'OneInputOneOutputNoEventLM.u1', 'OneInputOneOutputNoEventLM_2.u1'})
         self.assertSetEqual(m_composite.outputs, {'OneInputOneOutputNoEventLM.z1', 'OneInputOneOutputNoEventLM_2.z1'})
         self.assertSetEqual(m_composite.events, set())
-        self.assertSetEqual(m_composite.performance_metric_keys, set(), "Shouldn't have any performance metrics")
+        self.assertSetEqual(m_composite.performance_metric_keys, set(), "Shouldn't have any performance at metrics")
 
         x0 = m_composite.initialize()
         self.assertSetEqual(set(x0.keys()), {'OneInputOneOutputNoEventLM_2.x1', 'OneInputOneOutputNoEventLM.x1'})
@@ -218,7 +218,7 @@ class TestCompositeModel(unittest.TestCase):
         self.assertSetEqual(m_composite.events, set())
         x0 = m_composite.initialize()
         z = m_composite.output(x0)
-        self.assertSetEqual(set(z.keys()), {'OneInputOneOutputNoEventLM_2.z1', })
+        self.assertSetEqual(set(z.keys()), {'OneInputOneOutputNoEventLM_2.z1'})
 
         # With Names
         m_composite = CompositeModel([('m1', m1), ('m2', m2)], connections=[('m1.x1', 'm2.u1')])
