+++ tests/test_estimate_params.py
@@ -14,7 +14,7 @@ from prog_models.models import *
 from prog_models.models.test_models.linear_models import (
     OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
 from prog_models.models.thrown_object import LinearThrownObject
-from prog_models.models.test_models.thrown_object_models import defaultParams
+from prog_models.models.test_models.thrown_object_models import defaultParams, wrongInputStorage, wrongTimeValues
 
 class TestEstimateParams(unittest.TestCase):
     def test_estimate_params_works(self):
@@ -65,16 +65,33 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
+        m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333), (0, 3)))
+        # value1 = m.calc_error(results.times, results.inputs, results.outputs)
+        # m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
+        # value5 = m.calc_error(results.times, results.inputs, results.outputs)
+        # m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333), (0, 3)))
+        # sanityCheck = m.calc_error(results.times, results.inputs, results.outputs)
+
+        # Need at least one data point
+        with self.assertRaises(ValueError):
+            m.estimate_params(times=[], inputs=[], outputs=[])
+        with self.assertRaises(ValueError):
+            m.estimate_params(times=None, inputs=None, output=None)
+        with self.assertRaises(ValueError):
+            m.estimate_params(times='', inputs='', outputs='')
+
         # Now with limits that dont include the true values
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
+        value2 = m.calc_error(results.times, results.inputs, results.outputs)
         # Not the most accurate way of seeing when these values are not accurate.
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1, "Limits of the Bounds do not include the true values")
 
         # Now with limits that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
+        value3 = m.calc_error(results.times, results.inputs, results.outputs)
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2, "Limits of the Bounds do not include the true values")
         
@@ -85,7 +102,7 @@ class TestEstimateParams(unittest.TestCase):
 
         # Testing all bounds are incorrect
         with self.assertRaises(ValueError):
-            m.estimate_params(data, keys, bounds=((30, 7), (42, 20), (0, -20)))
+            m.estimate_params(data, keys, bounds=((30, 7), (40, 20), (0, -20)))
 
         #Implememnt different variations of lists and tuples and see if they work as intended
         # Try incomplete list:
@@ -147,11 +164,7 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(ValueError):
             # Item is a list of length 1
             m.estimate_params(data, keys, bounds={'g': [7]})
-        # with self.assertRaises(ValueError):
-            # This does not change our current values
-            # Defining two values
-            # Uses the second defined value
[…]
