### 2023-03-21 cac8a48
some progress on modular controller classes
