+++ src/prog_models/loading_fcns/controllers.py
@@ -1,0 +1,232 @@
+"""
+Controllers
+"""
+
+# IMPORTS
+# ========
+from utilities.imports_ import np
+
+
+# LQR general function
+# ====================
+def lqr_fn(A, B, Q, R):
+    """
+    Calculate control matrix K using LQR
+
+    The function estimates the gain matrix associated with a linear quadratic regulator(LQR) via Hamiltonian matrix.
+    The function assumes a linearized system:
+
+    dx / dt = Ax + Bu
+    y = Cx + Du
+
+    The closed loop uses - K * x as input dx / dt = (A - B * K) * x
+
+    :param A:       Nx x Nx state matrix, linearized w.r.t. current attitude and thrust
+    :param B:       Nu x Nx input matrix, linearized w.r.t. current attitude
+    :param Q:       Nx x Nx state cost matrix, higher the values of the (i,i)-th element is the more aggresive the controller is w.r.t. that variable
+    :param R:       Nx x Nu input cost matrix, higher values imply higher cost of producing that input (used to limit the amount of power)
+    :return K:      Nx x Nu, control gain matrix
+    :return E:      Nx x 1, eigenvalues of the matrix A - B * K
+    """
+    n     = A.shape[0]
+    R_inv = np.linalg.inv(R)
+
+    # --------- Generate hamiltonian matrix ------------ #
+    HM1 = A
+    HM2 = - np.dot( np.dot( B, R_inv), B.T )
+    HM3 = - Q
+    HM4 = - A.T
+
+    HM12 = np.concatenate((HM1, HM2), axis=1)
+    HM34 = np.concatenate((HM3, HM4), axis=1)
+    HM   = np.concatenate((HM12, HM34), axis=0)
+
+    # --------- Extract eigevectors whose eigenvalues have real part < 0 ------------ #
+    eig_val, eig_vec = np.linalg.eig(HM)
+    e = len(eig_val)
+    V_ = np.zeros((e, 1))
+    for ii in range(e):
+        if np.real(eig_val[ii]) < 0:
+            V_ = np.concatenate((V_, eig_vec[:, ii].reshape((e,1))), axis=1)
+    V_ = V_[:, 1:]
+
+    # --------- Partition matrix of resulting eigenvectors ------------ #
+    X = V_[:n, :n] # first half (row-wise)
+    Y = V_[n:, :n] # second half (row-wise)
+
+    # ----------- Estimate control gain ----------------- #
+    K = np.dot( np.dot( np.dot(R_inv, B.T), Y ), np.linalg.inv(X) )
+    K = np.real(K) # some spuriorus imaginary parts (1e-13) sometimes remain in the matrix. we manually remove them
+
+    # Calculate the eigenvalues of the matrix A - B*K
+    E, _ = np.linalg.eig( A - np.dot(B, K) )
+    return K, E
+
+    
+# CONTROLLERS: FULL STATE VECTOR CONTROL  (i.e., no powertrain)
+# ==============================================================
+class LQR():
+    """ Linear Quadratic Regulator"""
+    def __init__(self, n_states, n_inputs, Q=None, R=None) -> None:
+        self.type = 'LQR'
+        if Q is None:       Q = np.eye(n_states)
+        if R is None:       R = np.eye(n_inputs)
+
+        self.Q = Q
+        self.R = R
+        pass
+    
+    def compute_gain(self, A, B):
+        """ Compute controller gain given state of the system described by linear model A, B"""
+        self.K, self.E = lqr_fn(A, B, self.Q, self.R)
+        return self.K, self.E
+
+    def compute_input(self, gain, state_error):
+        """ Compute system input given the controller gain 'gain' and the error w.r.t. the reference state 'error' """
[…]
