+++ tests/test_estimate_params.py
@@ -65,7 +65,8 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-        m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333), (0, 3)))
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
+        # m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333), (0, 3)))
         # value1 = m.calc_error(results.times, results.inputs, results.outputs)
         # m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
         # value5 = m.calc_error(results.times, results.inputs, results.outputs)
@@ -79,6 +80,8 @@ class TestEstimateParams(unittest.TestCase):
             m.estimate_params(times=None, inputs=None, output=None)
         with self.assertRaises(ValueError):
             m.estimate_params(times='', inputs='', outputs='')
+        with self.assertRaises(ValueError):
+            m.estimate_params(times=[[]], inputs=[[]], outputs=[[]])
 
         # Now with limits that dont include the true values
         m.parameters['thrower_height'] = 1.5
@@ -88,6 +91,16 @@ class TestEstimateParams(unittest.TestCase):
         # Not the most accurate way of seeing when these values are not accurate.
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1, "Limits of the Bounds do not include the true values")
+        
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
+        check = m.calc_error(results.times, results.inputs, results.outputs)
+        m.estimate_params(data, keys, bounds=((1.243, 4.0), (19.13212, 39.12301), (-21.123, 2.000991)))
+        check2= m.calc_error(results.times, results.inputs, results.outputs)
+
+        # calc_error is a much bigger value compared to other values. Not including a good range?
+        # Should not affect other calls...
+        # m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333), (0, 3)))
+        # check3 = m.calc_error(results.times, results.inputs, results.outputs)
 
         # Now with limits that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
@@ -210,26 +223,9 @@ class TestEstimateParams(unittest.TestCase):
 
         m.estimate_params(data, keys, bounds=[[0, 4], (20, 42), [-4, 15]])
 
-        value1 = m.calc_error(results.times, results.inputs, results.outputs)
-
-        # This should error
+        # This should error as outputs is already a dictionary and we cannot place a hashable type within another.
         with self.assertRaises(TypeError):
             m.estimate_params(data, keys, bounds=[[-1, 5], (20, 40), {-5, 15}])
-
-        value2 = m.calc_error(results.times, results.inputs, results.outputs)
-
-        # self.assertAlmostEqual(value1, value2)
-
-        # interesting behavior where it states the lower bounds is greater than an upper bounds
-
-        # Throws a type error if bounds are Sets. Has to be specific enough to mention that Bounds cannot be undefined ordered Sequence?
-
-        # m.estimate_params(data, keys, bounds=[{0, 4}, {20, 42}, {-4, 15}])
-
-        # m.estimate_params(data, keys, bounds={{0, 4}, {20, 42}, {-4, 15}})
-
-        # This is a Pythonic Error, unhashable, cannot have sets within sets
-        # m.estimate_params(data, keys, bounds={{0, 4}, {20, 42}, {-4, 15}})
 
[…]
