+++ tests/test_estimate_params.py
@@ -1,3 +1,6 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the
+# National Aeronautics and Space Administration.  All Rights Reserved.
+
 from copy import deepcopy
 import io
 import numpy as np
@@ -37,18 +40,9 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         # Enforce if parameters are within bounds after estimate_params()
         check = m.calc_error(results.times, results.inputs, results.outputs)
-        # m.estimate_params(data, keys, bounds=((1.243, 4.0), (19.13212, 39.12301), (-21.123, 2.000991)))
-        # check2= m.calc_error(results.times, results.inputs, results.outputs)
 
-        # Note about estimate_params. Calling it on extremely big values and/or having bounds be very obviously incorrect will
-        # result in a incorrectly defined behavior. I can make examples detailing common examples of this occuring would be using  
-        # the battery and having some parameters be VERY wrong. Also need to look into making other parameters here very wrong.
-        # For the latter case, would this confirm that it would be possible for a chain of estimate_params to be called? Makes sense
-        # however, a bit interesting... After giving one bounds, possible for other bounds to change? With same bounds does not happen,
-        # with different bounds occurs? How are we supposed to clearly keep track of this behavior...
+        # Demonstrates furhter limitations of Parameter Estiamtion
 
-        # calc_error is a much bigger value compared to other values. Not including a good range?
-        # Should not affect other calls...
         before = m.parameters
         m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333)))
         check3 = m.calc_error(results.times, results.inputs, results.outputs)
@@ -57,7 +51,10 @@ class TestEstimateParams(unittest.TestCase):
         # Now with limits that do include the true values
         # only returning the upper bounds? Why after this defined behavior?
         # m.parameters['throwing_speed'] = 15
+
+        # Or two calls to estimate_params would resolve this issue
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42)))
+        # m.estimate_params(data, keys, bounds=(0, 8), (20, 42))
         ax = m.parameters
         value3 = m.calc_error(results.times, results.inputs, results.outputs)
         for key in keys:
@@ -67,16 +64,10 @@ class TestEstimateParams(unittest.TestCase):
 
 
     def test_estimate_params(self):
-        # gt copy works as intended. Going to a different object location
-
-        # Potentially include something that focuses on specifically the key value pairs between gt and m.parameters? mayb not
-
-        # Create a model that will easily raise an Exception with calc_error.
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
         data = [(results.times, results.inputs, results.outputs)]
         gt = m.parameters.copy()
-
 
         # Reset some parameters
         m.parameters['thrower_height'] = 1.5
@@ -96,6 +87,8 @@ class TestEstimateParams(unittest.TestCase):
             m.estimate_params(times=None, inputs=None, output=None)
         with self.assertRaises(ValueError):
             m.estimate_params(times='', inputs='', outputs='')
+        # with self.assertRaises(ValueError):
+        #     m.estimate_params(times='[]', inputs='[]', outputs='[]')
[…]
