### 2023-03-23 #5099 @tbsexton
@teubert Outstanding checklist items:

- Does the list of citations include reference for the datasets you have bundled with `ProgPy`? I couldn't find a relevant one.
- I don't see reference to state-of-the-community regarding existing packages, libraries, or comparisons benchmarks. Obviously this is a bit of a difficulty when there's not a huge python PHM community, but there most be _something_ out there, e.g. used in courses? Could also reference:
   - [Lifelines](https://lifelines.readthedocs.io/en/latest/) which gets used for survival analysis in maintenance contexts
   - [pomegranate](https://pomegranate.readthedocs.io/en/latest/) which has modular systems for stochastic process estimation, markov chains, bayes nets, HMMs, etc, all commonly used in PHM.
   - [Sklearn](scikit-learn.org) which is the backbone of most academic courses I've seen, e.g. for anomaly detection (despite not at all being meant for PHM specicifally, but nonetheless)
   - E.g. [Simantha](https://github.com/usnistgov/simantha) Isn't specifically a tool for PHM, and it's not nearly as close to "polished & ready" as ProgPy (iirc), but I know it has a plugin system for external degradation estimation when running the manufacturing sim. Maybe it's an opportunity for ProgPy models to be used? There's probably lots of examples like this! Would make the paper easier to relate to.

  Mentioning some things like this and discussing how `prog_models` fits relative to them would go a ways to say "how are people doing this now, and how can they do it better with `ProgPy`?". I know you all say "ad hoc" but ad hoc _with/using what_?
