### 2023-03-23 4f95e37
Removing regex and fixing raise errors
