### 2023-03-23 ef051ef
Delete thrown_object_models.py
