+++ tests/test_estimate_params.py
@@ -17,7 +17,6 @@ from prog_models.models import *
 from prog_models.models.test_models.linear_models import (
     OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
 from prog_models.models.thrown_object import LinearThrownObject
-from prog_models.models.test_models.thrown_object_models import defaultParams, wrongInputStorage, wrongTimeValues
 
 class TestEstimateParams(unittest.TestCase):
     def test_estimate_params_works(self):
