+++ paper.bib
@@ -61,3 +61,20 @@
   year={2022}
 }
 
+@misc{Dataset_RWBattery,
+  title = {Randomized Battery Usage Data Set},
+  author = {Bole, B. and Kulkarni, C. and Daigle, M.},
+  year = {2014},
+  publisher = {NASA Prognostics Data Repository},
+  journal = {NASA Ames Research Center, Moffett Field, CA},
+  url = {https://www.nasa.gov/content/prognostics-center-of-excellence-data-set-repository}
+}
+
+@misc{Dataset_CMAPPS,
+  title = {Turbofan Engine Degradation Simulation},
+  author = {Saxena, A. and Goebel, K.},
+  year = {2008},
+  publisher = {NASA Prognostics Data Repository},
+  journal = {NASA Ames Research Center, Moffett Field, CA},
+  url = {https://www.nasa.gov/content/prognostics-center-of-excellence-data-set-repository}
+}
