+++ paper.md
@@ -40,7 +40,7 @@ The [NASA Prognostics Python Packages (ProgPy)](https://nasa.github.io/progpy/)[
 # Statement of need
 Prognostics and Health Management (PHM) is a fast-growing field. Successful PHM application can reduce operational costs and prevent failure, making systems safer. There has been limited application of prognostics in practice. This is partially because prognostics is a multi-faceted and complex problem, including data availability, sensor design and placement, and, of interest to us, software. Often, software is written for an ad-hoc single prognostic application and cannot be transferred to others. Thus, there is a need for a foundational set of efficient tools to enable new PHM technologies. 
 
-ProgPy provides a set of support packages for individuals researching and developing prognostic technologies. ProgPy consists of three packages: prog_models, prog_algs, and prog_server. prog_models provides tools aiding the development, evaluation, simulation, and tuning of prognostic models, whether physics-based or data-driven. prog_algs supports uncertainty representation, state estimation, prognostics, the evaluation and visualization of prognostic results, and the creation of new prognostic algorithms. prog_server is a Service-Oriented Architecture for prognostics and state estimation. prog_server is also distributed with a python client, prog_client. 
+ProgPy provides a set of support packages for individuals researching and developing prognostic technologies. ProgPy consists of three packages: prog_models, prog_algs, and prog_server. prog_models provides tools aiding the development, evaluation, simulation, and tuning of prognostic models, whether physics-based or data-driven. prog_models also supports downloading select relevant datasets [@Dataset_RWBattery, @Dataset_CMAPPS]. prog_algs supports uncertainty representation, state estimation, prognostics, the evaluation and visualization of prognostic results, and the creation of new prognostic algorithms. prog_server is a Service-Oriented Architecture for prognostics and state estimation. prog_server is also distributed with a python client, prog_client. 
 
 The following sections describe some ways ProgPy could be used. There are many features and use cases beyond those illustrated here. See the [ProgPy documentation](https://nasa.github.io/progpy/) for more information.
 
