### 2023-03-26 66ffbd8
Update prognostics_model.py
