+++ src/prog_models/prognostics_model.py
@@ -6,9 +6,10 @@ from collections import abc, namedtuple
 from copy import deepcopy
 import itertools
 import json
+import re
 from numbers import Number
 import numpy as np
-from typing import Callable, Iterable, List
+from typing import Callable, Iterable, List, Sequence
 from warnings import warn
 
 from prog_models.exceptions import ProgModelInputException, ProgModelTypeError, ProgModelException, ProgModelStateLimitWarning
@@ -1026,11 +1027,8 @@ class PrognosticsModel(ABC):
             def load_eqn(t, x):
                 u = future_loading_eqn(t, x)
                 return self.InputContainer(u)
-<<<<<<< Updated upstream
-
-=======
         
->>>>>>> Stashed changes
+
         if not isinstance(self.next_state(x.copy(), u, dt0), DictLikeMatrixWrapper):
             # Wrapper around next_state
             def next_state(x, u, dt):
@@ -1152,13 +1150,13 @@ class PrognosticsModel(ABC):
         """Calculate Mean Squared Error (MSE) between simulated and observed
 
         Args:
-            times (list[float]): array of times for each sample
-            inputs (list[dict]): array of input dictionaries where input[x] corresponds to time[x]
-            outputs (list[dict]): array of output dictionaries where output[x] corresponds to time[x]
-        
+            times (list[float]): Array of times for each sample.
+            inputs (list[dict]): Array of input dictionaries where input[x] corresponds to time[x].
+            outputs (list[dict]): Array of output dictionaries where output[x] corresponds to time[x].
+
         Keyword Args:
-            x0 (dict, optional): Initial state
-            dt (double, optional): time step
+            x0 (dict, optional): Initial state.
+            dt (double, optional): Maximum time step.
 
         Returns:
             double: Total error
@@ -1183,11 +1181,7 @@ class PrognosticsModel(ABC):
         counter = 0  # Needed to account for skipped (i.e., none) values
         t_last = times[0]
         err_total = 0
-<<<<<<< Updated upstream
-        z_obs = self.output(x)  # Initialize
-=======
         z_obs = self.output(x)
->>>>>>> Stashed changes
         for t, u, z in zip(times, inputs, outputs):
             while t_last < t:
                 t_new = min(t_last + dt, t)
@@ -1209,7 +1203,7 @@ class PrognosticsModel(ABC):
         """Estimate the model parameters given data. Overrides model parameters
 
         Keyword Args:
-            keys (list[str]): 
+            keys (list[str]):
                 Parameter keys to optimize
             times (list[float]):
                 Array of times for each sample
@@ -1217,13 +1211,13 @@ class PrognosticsModel(ABC):
                 Array of input containers where input[x] corresponds to time[x]
             outputs (list[OutputContainer]):
                 Array of output containers where output[x] corresponds to time[x]
-            method (str, optional): 
+            method (str, optional):
                 Optimization method- see scipy.optimize.minimize for options
-            bounds (tuple or dict): 
+            bounds (tuple or dict, optional):
                 Bounds for optimization in format ((lower1, upper1), (lower2, upper2), ...) or {key1: (lower1, upper1), key2: (lower2, upper2), ...}
-            options (dict): 
+            options (dict, optional):
                 Options passed to optimizer. see scipy.optimize.minimize for options
-            runs (list[tuple], depreciated): 
+            runs (list[tuple], depreciated):
                 data from all runs, where runs[0] is the data from run 0. Each run consists of a tuple of arrays of times, input dicts, and output dicts. Use inputs, outputs, states, times, etc. instead
 
         See: examples.param_est
@@ -1233,21 +1227,18 @@ class PrognosticsModel(ABC):
         if keys is None:
[…]
