+++ tests/test_estimate_params.py
@@ -4,9 +4,6 @@
 from copy import deepcopy
 import io
 import numpy as np
-from os.path import dirname, join
-import pickle
-import sys
 import unittest
 
 # This ensures that the directory containing ProgModelTemplate is in the python search directory
@@ -14,9 +11,6 @@ sys.path.append(join(dirname(__file__), ".."))
 
 from prog_models import *
 from prog_models.models import *
-from prog_models.models.test_models.linear_models import (
-    OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
-from prog_models.models.thrown_object import LinearThrownObject
 
 class TestEstimateParams(unittest.TestCase):
     def test_estimate_params_works(self):
@@ -25,37 +19,45 @@ class TestEstimateParams(unittest.TestCase):
         data = [(results.times, results.inputs, results.outputs)]
         gt = m.parameters.copy()
 
-        self.assertTrue(m.parameters == gt)
         self.assertEqual(m.parameters, gt)
 
-        # Now lets reset some parameters
+        # Now lets incorrectly set some parameters
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         keys = ['thrower_height', 'throwing_speed']
         m.estimate_params(data, keys)
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+        zero = m.calc_error(results.times, results.inputs, results.outputs)
         
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         # Enforce if parameters are within bounds after estimate_params()
-        check = m.calc_error(results.times, results.inputs, results.outputs)
+        first = m.calc_error(results.times, results.inputs, results.outputs)
 
-        # Demonstrates furhter limitations of Parameter Estiamtion
-
-        before = m.parameters
-        m.estimate_params(data, keys, bounds=((1.243, 4.0), (-12.9234, 3.33333333)))
-        check3 = m.calc_error(results.times, results.inputs, results.outputs)
-        after = m.parameters
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=((0, 4), (40, 40)))
+        case = m.calc_error(results.times, results.inputs, results.outputs)
+        # Demonstrates further limitations of Parameter Estiamtion
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 41.99)))
+        for key in keys:
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+        second = m.calc_error(results.times, results.inputs, results.outputs)
 
         # Now with limits that do include the true values
         # only returning the upper bounds? Why after this defined behavior?
         # m.parameters['throwing_speed'] = 15
 
         # Or two calls to estimate_params would resolve this issue
-        m.estimate_params(data, keys, bounds=((0, 8), (20, 42)))
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         # m.estimate_params(data, keys, bounds=(0, 8), (20, 42))
-        ax = m.parameters
-        value3 = m.calc_error(results.times, results.inputs, results.outputs)
[…]
