+++ src/prog_models/prognostics_model.py
@@ -1229,9 +1229,14 @@ class PrognosticsModel(ABC):
             # if no keys provided, use all
             keys = [key for key in self.parameters.keys() if isinstance(self.parameters[key], Number)]
         
+        if isinstance(keys, set):
+            raise ValueError(f"Can not pass in keys as a Set. Sets are unordered by construction, so bounds may be out of order.")
+        
         for key in keys:
+            # if isinstance(key, Sequence):
+            #     raise ValueError(f"Key '{key}' cannot be a Sequence")
             if key not in self.parameters:
-                raise ValueError("Passing in key that does not exist.")
+                raise ValueError(f"Key '{key}' not in model parameters")
 
         config = {
             'method': 'nelder-mead',
@@ -1270,10 +1275,27 @@ class PrognosticsModel(ABC):
                 raise ValueError("Must pass in outputs")
             raise ValueError("Must provide either runs")
         if runs is None:
-            if len(times) != len(inputs) or len(outputs) != len(inputs):
+            # If depreciated feature runs is not provided (will be removed in future version)
+            # Check if required times, inputs, and outputs are present
+            missing_args = []
+            for arg in ('times', 'inputs', 'outputs'):
+                # convert = type(locals.get(arg))
+                if locals().get(arg) is None:
+                    missing_args.append(arg)
+                if arg not in missing_args:
+                    convert = list(locals().get(arg))
+                    if convert == []: missing_args.append(arg)
+            if len(missing_args) > 0:
+                # Concat into string
+                missing_args_str = ', '.join(missing_args)
+                # missing_args_str = missing_args_str[:-2] # Remove last comma and space
+                raise ValueError(f"Missing keyword arguments {missing_args_str}")
+            
+            # Check lengths of args
+            if len(times) != len(inputs): 
                 raise ValueError(f"Times, inputs, and outputs must be same length. Length of times: {len(times)}, Length of inputs: {len(inputs)}, Length of outputs: {len(outputs)}")
             if len(times) == 0:
-                # since they are all the same length, does not matter as much
+                # Since inputs, times, and outputs are already confirmed to be the same length, only check that one is not empty
                 raise ValueError(f"Times, inputs, and outputs must have at least one element")
             # For now- convert to runs
             runs = [(t, u, z) for t, u, z in zip(times, inputs, outputs)]
@@ -1311,7 +1333,6 @@ class PrognosticsModel(ABC):
             if len(times) == 0:
                 raise ValueError(f"Times, inputs, and outputs for Run {i} must have at least one element")
             if not isinstance(inputs[0], self.InputContainer):
-                # Error for recent test occurs here.
                 inputs = [self.InputContainer(u_i) for u_i in inputs]
                 has_changed = True
             if isinstance(outputs, np.ndarray):
