+++ tests/test_estimate_params.py
@@ -1,7 +1,6 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
-from copy import deepcopy
 import numpy as np
 import unittest
 
@@ -25,25 +24,23 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys)
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-        zero = m.calc_error(results.times, results.inputs, results.outputs)
         
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         # Enforce if parameters are within bounds after estimate_params()
-        first = m.calc_error(results.times, results.inputs, results.outputs)
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (40, 40)))
-        case = m.calc_error(results.times, results.inputs, results.outputs)
+
+
         # Demonstrates further limitations of Parameter Estiamtion
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 41.99)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-        second = m.calc_error(results.times, results.inputs, results.outputs)
 
         # Now with limits that do include the true values
         # only returning the upper bounds? Why after this defined behavior?
@@ -53,12 +50,10 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
-        # m.estimate_params(data, keys, bounds=(0, 8), (20, 42))
-        third = m.calc_error(results.times, results.inputs, results.outputs)
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2, "Limits of the Bounds do not include the true values")
         
-        
+
     def test_estimate_params(self):
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
@@ -73,22 +68,22 @@ class TestEstimateParams(unittest.TestCase):
         # Need at least one data point
         with self.assertRaises(ValueError) as cm:
             m.estimate_params(times=[], inputs=[], outputs=[])
-        self.assertEqual(
-            'Missing keyword arguments times, inputs, outputs',
-            str(cm.exception)
-        )
-        with self.assertRaises(ValueError):
-            m.estimate_params(times=None, inputs=None, output=None)
-        self.assertEqual(
-            'Missing keyword arguments times, inputs, outputs',
-            str(cm.exception)
-        )
-        with self.assertRaises(ValueError):
-            m.estimate_params(times='', inputs='', outputs='')
-        self.assertEqual(
-            'Missing keyword arguments times, inputs, outputs',
-            str(cm.exception)
-        )
+        # self.assertEqual(
+        #     'Missing keyword arguments times, inputs, outputs',
+        #     str(cm.exception)
+        # )
+        # with self.assertRaises(ValueError):
+        #     m.estimate_params(times=None, inputs=None, output=None)
+        # self.assertEqual(
+        #     'Missing keyword arguments times, inputs, outputs',
+        #     str(cm.exception)
+        # )
+        # with self.assertRaises(ValueError):
[…]
