+++ blub.txt
@@ -1,334 +1,0 @@
-estimate_params from this link needs to be updated: https://nasa.github.io/progpy/api_ref/prog_models/PrognosticModel.html#prog_models.PrognosticsModel.estimate_params
-, updated in the function definition, however not on the website.
-
- - Created test to see what happens if bounds go the wrong way. i.e (30, 7) instead of (7, 30)
-
-Think of ways to create more descriptive error messages
-
-Begin checking other parameters for params_estimate
-
-Statement ____ Swap ordering in params_estimate for bounds and method. ^^website
-
-Question*** Shouldn't changing optimization method chagne end values?
-
-Are default options set to make it various questions more optimal
-
-Idea: on line 1238 of pringistcs_model, I provide change the logic to provide a more acurate error message and detail what exactly is not equal to one another.
-            and potneitally have their values shown.
-
-Question*** Is runs 'depreciated' since one of the parameters is marked as 'depreciated'
-
-        # with self.assertRaises(ValueError):
-        # Does not do anything. Probably want to include an error message or some indicator to let the user know
-        # in case user is expected change however is not seeing it.
-
-Line 148 does not work as intended?
-
-Talk about runs parameter.
-
-        # Consider giving a warning LINE121 estimate_params
-
-#1240 
-
-                # not in for loops perferably, runtime
-
-# tolerance parameter in minimize function - (tol)
-
-# larger tolerance = larger error
-
-
-        # If not in a list, and only one main data passed in, then you want to run without failing, additional feature needed
-
-
-# Review simulate_to_threshold and make sure previous tests are working as intended
-
-
-        simulated_results = m.simulate_to(2000, future_loading, **options)
-
-
-        value = m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=2)
-
-        m.parameters['alpha'] = 3000
-        keys = ['alpha']
-        
-        error = m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=2)
-        
-        # error returns 0
-
-
-# talk about the different ways runs is created
-
-
-
-test_estimate_params 
-# Should be passing.
-m.estimate_params(times=[[times]], inputs=[[inputs]], outputs=[[outputs]])
-        # This will work after future change 
-
-prognostics_model
-                # try:
-                #     outputs = [self.OutputContainer(u_i) for u_i in outputs]
-                # except:
-                #     raise TypeError("Passing in too many lists aorund outputs")
-        
-1307, if not isinstance(inputs[0], self.InputContainer):
-
-# Also keeping track of which run specifcally is failing.
-        - run 0 vs run 1
-        - Note about specific implementation. Even if the user does not pass in a particular 'run' persay, the error message will
-still communicate that they are failing on the first run. Easy fix if it is worth the effort.
-
-separate pull request for tol feature in params_estimate.
-# Move all changes here:
-
-        
-        # Testing if tolerance works
-        m.estimate_params(data, keys, bounds=bound, method='Powell', tol=2.00)
-        m1.estimate_params(data, keys, bounds=bound, method='Powell', tol=1e-6)
-
-        self.assertNotAlmostEqual(m.calc_error(times, inputs, outputs), m1.calc_error(times, inputs, outputs))
-
-
-        m.estimate_params(data, keys, bounds=bound, method='Powell', tol=1e-6)
-        m1.estimate_params(data, keys, bounds=bound, method='Powell', tol=1e-6)
-
[…]
