+++ tests/test_estimate_params.py
@@ -578,16 +578,14 @@ class TestEstimateParams(unittest.TestCase):
         m1.parameters['throwing_speed'] = 25
 
         
-        # using battery model to see when calc_errors do not equate each other
+        # using thrown object model to see when calc_errors do not equate each other
         m.estimate_params(data, keys, bounds=bound, method='Powell')
         check1 = m.calc_error(results.times, results.inputs, results.outputs)
         m1.estimate_params(data, keys, bounds=bound, method='CG')
 
-        # For Simple models, there shouldn't be too much change
         self.assertNotEqual(m.calc_error(results.times, results.inputs, results.outputs), 
                             m1.calc_error(results.times, results.inputs, results.outputs))
         
-        # For Simple models, there shouldn't be too much change
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
