+++ tests/test_estimate_params.py
@@ -51,7 +51,7 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
-            self.assertAlmostEqual(m.parameters[key], gt[key], 2, "Limits of the Bounds do not include the true values")
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
 
     def test_estimate_params(self):
@@ -68,23 +68,22 @@ class TestEstimateParams(unittest.TestCase):
         # Need at least one data point
         with self.assertRaises(ValueError) as cm:
             m.estimate_params(times=[], inputs=[], outputs=[])
-        # self.assertEqual(
-        #     'Missing keyword arguments times, inputs, outputs',
-        #     str(cm.exception)
-        # )
-        # with self.assertRaises(ValueError):
-        #     m.estimate_params(times=None, inputs=None, output=None)
-        # self.assertEqual(
-        #     'Missing keyword arguments times, inputs, outputs',
-        #     str(cm.exception)
-        # )
-        # with self.assertRaises(ValueError):
-        #     m.estimate_params(times='', inputs='', outputs='')
-        # self.assertEqual(
-        #     'Missing keyword arguments times, inputs, outputs',
-        #     str(cm.exception) 
-        # )
-
+        self.assertEqual(
+            'Times, inputs, and outputs must have at least one element',
+            str(cm.exception)
+        )
+        with self.assertRaises(ValueError) as cm:
+            m.estimate_params(times=None, inputs=None, output=None)
+        self.assertEqual(
+            'Missing keyword arguments times, inputs, outputs',
+            str(cm.exception)
+        )
+        with self.assertRaises(ValueError) as cm:
+            m.estimate_params(times='', inputs='', outputs='')
+        self.assertEqual(
+            'Times, inputs, and outputs must have at least one element',
+            str(cm.exception) 
+        )
         with self.assertRaises(ValueError) as cm:
             m.estimate_params(times=[[]], inputs=[[]], outputs=[[]])
         self.assertEqual(
@@ -99,7 +98,13 @@ class TestEstimateParams(unittest.TestCase):
              'Times, inputs, and outputs must be same length. Length of times: 0, Length of inputs: 0, Length of outputs: 9',
              str(cm.exception)
         )
-
+        
+        with self.assertRaises(ValueError) as cm:
+            m.estimate_params(times=results.times, inputs=None, outputs=None)
+        self.assertEqual(
+            'Missing keyword arguments inputs, outputs',
+            str(cm.exception)
+        )
         with self.assertRaises(ValueError) as cm:
             m.estimate_params(times=results.times, inputs=results.inputs, outputs=None)
         self.assertEqual(
@@ -128,13 +133,13 @@ class TestEstimateParams(unittest.TestCase):
         value2 = m.calc_error(results.times, results.inputs, results.outputs)
         # Not the most accurate way of seeing when these values are not accurate.
         for key in keys:
-            self.assertNotAlmostEqual(m.parameters[key], gt[key], 1, "Limits of the Bounds do not include the true values")
+            self.assertNotAlmostEqual(m.parameters[key], gt[key], 1)
 
         # Now with limits that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
         value3 = m.calc_error(results.times, results.inputs, results.outputs)
         for key in keys:
-            self.assertAlmostEqual(m.parameters[key], gt[key], 2, "Limits of the Bounds do not include the true values")
[…]
