### 2023-03-29 a1b1fb0
First iteration of fixing division by zero
