+++ tests/test_estimate_params.py
@@ -20,25 +20,31 @@ class TestEstimateParams(unittest.TestCase):
         # Now lets incorrectly set some parameters
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        keys = ['thrower_height', 'throwing_speed']
+        m.parameters['g'] = -8.5
+        keys = ['thrower_height', 'throwing_speed', 'g']
         m.estimate_params(data, keys)
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
-        # Enforce if parameters are within bounds after estimate_params()
+        m.parameters['g'] = -8.5
+
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, -5)))
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (40, 40)))
+        m.parameters['g'] = -8.5
+
+        m.estimate_params(data, keys, bounds=((-40, 40), (-40, 40), (-40, 40)))
 
 
         # Demonstrates further limitations of Parameter Estiamtion
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 41.99)))
+        m.parameters['g'] = -8.5
+
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 41.99), (-20, -5)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
@@ -49,7 +55,9 @@ class TestEstimateParams(unittest.TestCase):
         # Or two calls to estimate_params would resolve this issue
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
+        m.parameters['g'] = -8.5
+
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, 10)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
@@ -519,12 +527,15 @@ class TestEstimateParams(unittest.TestCase):
         # Now lets reset some parameters
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
+        m.parameters['g'] = -8.5
         keys = ['thrower_height', 'throwing_speed', 'g']
         bound=((0, 8), (20, 42), (-20, -5))
 
-        m.parameters['thrower_height'] = 1.5
+        m.parameters['thrower_height'] = 1.2
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=bound, method='Nelder-Mead')
+        m.parameters['g'] = -8.5
+
+        m.estimate_params(data, keys, bounds=bound, method='Nelder-Mead', options={'maxiter':1000})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
@@ -532,15 +543,19 @@ class TestEstimateParams(unittest.TestCase):
         # Passing in Method to see if it works
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=bound, method='TNC', options={'maxfun': 1000, 'disp': False})
+        m.parameters['g'] = -8.5
+
[…]
