+++ paper.bib
@@ -78,3 +78,23 @@
   journal = {NASA Ames Research Center, Moffett Field, CA},
   url = {https://www.nasa.gov/content/prognostics-center-of-excellence-data-set-repository}
 }
+
+@misc{Simantha, 
+  title = {Simantha},
+  author = {Mehdi Dadfarnia and Serghei Drozdov},
+  year = {2013},
+  publisher = {National Institute of Standards and Technology [Software]},
+  journal = {NIST},
+  url = {https://github.com/usnistgov/simantha}
+}
+
+@misc{davidson_pilon_cameron_2022_7329096,
+  author = {Davidson-Pilon, Cameron},
+  title = {lifelines, survival analysis in Python},
+  month = nov,
+  year = 2022,
+  publisher = {Zenodo},
+  version = {v0.27.4},
+  doi = {10.5281/zenodo.7329096},
+  url = {https://doi.org/10.5281/zenodo.7329096}
+}
