+++ .github/workflows/python-package.yml
@@ -1,135 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: Python package
-
-on:
-  push:
-  pull_request:
-    paths:
-      - prog_models
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ['3.7']
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install -e .
-        python -m pip install notebook
-        python -m pip install testbook
-        python -m pip install requests
-    - name: Run tests
-      run: python -m tests
-  # test_prog_algs_released:
-  #   runs-on: ubuntu-latest
-  #   strategy:
-  #     matrix:
-  #       python-version: ['3.9']
-  #   steps:
-  #   - uses: actions/checkout@v2
-  #   - name: Set up Python ${{ matrix.python-version }}
-  #     uses: actions/setup-python@v2
-  #     with:
-  #       python-version: ${{ matrix.python-version }}
-  #   - name: Install dependencies
-  #     run: |
-  #       python -m pip install --upgrade pip
-  #       python -m pip install -e .
-  #       python -m pip install notebook
-  #       python -m pip install testbook
-  #       python -m pip install requests
-  #   - name: Clone prog_algs
-  #     uses: actions/checkout@v2
-  #     with:
-  #       repository: nasa/prog_algs 
-  #       ref: refs/heads/master
-  #       path: prog_algs
-  #   - name: Install prog_algs
-  #     run: |
-  #       python -m pip install -e ./prog_algs
-  #   - name: Run tests
-  #     run: |
-  #       cd prog_algs 
-  #       python -m tests
-  test_prog_algs_dev:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [ '3.9']
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install -e .
-        pip install notebook
-        pip install testbook
-        pip install requests
-    - name: Clone prog_algs
-      uses: actions/checkout@v3
-      with:
-        repository: nasa/prog_algs 
-        ref: refs/heads/dev
-        path: prog_algs
-    - name: Install prog_algs
-      run: |
-        python -m pip install -e ./prog_algs
-    - name: Run tests
-      run: |
-        cd prog_algs 
-        python -m tests
-  copyright:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ['3.9']
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Run copyright check
-      run: |
-        python scripts/test_copyright.py
-  # coverage:
-  #   runs-on: ubuntu-latest
-  #   strategy:
-  #     matrix:
-  #       python-version: ['3.9']
-  #   steps:
-  #   - uses: actions/checkout@v3
-  #   - name: Set up Python ${{ matrix.python-version }}
[…]
