+++ tests/test_estimate_params.py
@@ -31,11 +31,11 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+        # Checking to see if our estimated parameters are wihtin the bounds.
         self.assertLessEqual(m.parameters['thrower_height'], 4)
         self.assertGreaterEqual(m.parameters['thrower_height'], 0)
         self.assertLessEqual(m.parameters['throwing_speed'], 42)
         self.assertGreaterEqual(m.parameters['throwing_speed'], 20)
-        # Enforce if parameters are within bounds after estimate_params()
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
@@ -62,21 +62,20 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-        # What occurs when our given parameters are so far away from the original values
+        # When our incorrecltly set parameters that are not within the bounds themselves
         m.parameters['thrower_height'] = 10
         m.parameters['throwing_speed'] = -10
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
-        # These values are still within the bounds, they are just not close to the original value at all.
-        self.assertLessEqual(m.parameters['thrower_height'], 4)
-        self.assertGreaterEqual(m.parameters['thrower_height'], 0)
-        self.assertLessEqual(m.parameters['throwing_speed'], 42)
-        self.assertGreaterEqual(m.parameters['throwing_speed'], 20)
+        # These returned values are still within the bounds, they are just not close to the original value at all.
+        # This results in our estimate parameters to result in the upper extremas
+        self.assertEqual(m.parameters['thrower_height'], 4)
+        self.assertEqual(m.parameters['throwing_speed'], 42)
 
         # Show casing results of having a local min/max
         # Even though all our bounds accomodate for the original model params values,
-            # our estimate_parmams returns lower/upper bound values.
+        # our estimate_parmams returns lower/upper bound values.
         m.parameters['thrower_height'] = 4
         m.parameters['throwing_speed'] = 24
         m.parameters['g'] = -20
@@ -208,14 +207,12 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
-        value2 = m.calc_error(results.times, results.inputs, results.outputs)
         # Not the most accurate way of seeing when these values are not accurate.
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1)
 
         # Now with limits that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
-        value3 = m.calc_error(results.times, results.inputs, results.outputs)
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
@@ -529,16 +526,19 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
 
+        # Setting kesy to incorrect values
         keys = ['x', 'y', 'g']
         bound=((0, 8), (20, 42), (-20, -5))
         with self.assertRaises(ValueError):
[…]
