### 2023-04-03 019730e
Implementing Tests for calc_error
