+++ tests/test_estimate_params.py
@@ -623,7 +623,7 @@ class TestEstimateParams(unittest.TestCase):
         m1.parameters['thrower_height'] = 1.5
         m1.parameters['throwing_speed'] = 25
         
-        # Increasing total amount of iterations, having different methods would matter proportionally.
+        # Decreasing total amount of iterations, having different methods would matter proportionally.
         m.estimate_params(data, keys, bounds=bound, method='Powell', options={'maxiter': 50, 'disp': False})
         m1.estimate_params(data, keys, bounds=bound, method='CG', options={'maxiter': 50, 'disp': False})
 
@@ -828,7 +828,7 @@ class TestEstimateParams(unittest.TestCase):
 
 
     # @unittest.skip
-    def test_big_example(self):
+    def test_calc_error(self):
         # Note, lowering timesteps or increasing simulate threshold may cause this model to not run (takes too long)
         m = BatteryElectroChemEOD()
 
@@ -850,31 +850,47 @@ class TestEstimateParams(unittest.TestCase):
                 i = 3
             return m.InputContainer({'i': i})
     
-        simulated_results = m.simulate_to(1200, future_loading, **options)
+        simulated_results = m.simulate_to(2000, future_loading, **options)
 
-        value = m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
+        # Running calc_error before setting incorrect parameters
+        m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
 
-        # Creating errors
-        m.parameters['qMax'] = 12000
+        # Initalizing parameters to very erreneous values       
+        m.parameters['qMax'] = 4000
         keys = ['qMax']
 
-        error = m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
+        # Before running estimate_params
+        with self.assertRaises(ValueError):
+            m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
 
         m.estimate_params([(simulated_results.times, simulated_results.inputs, simulated_results.outputs)], keys, dt=0.5)
 
-        value1 = m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
-        
-        # After estimating parameters, value1 needs to at least be less than error
-        self.assertLessEqual(value1, error)
+        # After running estimate_params. Note that this would not change the outcome of the result
+        with self.assertRaises(ValueError):
+            m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)    
 
+        # Running through various dt values that could work
+        for i in np.arange(0, 1, 0.1):
+            with self.assertRaises(ValueError):
+                m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=i)
+
+        for i in range(2, 10):
+            with self.assertRaises(ValueError):
+                m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=i)
+ 
+        # Creating duplicate model
         m1 = BatteryElectroChemEOD()
 
+        orig_params = m1.parameters.copy()
+
+        # Much bigger parameter initialization
         m.parameters['kp'] = m1.parameters['kp'] = 10000
-        # m.parameters['vMax'] = m1.parameters['vMax'] = 3000
-        m.parameters['kn'] = m1.parameters['kn'] = 1000 #Does not error out even though value is unbelivably off
+        m.parameters['kn'] = m1.parameters['kn'] = 1000 
[…]
