### 2023-04-05 6969f91
updating the controller class LQR_I with some general improvements.
