+++ src/prog_models/loading_fcns/controllers.py
@@ -74,7 +74,7 @@ class LQR():
 
         self.Q = Q
         self.R = R
-        pass
+        
     
     def compute_gain(self, A, B):
         """ Compute controller gain given state of the system described by linear model A, B"""
@@ -110,34 +110,34 @@ class LQR():
 
 class LQR_I():
     """ Linear Quadratic Regulator with Integral Effect"""
-    
-    # def __init__(self, n_states, n_inputs, n_outputs, dt, int_lag=np.inf, C=None, Q=None, R=None, qi=None) -> None:
+
     def __init__(self, x_ref, vehicle, **kwargs):
 
-        self.type      = 'LQR_I'    # type of controller
-        self.states = vehicle.states
-        self.n_states  = len(self.states)   # number of states
-        self.outputs = vehicle.outputs[:3]
-        self.n_outputs = 3                     # number of outputs
-        self.inputs = vehicle.inputs
-        self.n_inputs  = len(self.inputs)   # number of inputs
-        self.err_hist   = []         # error history (integral)
-        self.ref_traj = x_ref
+        self.type      = 'LQR_I'                # type of controller
+        self.states    = vehicle.states         # state variables of the system to be controlled (x, y, z, phi, theta, psi)
+        self.n_states  = len(self.states)       # number of states
+        self.outputs   = vehicle.outputs[:3]    # output variables of the system to be controlled (x, y, z only)
+        self.n_outputs = 3                      # number of outputs
+        self.inputs    = vehicle.inputs         # input variables of the system to be controlled ()
+        self.n_inputs  = len(self.inputs)       # number of inputs
+        self.err_hist  = []                     # error history (integral)
+        self.ref_traj  = x_ref                  # reference state to follow during simulation (x_ref, y_ref, z_ref, phi_ref, theta_ref, psi_ref, ...)
 
         # Default control parameters
         # --------------------------------
-        self.parameters = dict(int_lag=100,
-                              Q=np.eye(self.n_states),
-                              R=np.eye(self.n_inputs),
-                              qi=np.ones((self.n_outputs,)),
-                              scheduled_var='psi',
-                              index_scheduled_var=5)
-        self.parameters.update(kwargs)
+        self.parameters = dict(int_lag=100,             # error integral length: how far back in time to compute integral error
+                               Q=np.eye(self.n_states), # state error penalty matrix: how 'bad' is an error in the state vector w.r.t. the reference state vector
+                               R=np.eye(self.n_inputs), # input penalty matrix: how 'hard' it is to produce the desired input (thrust and three moments along three axes)
+                               qi=np.ones((self.n_outputs,)),   # integral error penalty: how much the error history is contributing to the overall error
+                               scheduled_var='psi',     # variable used to create the scheduled controller gains (only psi allowed for now)
+                               index_scheduled_var=5)   # index corresponding to the scheduled_var (psi) in the state vector x; i.e., x[5] = psi
+        self.parameters.update(kwargs)                  # update control parameters according to user
 
-        # Get scheduled variable index
+        # Get scheduled variable index (only necessary if scheduled_var is changed, which is not happening at the moment)
         self.parameters['index_scheduled_var'] = self.states.find(self.parameters['scheduled_var'])
 
-
+        # Initialize other controller-related variables
+        # ---------------------------------------------
         self.dt        = vehicle.parameters['dt']
         self.C         = np.zeros((self.n_outputs, self.n_states))
[…]
