### 2023-04-05 8fbc3fd
Split out calc_error and add method
