+++ src/prog_models/prognostics_model.py
@@ -14,6 +14,7 @@ from warnings import warn
 from prog_models.exceptions import ProgModelInputException, ProgModelTypeError, ProgModelException, ProgModelStateLimitWarning
 from prog_models.sim_result import SimResult, LazySimResult
 from prog_models.utils import ProgressBar
+from prog_models.utils import calc_error
 from prog_models.utils.containers import DictLikeMatrixWrapper
 from prog_models.utils.parameters import PrognosticsModelParameters
 from prog_models.utils.serialization import *
@@ -1153,51 +1154,25 @@ class PrognosticsModel(ABC):
             outputs (list[dict]): array of output dictionaries where output[x] corresponds to time[x]
         
         Keyword Args:
+            method (str, optional): Error method to use. Supported methods include:
+                * MSE (Mean Squared Error)
             x0 (dict, optional): Initial state
-            dt (double, optional): time step
+            dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
 
         Returns:
-            double: Total error
+            float: error
+
+        See Also:
+            :func:`calc_error.MSE`
         """
-        if isinstance(times[0], Iterable):
-            # Calculate error for each
-            error = [self.calc_error(t, i, z, **kwargs) for (t, i, z) in zip(times, inputs, outputs)]
-            return sum(error)/len(error)
+        method = kwargs.get('method', 'MSE')
 
-        x = kwargs.get('x0', self.initialize(inputs[0], outputs[0]))
-        dt = kwargs.get('dt', 1e99)
-
-        if not isinstance(x, self.StateContainer):
-            x = [self.StateContainer(x_i) for x_i in x]
-
-        if not isinstance(inputs[0], self.InputContainer):
-            inputs = [self.InputContainer(u_i) for u_i in inputs]
-        
-        if not isinstance(outputs[0], self.OutputContainer):
-            outputs = [self.OutputContainer(z_i) for z_i in outputs]
-
-        counter = 0  # Needed to account for skipped (i.e., none) values
-        t_last = times[0]
-        err_total = 0
-        z_obs = self.output(x)  # Initialize
-        for t, u, z in zip(times, inputs, outputs):
-            while t_last < t:
-                t_new = min(t_last + dt, t)
-                x = self.next_state(x, u, t_new-t_last)
-                t_last = t_new
-                if t >= t_last:
-                    # Only recalculate if required
-                    z_obs = self.output(x)
-            if not (None in z_obs.matrix or None in z.matrix):
-                if any(np.isnan(z_obs.matrix)):
-                    warn("Model unstable- NaN reached in simulation (t={})".format(t))
-                    break
-                err_total += np.sum(np.square(z.matrix - z_obs.matrix), where= ~np.isnan(z.matrix))
-                counter += 1
-
-        return err_total/counter
+        if method.lower() == 'mse':
+            return calc_error.MSE(self, times, inputs, outputs, **kwargs)
+        else:
+            raise ProgModelInputException(f"Error method '{method}' not supported")
     
-    def estimate_params(self, runs : List[tuple] = None, keys : List[str] = None, times = None, inputs = None, outputs = None, **kwargs) -> None:
+    def estimate_params(self, runs : List[tuple] = None, keys : List[str] = None, times = None, inputs = None, outputs = None, method = 'nelder-mead', **kwargs) -> None:
         """Estimate the model parameters given data. Overrides model parameters
 
         Keyword Args:
@@ -1211,6 +1186,8 @@ class PrognosticsModel(ABC):
                 Array of output containers where output[x] corresponds to time[x]
             method (str, optional): 
                 Optimization method- see scipy.optimize.minimize for options
+            error_method (str, optional):
[…]
