### 2023-04-05 6823a37
Add pomegranate notes
