+++ paper.bib
@@ -98,3 +98,14 @@
   doi = {10.5281/zenodo.7329096},
   url = {https://doi.org/10.5281/zenodo.7329096}
 }
+
+@article{schreiber2017pomegranate,
+  title={Pomegranate: fast and flexible probabilistic modeling in python},
+  author={Schreiber, Jacob},
+  journal={The Journal of Machine Learning Research},
+  volume={18},
+  number={1},
+  pages={5992--5997},
+  year={2017},
+  publisher={JMLR. org}
+}
