+++ .github/workflows/python-package.yml
@@ -110,31 +110,31 @@ jobs:
     - name: Run copyright check
       run: |
         python scripts/test_copyright.py
-  # coverage:
-  #   timeout-minutes: 30
-  #   runs-on: ubuntu-latest
-  #   strategy:
-  #     matrix:
-  #       python-version: ['3.9']
-  #   steps:
-  #   - uses: actions/checkout@v3
-  #   - name: Set up Python ${{ matrix.python-version }}
-  #     uses: actions/setup-python@v4
-  #     with:
-  #       python-version: ${{ matrix.python-version }}
-  #   - name: Install dependencies
-  #     run: |
-  #       python -m pip install --upgrade pip
-  #       python -m pip install -e .
-  #       pip install coverage
-  #       pip install notebook
-  #       pip install testbook
-  #       pip install requests
-  #   - name: Run coverage
-  #     run: |
-  #       coverage run -m tests
-  #       coverage xml
-  #   - name: "Upload coverage to Codecov"
-  #     uses: codecov/codecov-action@v3
-  #     with:
-  #       fail_ci_if_error: true
+  coverage:
+    timeout-minutes: 30
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ['3.10']
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install -e .
+        pip install coverage
+        pip install notebook
+        pip install testbook
+        pip install requests
+    - name: Run coverage
+      run: |
+        coverage run -m tests
+        coverage xml -i
+    - name: "Upload coverage to Codecov"
+      uses: codecov/codecov-action@v3
+      with:
+        fail_ci_if_error: true
