### 2023-04-07 4d9053d
Add data_model remove examples
