### 2023-04-07 2d695df
add estimate params tests
