+++ tests/test_ensemble.py
@@ -2,11 +2,12 @@
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
 from io import StringIO
+import numpy as np
 import sys
 import unittest
 
 from prog_models import EnsembleModel
-from prog_models.models.test_models.linear_models import OneInputOneOutputOneEventLM
+from prog_models.models.test_models.linear_models import OneInputOneOutputOneEventLM, OneInputOneOutputOneEventAltLM
 
 
 class TestEnsemble(unittest.TestCase):
@@ -39,8 +40,16 @@ class TestEnsemble(unittest.TestCase):
             EnsembleModel([m, m, m, 77])
 
     def test_two_models_identical(self):
+        """
+        this tests that the ensemble model works with two identical model-types with identical states, inputs, etc., one with slightly altered parameters
+        """
         m = OneInputOneOutputOneEventLM()
         m2 = OneInputOneOutputOneEventLM(x0={'x1': 2})
+
+        # Make sure they're not the same - 3x the impact
+        m2.B = np.array([[3]])
+        m2.C = np.array([[3]])
+        m2.F = np.array([[-0.3]])
 
         # An ensemble model with two models should work
         em = EnsembleModel([m, m2])
@@ -51,9 +60,123 @@ class TestEnsemble(unittest.TestCase):
         self.assertEqual(em.outputs, m.outputs)
         self.assertEqual(em.events, m.events)
 
-        # The resulting initial state should be between the two:
+        # The resulting initial state should be exactly between the two:
         x0 = em.initialize()
         self.assertEqual(x0['x1'], 1)
+
+        # Same with state transition
+        u = em.InputContainer({'u1': 1})
+        x1 = em.next_state(x0, u, 1)
+        # m would give 2, m2 would give 4
+        self.assertEqual(x1['x1'], 3)
+
+        # Same with output
+        z = em.output(x1)
+        # m would give 3, m2 would give 9
+        self.assertEqual(z['z1'], 6)
+
+        # Same with event state
+        es = em.event_state(x1)
+        # m would give 0.7, m2 would give 0.1
+        self.assertEqual(es['x1 == 10'], 0.4)
+
+        # threshold met should be false
+        self.assertFalse(em.threshold_met(x1)['x1 == 10'])
+
+        # Transition again
+        x2 = em.next_state(x1, u, 1)
+        # threshold met should be true (because one of 2 models says it is)
+        self.assertTrue(em.threshold_met(x2)['x1 == 10'])
+
+    def test_two_models_different(self):
+        """
+        this tests that the ensemble model works with two different model-types with different states, inputs, etc. Tests how the ensemble model handles the different values
+        """
+        m = OneInputOneOutputOneEventLM()
+        m2 = OneInputOneOutputOneEventAltLM()
+        em = EnsembleModel([m, m2])
+
+        # inputs, states, outputs, events should be a combination of the two models
+        self.assertSetEqual(set(em.inputs), {'u1', 'u2'})
+        self.assertSetEqual(set(em.states), {'x1', 'x2'})
+        self.assertSetEqual(set(em.outputs), {'z1', 'z2'})
+        self.assertSetEqual(set(em.events), {'x1 == 10', 'x2 == 5'})
+
+        # Initialize - should be combination of the two
+        x0 = em.initialize()
+        self.assertEqual(x0['x1'], 0)
+        self.assertEqual(x0['x2'], 0)
+
+        # State transition - should be combination of the two
+        u = em.InputContainer({'u1': 1, 'u2': 2})
+        x1 = em.next_state(x0, u, 1)
+        self.assertEqual(x1['x1'], 1)
[…]
