### 2023-04-07 4c4dd47
add to auto tests
