+++ tests/__main__.py
@@ -16,6 +16,7 @@ from tests.test_direct import main as direct_main
 from tests.test_linear_model import main as linear_main
 from tests.test_composite import main as composite_main
 from .test_serialization import main as serialization_main
+from .test_ensemble import main as ensemble_main
 
 from io import StringIO
 import matplotlib.pyplot as plt
@@ -128,6 +129,11 @@ if __name__ == '__main__':
 
     try:
         serialization_main()
+    except Exception:
+        was_successful = False
+
+    try:
+        ensemble_main()
     except Exception:
         was_successful = False
 
