+++ src/prog_models/models/test_models/linear_models.py
@@ -67,10 +67,13 @@ class OneInputOneOutputNoEventLM(LinearModel):
 
 class OneInputOneOutputOneEventLM(OneInputOneOutputNoEventLM):
     events = ['x1 == 10']
+    performance_metric_keys = ['pm1']
 
     F = np.array([[-0.1]])
     G = np.array([[1]])
 
+    def performance_metrics(self, x) -> dict:
+        return {'pm1': x['x1'] + 1}
 
 class OneInputOneOutputOneEventAltLM(LinearModel):
     """
@@ -79,6 +82,7 @@ class OneInputOneOutputOneEventAltLM(LinearModel):
     inputs = ['u2']
     states = ['x2']
     outputs = ['z2']
+    performance_metric_keys = ['pm2']
     events = ['x2 == 5']
 
     A = np.array([[0]])
@@ -93,6 +97,11 @@ class OneInputOneOutputOneEventAltLM(LinearModel):
             'x2': 0
         }
     }
+
+    def performance_metrics(self, x) -> dict:
+        return {'pm2': x['x2'] + 1}
+
+
 class OneInputOneOutputNoEventLMPM(OneInputOneOutputNoEventLM):
     """
     Same as OneInputOneOutputNoEventLM, but with performance metrics defined as a function. Has a single performance metric that is always the state, plus 1
