+++ paper.md
@@ -40,7 +40,7 @@ The [NASA Prognostics Python Packages (ProgPy)](https://nasa.github.io/progpy/)[
 # Statement of need
 Prognostics and Health Management (PHM) is a fast-growing field. Successful PHM application can reduce operational costs and prevent failure, making systems safer. There has been limited application of prognostics in practice. This is partially because prognostics is a multi-faceted and complex problem, including data availability, sensor design and placement, and, of interest to us, software. 
 
-Often, software is written for an ad-hoc single prognostic application and cannot be transferred to others, or is limited in scope. A few related softwares are described here. Simantha is a discrete manufacturing system simulation package that simulates degradation, but it is limted to a Discrete-Time Markov Chain and doesn't include prognostic capabilities [@Simantha]. Lifelines is a survival analysis tool that can be used for reliability analysis to establish fixed-interval maintenance schedules, a different problem than that solved by ProgPy [@davidson_pilon_cameron_2022_7329096]. Pomegranate is a python probabilistic modeling package effective for data science applications [@schreiber2017pomegranate]. However, Pomegranate does not include explicit state estimation capabilities, prognostics tools, or physics-based degradation modeling features.
+Often, software is written for an ad-hoc single prognostic application and cannot be transferred to others, or is limited in scope. A few related softwares are described here. Simantha is a discrete manufacturing system simulation package that simulates degradation, but it is limted to a Discrete-Time Markov Chain and doesn't include prognostic capabilities [@Simantha]. Lifelines is a survival analysis tool that can be used for reliability analysis to establish fixed-interval maintenance schedules, a different problem than that solved by ProgPy [@davidson_pilon_cameron_2022_7329096]. Pomegranate is a python probabilistic modeling package effective for data science applications [@schreiber2017pomegranate]. However, Pomegranate does not include explicit state estimation capabilities, prognostics tools, or physics-based degradation modeling features. Finally, there are a number of general machine-learning packages such as tensorflow, scikit learn, and pytourch. These are general tools that can be used for diagnostics and prognostics, but are not designed specifically for that application.
 
 There is a need for a foundational set of efficient tools to enable new PHM technologies. 
 
