+++ tests/test_estimate_params.py
@@ -309,8 +309,7 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(ValueError):
             m.estimate_params()
 
-        #Testing with Arrays
-
+        # Testing with Arrays
         m.estimate_params(data, keys, bounds=[(0, 4), (20, 42), (-20, 15)])
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
@@ -444,6 +443,7 @@ class TestEstimateParams(unittest.TestCase):
         # Testing passing in strings with standard bounds
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
+        m.parameters['g'] = -8
         m.estimate_params(data, keys, bounds=(('-3', '12'), ('1', '20'), ('-5', '30')))
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
@@ -601,13 +601,15 @@ class TestEstimateParams(unittest.TestCase):
         keys = ['thrower_height', 'throwing_speed', 'g']
         bound=((0, 8), (20, 42), (-20, -5))
 
+        # Testing that defualt method passed in works as intended
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=bound, method='Nelder-Mead')
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-        # Passing in Method to see if it works
+        # Passing 'TNC' method along with 'maxfun 1000' as our options
+        # to see what happens if there are changed when we pass in 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=bound, method='TNC', options={'maxfun': 1000, 'disp': False})
@@ -619,7 +621,7 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=bound, method='TNC')
         if all(abs(m.parameters[key] - gt[key]) <= 0.02 for key in keys):
-            raise ValueError("m.parameter shouldn't be too close to the original parameters")
+            self.fail("m.parameter shouldn't be too close to the original parameters")
 
         # Passing in Method that does not exist
         with self.assertRaises(ValueError):
@@ -629,9 +631,14 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
 
+        # Additionally, remove the key 'g' that was never set and its bounds
+        keys = ['thrower_height', 'throwing_speed']
+        bound=((0, 8), (20, 42))
+
         m.estimate_params(data, keys, bounds=bound, method='Powell', options={'maxiter': 0, 'disp': False})
+        # self.assertFalse(all(abs(m.parameters[key] - gt[key]) <= 0.02 for key in keys))
         if all(abs(m.parameters[key] - gt[key]) <= 0.02 for key in keys):
-            raise ValueError("m.parameter shouldn't be too close to the original parameters")
+            self.fail("m.parameter shouldn't be too close to the original parameters")
 
         # Passing in arbitrary options should not error that follow our format.
         m.estimate_params(data, keys, bounds=bound, method='Powell', options= {'1':3, 'disp':1})
@@ -643,56 +650,65 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(TypeError):
             m.estimate_params(data, keys, bounds=bound, method='Powell', options={'maxiter': '9999', 'disp': False})
 
-        # Keys that are not defined in specs
[…]
