+++ src/prog_models/prognostics_model.py
@@ -1268,7 +1268,7 @@ class PrognosticsModel(ABC):
                 raise ValueError(f"Missing keyword arguments {missing_args_str}")
             
             # Check lengths of args
-            if len(times) != len(inputs) or len(inputs) != len(outputs): 
+            if len(times) != len(inputs) or len(inputs) != len(outputs):
                 raise ValueError(f"Times, inputs, and outputs must be same length. Length of times: {len(times)}, Length of inputs: {len(inputs)}, Length of outputs: {len(outputs)}")
             if len(times) == 0:
                 # Since inputs, times, and outputs are already confirmed to be the same length, only check that one is not empty
