+++ src/prog_models/utils/calc_error.py
@@ -2,9 +2,9 @@
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
 from collections.abc import Iterable
-import numpy as np
 from warnings import warn
-
+import math
+import numpy as np
 
 def MAX_E(m, times, inputs, outputs, **kwargs):
     """
@@ -87,63 +87,78 @@ def RMSE(m, times, inputs, outputs, **kwargs):
     return np.sqrt(MSE(m, times, inputs, outputs, **kwargs))
 
 
-def MSE(m, times, inputs, outputs, **kwargs):
-    """
-    Calculate the Mean Squared Error between model behavior and some collected data.
+def MSE(self, times, inputs, outputs, **kwargs) -> float:
+    """Calculate Mean Squared Error (MSE) between simulated and observed
 
     Args:
-        m (PrognosticsModel): Model to use for comparison
-        times (list[float]): array of times for each sample
-        inputs (list[dict]): array of input dictionaries where input[x] corresponds to time[x]
-        outputs (list[dict]): array of output dictionaries where output[x] corresponds to time[x]
+        times (list[float]): Array of times for each sample.
+        inputs (list[dict]): Array of input dictionaries where input[x] corresponds to time[x].
+        outputs (list[dict]): Array of output dictionaries where output[x] corresponds to time[x].
 
     Keyword Args:
-        x0 (StateContainer): Current State of the model
-        dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
+        x0 (dict, optional): Initial state.
+        dt (double, optional): Maximum time step.
+        stability_tol (double, optional): Configurable parameter.
+            Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
+            In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
+            stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
+            before the model goes unstable in order to produce a valid estimate of mean squared error. 
+
+            If the model goes unstable before stability_tol is met, NaN is returned. 
+            Else, model goes unstable after stability_tol is met, the mean squared error calculated from data up to the instability is returned.
 
     Returns:
-        float: MSE between model and data
+        double: Total error
     """
     if isinstance(times[0], Iterable):
         # Calculate error for each
-        error = [MSE(t, i, z, **kwargs) for (t, i, z) in zip(times, inputs, outputs)]
+        error = [self.calc_error(t, i, z, **kwargs) for (t, i, z) in zip(times, inputs, outputs)]
         return sum(error)/len(error)
 
-    x = kwargs.get('x0', m.initialize(inputs[0], outputs[0]))
+    x = kwargs.get('x0', self.initialize(inputs[0], outputs[0]))
     dt = kwargs.get('dt', 1e99)
+    stability_tol = kwargs.get('stability_tol', 0.95)
 
-    if not isinstance(x, m.StateContainer):
-        x = [m.StateContainer(x_i) for x_i in x]
+    if not isinstance(x, self.StateContainer):
+        x = [self.StateContainer(x_i) for x_i in x]
 
-    if not isinstance(inputs[0], m.InputContainer):
-        inputs = [m.InputContainer(u_i) for u_i in inputs]
+    if not isinstance(inputs[0], self.InputContainer):
+        inputs = [self.InputContainer(u_i) for u_i in inputs]
     
-    if not isinstance(outputs[0], m.OutputContainer):
-        outputs = [m.OutputContainer(z_i) for z_i in outputs]
+    if not isinstance(outputs[0], self.OutputContainer):
+        outputs = [self.OutputContainer(z_i) for z_i in outputs]
+
[…]
