### 2023-04-10 a1b0247
Adding changes to other calc_error methods.
