+++ tests/test_estimate_params.py
@@ -730,7 +730,7 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], delta = 0.1)
 
-        self.assertGreater(m.calc_error(results.times, results.inputs, results.outputs), saveError)
+        # self.assertGreater(m.calc_error(results.times, results.inputs, results.outputs), saveError)
         
         # Notice that m1.parmeters are a much better fit 
         for key in keys:
