+++ src/prog_models/utils/containers.py
@@ -163,7 +163,7 @@ class DictLikeMatrixWrapper():
 
         example
         -------
-        >>> from prog_models.utils.containers_srs import DictLikeMatrixWrapper
+        >>> from prog_models.utils.containers import DictLikeMatrixWrapper
         >>> dlmw = DictLikeMatrixWrapper(['a', 'b', 'c'], {'a': 1, 'b': 2, 'c': 3})
         >>> 'a' in dlmw  # True
         """
