+++ tests/test_ensemble.py
@@ -18,6 +18,10 @@ class TestEnsemble(unittest.TestCase):
     def tearDown(self):
         sys.stdout = sys.__stdout__
 
+    def test_no_model(self):
+        with self.assertRaises(ValueError):
+            EnsembleModel([])
+
     def test_single_model(self):
         # An ensemble model with one model should raise an exception
 
@@ -41,12 +45,16 @@ class TestEnsemble(unittest.TestCase):
 
     def test_two_models_identical(self):
         """
-        this tests that the ensemble model works with two identical model-types with identical states, inputs, etc., one with slightly altered parameters
+        This tests that the ensemble model works with two identical model-types with identical states, inputs, etc., one with slightly altered parameters.
+
+        The result is that each state, output, etc. is combined using the specified aggregation method.
         """
         m = OneInputOneOutputOneEventLM()
         m2 = OneInputOneOutputOneEventLM(x0={'x1': 2})
 
-        # Make sure they're not the same - 3x the impact
+        # Make sure they're not the same - matricies are 3x their original values
+        # The result is a model where state changes 3x as fast.
+        # Event state degrades 9x as fast, since B and F compound
         m2.B = np.array([[3]])
         m2.C = np.array([[3]])
         m2.F = np.array([[-0.3]])
@@ -61,44 +69,46 @@ class TestEnsemble(unittest.TestCase):
         self.assertEqual(em.events, m.events)
 
         # The resulting initial state should be exactly between the two:
-        x0 = em.initialize()
-        self.assertEqual(x0['x1'], 1)
+        x_t0 = em.initialize()
+        self.assertEqual(x_t0['x1'], 1)
 
         # Same with state transition
         u = em.InputContainer({'u1': 1})
-        x1 = em.next_state(x0, u, 1)
+        x_t1 = em.next_state(x_t0, u, 1)
         # m would give 2, m2 would give 4
-        self.assertEqual(x1['x1'], 3)
+        self.assertEqual(x_t1['x1'], 3)
 
         # Same with output
-        z = em.output(x1)
+        z = em.output(x_t1)
         # m would give 3, m2 would give 9
         self.assertEqual(z['z1'], 6)
 
         # Same with event state
-        es = em.event_state(x1)
+        es = em.event_state(x_t1)
         # m would give 0.7, m2 would give 0.1
         self.assertEqual(es['x1 == 10'], 0.4)
 
         # performance metrics
-        pm = em.performance_metrics(x1)
+        pm = em.performance_metrics(x_t1)
         self.assertEqual(pm['pm1'], 4)
 
         # Time of event 
-        toe = em.time_of_event(x0, lambda t, x=None: u, dt=1e-3)
+        toe = em.time_of_event(x_t0, lambda t, x=None: u, dt=1e-3)
         self.assertAlmostEqual(toe['x1 == 10'], 4.8895)
 
         # threshold met should be false
-        self.assertFalse(em.threshold_met(x1)['x1 == 10'])
+        self.assertFalse(em.threshold_met(x_t1)['x1 == 10'])
 
         # Transition again
-        x2 = em.next_state(x1, u, 1)
+        x_t2 = em.next_state(x_t1, u, 1)
         # threshold met should be true (because one of 2 models says it is)
-        self.assertTrue(em.threshold_met(x2)['x1 == 10'])
+        self.assertTrue(em.threshold_met(x_t2)['x1 == 10'])
 
     def test_two_models_different(self):
         """
[…]
