+++ tests/test_estimate_params.py
@@ -582,21 +582,29 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         
+        # Keys within a list of tuples without any commas
+        # Same as writing - ['thrower_height', 'throwing_speed', 'g']
         keys = [('thrower_height'), ('throwing_speed'), ('g')]
         m.estimate_params(data, keys, bounds=bound)  
-
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 4)
 
+        # Testing keys within tuples that are a length of one
         keys = [('thrower_height', ), ('throwing_speed', ), ('g', )]
         with self.assertRaises(ValueError):
             m.estimate_params(data, keys, bounds=bound)
 
+        # Testing lists within lists
         keys = [['thrower_height'], ['throwing_speed'], ['g']]
         with self.assertRaises(TypeError):
             m.estimate_params(data, keys, bounds=bound)
 
+
     def test_parameters(self):
+        """
+        Testing if passing in other keyword arguments works as intended
+        """
+
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
         data = [(results.times, results.inputs, results.outputs)]
@@ -608,12 +616,20 @@ class TestEstimateParams(unittest.TestCase):
         keys = ['thrower_height', 'throwing_speed', 'g']
         bound=((0, 8), (20, 42), (-20, -5))
 
-        # Testing that defualt method passed in works as intended
+        # Note that not every one of the keys would comply with this system
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=bound, method='Nelder-Mead')
+        m.estimate_params(data, keys, bounds=bound, method='TNC')
+        saveError = m.calc_error(results.times, results.inputs, results.outputs)
+
+        # Testing that default method passed in works as intended
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=bound)
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+
+        self.assertNotEqual(saveError, m.calc_error(results.times, results.inputs, results.outputs))
 
         # Passing 'TNC' method along with 'maxfun 1000' as our options
         # to see what happens if there are changed when we pass in 
@@ -622,13 +638,8 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys, bounds=bound, method='TNC', options={'maxfun': 1000, 'disp': False})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-
-        # Note that not every one of the keys would comply with this system
-        m.parameters['thrower_height'] = 1.5
-        m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=bound, method='TNC')
-        # if all(abs(m.parameters[key] - gt[key]) <= 0.02 for key in keys):
-        #     self.fail("m.parameter shouldn't be too close to the original parameters")
+        
+        self.assertGreater(saveError, m.calc_error(results.times, results.inputs, results.outputs))
 
         # Passing in Method that does not exist
         with self.assertRaises(ValueError):
@@ -637,18 +648,9 @@ class TestEstimateParams(unittest.TestCase):
         # Reset incorrect parameters
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-
-        # Additionally, remove the key 'g' that was never set and its bounds
         keys = ['thrower_height', 'throwing_speed']
         bound=((0, 8), (20, 42))
-
[…]
