### 2023-04-11 523bd73
removing the 'set' tests, and reformatting a bit.
