### 2023-04-11 8cac797
Removing checks to see if new estimated params are close to og
