+++ tests/test_estimate_params.py
@@ -20,49 +20,100 @@ class TestEstimateParams(unittest.TestCase):
         # Now lets incorrectly set some parameters
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.parameters['g'] = -8.5
-        keys = ['thrower_height', 'throwing_speed', 'g']
+        keys = ['thrower_height', 'throwing_speed']
         m.estimate_params(data, keys)
         for key in keys: # using assert not equal also works.
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
+        # Incorrectly set parameters, but provide a bounds for what said values should be.
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.parameters['g'] = -8.5
-
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, -5)))
-        tracker1 = m.calc_error(results.times, results.inputs, results.outputs)
-
-        m.parameters['thrower_height'] = 1.5
-        m.parameters['throwing_speed'] = 25
-        m.parameters['g'] = -8.5
-
-        m.estimate_params(data, keys, bounds=((-40, 40), (-40, 40), (-40, 40)))
-        tracker2 = m.calc_error(results.times, results.inputs, results.outputs)
-
-
-        # Demonstrates further limitations of Parameter Estiamtion
-        m.parameters['thrower_height'] = 1.5
-        m.parameters['throwing_speed'] = 25
-        m.parameters['g'] = -8.5
-
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 41.99), (-20, -5)))
+        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
+        # Checking to see if our estimated parameters are wihtin the bounds.
+        self.assertLessEqual(m.parameters['thrower_height'], 4)
+        self.assertGreaterEqual(m.parameters['thrower_height'], 0)
+        self.assertLessEqual(m.parameters['throwing_speed'], 42)
+        self.assertGreaterEqual(m.parameters['throwing_speed'], 20)
 
-        # Now with limits that do include the true values
-        # only returning the upper bounds? Why after this defined behavior?
-        # m.parameters['throwing_speed'] = 15
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.estimate_params(data, keys, bounds=((0, 4), (40, 40)))
+        for key in keys:
+            self.assertAlmostEqual(m.parameters[key], gt[key], 1)
+        self.assertLessEqual(m.parameters['thrower_height'], 4)
+        self.assertGreaterEqual(m.parameters['thrower_height'], 0)
+        self.assertEqual(m.parameters['throwing_speed'], 40)
+
+        # Demonstrates further limitations of Parameter Estimation
+        m.parameters['thrower_height'] = 5
+        m.parameters['throwing_speed'] = 19
+        m.estimate_params(data, keys, bounds=((1.231, 4), (20, 41.99)))
+
+        # Notice how the estimated parameters equal to the upper bounds of their respective bounds.
+        self.assertEqual(m.parameters['thrower_height'], 4)
+        self.assertEqual(m.parameters['throwing_speed'], 41.99)
 
         # Or two calls to estimate_params would resolve this issue
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.parameters['g'] = -8.5
-
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, 10)))
[…]
