+++ tests/test_estimate_params.py
@@ -926,23 +926,21 @@ class TestEstimateParams(unittest.TestCase):
         # Now lets reset some parameters
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
-        m.parameters['g'] = -8
+        m.parameters['g'] = 10
 
         # High tolerance would result in a higher calc_error
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs,
                           bounds=bound, keys = keys, tol = 100)
-        if all(abs(m.parameters[key] - gt[key]) <= 0.02 for key in keys):
+        if not all(abs(m.parameters[key] - gt[key]) > 0.02 for key in keys):
             raise ValueError("m.parameter shouldn't be too close to the original parameters")
         check = m.calc_error(results.times, results.inputs, results.outputs)
 
         # Reset parameters
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
-        m.parameters['g'] = -15
+        m.parameters['g'] = 10
 
-
-        # Including bounds has this feature where
-        # Low tolernace would result in a lower calc_error
+        # Not including bounds works as intended here, whereas including bounds does not get a good fit for the parameters.
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = 1e-9)
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
@@ -956,6 +954,9 @@ class TestEstimateParams(unittest.TestCase):
                               bounds=bound, keys = keys , tol = "12")
 
         # When tolerance is in a list, it rasies a ValueError.
+        # with self.assertRaises(TypeError):
+        #     m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, 
+        #                       bounds=bound, keys = keys , tol = [(1)])             
 
         # These cases are working?
         m.parameters['thrower_height'] = 3.1
@@ -967,7 +968,9 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
-        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = [1])
+
+        # Tolerance works as intended as long as it is within a sequence of length 1
+        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = [15])
         hold2 = m.calc_error(results.times, results.inputs, results.outputs)
 
         # self.assertEqual(hold2, check2)
@@ -976,21 +979,37 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
-        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = (1))
+        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = 15)
         hold1 = m.calc_error(results.times, results.inputs, results.outputs)
 
         self.assertEqual(hold1, hold2)
 
+        m.parameters['thrower_height'] = 3.1
+        m.parameters['throwing_speed'] = 29
+        m.parameters['g'] = 10
+        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = (15))
+        hold2 = m.calc_error(results.times, results.inputs, results.outputs)
 
+        self.assertEqual(hold1, hold2)
+
+        # Cannot pass Sets into tolerance
[…]
