+++ src/prog_models/prognostics_model.py
@@ -1233,8 +1233,6 @@ class PrognosticsModel(ABC):
             raise ValueError(f"Can not pass in keys as a Set. Sets are unordered by construction, so bounds may be out of order.")
         
         for key in keys:
-            # if isinstance(key, Sequence):
-            #     raise ValueError(f"Key '{key}' cannot be a Sequence")
             if key not in self.parameters:
                 raise ValueError(f"Key '{key}' not in model parameters")
 
@@ -1249,7 +1247,14 @@ class PrognosticsModel(ABC):
             raise TypeError(f"Times, Inputs, and Outputs cannot be a Set. Sets are unordered by definition, so passing in arguments as Sets may have undefined behavior.")
 
         # if parameters not in parent wrapper sequence, then place them into one.
+        if isinstance(times, np.ndarray):
+            times = times.tolist()
+        if isinstance(inputs, np.ndarray):
+            inputs = inputs.tolist()
+        if isinstance(outputs, np.ndarray):
+            outputs = outputs.tolist()
         if not runs and times and inputs and outputs:
+            # for i, (times, inputs, outputs) in runs:
             if not isinstance(times[0], (Sequence, np.ndarray)):
                 times = [times]
             if not isinstance(inputs[0], (Sequence, np.ndarray)):
