+++ .github/workflows/python-package.yml
@@ -111,7 +111,7 @@ jobs:
       run: |
         python scripts/test_copyright.py
   coverage:
-    timeout-minutes: 45
+    timeout-minutes: 30
     runs-on: ubuntu-latest
     strategy:
       matrix:
