+++ src/prog_models/linear_model.py
@@ -3,7 +3,7 @@
 
 from abc import ABC, abstractmethod
 import numpy as np
-from prog_models import PrognosticsModel
+from prog_models.prognostics_model import PrognosticsModel
 
 
 class LinearModel(PrognosticsModel, ABC):
