+++ src/prog_models/utils/containers.py
@@ -3,111 +3,177 @@
 
 import numpy as np
 from typing import Union
+import pandas as pd
 
-from ..exceptions import ProgModelTypeError
+from prog_models.exceptions import ProgModelTypeError
 
 
 class DictLikeMatrixWrapper():
     """
     A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-    
-    Arguments
-    ---------
-    keys: list
-        The keys of the dictionary. e.g., model.states or model.inputs
-    data: dict or numpy array
-        The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
+
+    Arguments:
+        keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
+        data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
-    def __init__(self, keys : list, data : Union[dict, np.array]):
+
+    def __init__(self, keys: list, data: Union[dict, np.array, pd.Series]):
+        """
+        Initializes the container
+        """
         if not isinstance(keys, list):
-            keys = list(keys)
-
-        self._keys = keys.copy()
+            keys = list(keys)  # creates list with keys
+        temp_keys = keys.copy()
         if isinstance(data, np.matrix):
-            self.matrix = np.array(data, dtype=np.float64)
-        elif isinstance(data, np.ndarray):
-            if data.ndim == 1:
-                data = data[np.newaxis].T
-            self.matrix = data
-        elif isinstance(data, (dict, DictLikeMatrixWrapper)):
-            self.matrix = np.array(
-                [
-                    [data[key]] if key in data else [None] for key in keys
-                ], dtype=np.float64)
+            self.data = pd.DataFrame(np.array(data, dtype=np.float64), temp_keys)
+        elif isinstance(data, np.ndarray):  # data is a multidimensional array, in column vector form
+            self.data = pd.DataFrame(data, temp_keys)
+        elif isinstance(data, (dict, DictLikeMatrixWrapper)):  # data is not in column vector form
+            self.data = pd.DataFrame(data, index=[0]).T
         else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")     
-
+            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
+        self.matrix = self.data.to_numpy()
+        self._keys = self.data.index.to_list()
     def __reduce__(self):
-        return (DictLikeMatrixWrapper, (self._keys, self.matrix))
+        """
+        reduce is overridden for pickles
+        """
+        keys = self.data.index.to_list()
+        matrix = self.data.to_numpy()
+        return (DictLikeMatrixWrapper, (keys, matrix))
+    def __getitem__(self, key: str) -> int:
+        """
+        get all values associated with a key, ex: all values of 'i'
+        """
+        row = self.data.loc[key].to_list()  # creates list from a row of the DataFrame data
+        if len(self.data.loc[key]) == 1:  # list contains 1 value, returns that value (non-vectorized)
+            return self.data.loc[key, 0]
+        return row  # returns entire row/list (vectorized case)
 
-    def __getitem__(self, key : str) -> int:
-        row = self.matrix[self._keys.index(key)]
-        if len(row) == 1:
-            return self.matrix[self._keys.index(key)][0]
-        return self.matrix[self._keys.index(key)]
+    def __setitem__(self, key: str, value: int) -> None:
+        """
+        sets a row at the key given
+        """
[…]
