### 2023-04-12 092d19e
Delete battery_circuit_df.py
accidental
