### 2023-04-12 91aaded
Removing battery test.
