+++ src/prog_models/loading_fcns/controllers.py
@@ -4,7 +4,8 @@ Controllers
 
 # IMPORTS
 # ========
-from utilities.imports_ import np
+# from utilities.imports_ import np
+import numpy as np
 
 
 # LQR general function
@@ -52,7 +53,7 @@ def lqr_fn(A, B, Q, R):
 
     # --------- Partition matrix of resulting eigenvectors ------------ #
     X = V_[:n, :n] # first half (row-wise)
-    Y = V_[n:, :n] # second half (row-wise)
+    Y = V_[n:, :n] # second half (row-wise) 
 
     # ----------- Estimate control gain ----------------- #
     K = np.dot( np.dot( np.dot(R_inv, B.T), Y ), np.linalg.inv(X) )
@@ -63,51 +64,50 @@ def lqr_fn(A, B, Q, R):
     return K, E
 
     
-# CONTROLLERS: FULL STATE VECTOR CONTROL  (i.e., no powertrain)
-# ==============================================================
-class LQR():
-    """ Linear Quadratic Regulator"""
-    def __init__(self, n_states, n_inputs, Q=None, R=None) -> None:
-        self.type = 'LQR'
-        if Q is None:       Q = np.eye(n_states)
-        if R is None:       R = np.eye(n_inputs)
+# # CONTROLLERS: FULL STATE VECTOR CONTROL  (i.e., no powertrain)
+# # ==============================================================
+# class LQR():
+#     """ Linear Quadratic Regulator"""
+#     def __init__(self, n_states, n_inputs, Q=None, R=None) -> None:
+#         self.type = 'LQR'
+#         if Q is None:       Q = np.eye(n_states)
+#         if R is None:       R = np.eye(n_inputs)
 
-        self.Q = Q
-        self.R = R
+#         self.Q = Q
+#         self.R = R
         
     
-    def compute_gain(self, A, B):
-        """ Compute controller gain given state of the system described by linear model A, B"""
-        self.K, self.E = lqr_fn(A, B, self.Q, self.R)
-        return self.K, self.E
+#     def compute_gain(self, A, B):
+#         """ Compute controller gain given state of the system described by linear model A, B"""
+#         self.K, self.E = lqr_fn(A, B, self.Q, self.R)
+#         return self.K, self.E
 
-    def compute_input(self, gain, state_error):
-        """ Compute system input given the controller gain 'gain' and the error w.r.t. the reference state 'error' """
-        return - np.dot(gain, state_error)
+#     def compute_input(self, gain, state_error):
+#         """ Compute system input given the controller gain 'gain' and the error w.r.t. the reference state 'error' """
+#         return - np.dot(gain, state_error)
     
-    def build_scheduled_control(self, system_linear_model_fun, input_vector, state_vector_vals=None, index_scheduled_var=None):
+#     def build_scheduled_control(self, system_linear_model_fun, input_vector, state_vector_vals=None, index_scheduled_var=None):
 
-        if state_vector_vals is None:
-            # using psi (yaw angle) as scheduled variable as the LQR control cannot work with yaw=0 since it's in the inertial frame.
-            n_schedule_grid = 360*2 + 1
-            index_scheduled_var = 5
-            state_vector_vals = np.zeros((self.n_states, n_schedule_grid))
-            state_vector_vals[index_scheduled_var, :] = np.linspace(-2.0*np.pi, 2.0*np.pi, n_schedule_grid)
+#         if state_vector_vals is None:
+#             # using psi (yaw angle) as scheduled variable as the LQR control cannot work with yaw=0 since it's in the inertial frame.
+#             n_schedule_grid = 360*2 + 1
+#             index_scheduled_var = 5
[…]
