### 2023-04-12 00fcba8
Removed battery test
