+++ tests/test_estimate_params.py
@@ -521,6 +521,15 @@ class TestEstimateParams(unittest.TestCase):
         # Testing bounds equality with strings in np.array
         m.estimate_params(data, keys, bounds=[np.array(['9', '9']), np.array(['2', '3']), np.array(['4', '5'])])
 
+        # Resetting parameters
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
+        m.parameters['g'] = -8
+        times = np.array([0.0, 0.5, 1.0])
+        inputs = np.array([{}, {}, {}])
+        outputs = np.array([{'x': 1.83}, {'x': 21.83}, {'x': 38.78612068965517}])
+        m.estimate_params(times = times, inputs = inputs, outputs = outputs)
+
 
 # Testing features where keys are not parameters in the model
     def test_keys(self):
@@ -686,6 +695,9 @@ class TestEstimateParams(unittest.TestCase):
         self.assertAlmostEqual(saveError, m.calc_error(results.times, results.inputs, results.outputs), delta = 0.00001)
 
     def test_multiple_runs(self):
+        """
+        In this test, we are examining the behavior of estimate_params when there are multiple runs.
+        """
         m = ThrownObject()
 
         # The value of time1, time2, inputs, and outputs are arbitrary values
@@ -755,6 +767,7 @@ class TestEstimateParams(unittest.TestCase):
             str(cm.exception)
         )
 
+        # Incorrect lengths for times.
         with self.assertRaises(ValueError):
             m.estimate_params(times=[time1, [time2]], inputs=inputs, outputs=outputs)
         self.assertEqual(
@@ -771,9 +784,33 @@ class TestEstimateParams(unittest.TestCase):
         time2 = [0, 1, 2, 3]
 
         # Confirming estimate_params works when different runs are passed in as different data types.
+        # Passing in time1 as a np.array datatype.
         m.estimate_params(times=[time1, time2], inputs=inputs, outputs=outputs)
 
-        # Another test case that would be fixed with future changes to Containers
+        time1 = [0, 1, 2, 4, 5, 6, 7, 8, 9]
+        time2 = [0, 1, 2, 3]
+
+        inputs = [[{}]*9, [{}]*4]
+        outputs = [[{'x': 1.83},
+            {'x': 36.95},
+            {'x': 62.36},
+            {'x': 77.81},
+            {'x': 83.45},
+            {'x': 79.28},
+            {'x': 65.3},
+            {'x': 41.51},
+            {'x': 7.91},], 
+            np.array([
+                {'x': 1.83},
+                {'x': 36.95},
+                {'x': 62.36},
+                {'x': 77.81},
+            ])]
+
+        # Passing in an np.array structure as a runs for outputs.
+        m.estimate_params(times=[time1, time2], inputs=inputs, outputs=outputs)
+
+        # Test case that would be fixed with future changes to Containers
         # with self.assertRaises(ValueError):
         #     m.estimate_params(times=[incorrectTimesLen], inputs=[inputs], outputs=[outputs])
 
@@ -850,6 +887,15 @@ class TestEstimateParams(unittest.TestCase):
             str(cm.exception)
         )
 
+        # Length error expected, 1, 9, 1.
+        with self.assertRaises(ValueError) as cm:
+            m.estimate_params(times=[[results.times]], inputs=[results.inputs], outputs=[[results.outputs]])
+        self.assertEqual(
+            'Times, inputs, and outputs must be same length for the run at index 0. Length of times: 1, Length of inputs: 9, Length of outputs: 1',
+            str(cm.exception)
+        )
+
+
         # Passing in incorrect Runs
         with self.assertRaises(ValueError):
             m.estimate_params(wrongData)
@@ -883,28 +929,26 @@ class TestEstimateParams(unittest.TestCase):
[…]
