+++ tests/__main__.py
@@ -67,7 +67,6 @@ if __name__ == '__main__':
         was_successful = False
 
     try:
-    
         examples_main()
     except Exception:
         was_successful = False
