### 2023-04-12 a440169
Update prognostics_model.py
