+++ src/prog_models/prognostics_model.py
@@ -1211,6 +1211,8 @@ class PrognosticsModel(ABC):
                 Array of output containers where output[x] corresponds to time[x]
             method (str, optional):
                 Optimization method- see scipy.optimize.minimize for options
+            tol (int, optional):
+                Tolerance for termination. Depending on the provided minimization method, specifying tolerance sets solver-specific options to tol
             bounds (tuple or dict, optional):
                 Bounds for optimization in format ((lower1, upper1), (lower2, upper2), ...) or {key1: (lower1, upper1), key2: (lower2, upper2), ...}
             options (dict, optional):
