+++ src/prog_models/utils/containers_df.py
@@ -27,6 +27,8 @@ class DictLikeMatrixWrapper():
         if isinstance(data, np.matrix):
             self.data = pd.DataFrame(np.array(data, dtype=np.float64), temp_keys)
         elif isinstance(data, np.ndarray):  # data is a multidimensional array, in column vector form
+            if data.ndim == 1:
+                data = data[np.newaxis].T
             self.data = pd.DataFrame(data, temp_keys)
         elif isinstance(data, (dict, DictLikeMatrixWrapper)):  # data is not in column vector form
             self.data = pd.DataFrame(data, index=[0]).T
@@ -54,7 +56,8 @@ class DictLikeMatrixWrapper():
         """
         sets a row at the key given
         """
-        self.data.loc[key] = np.atleast_1d(value)  # using the key to find the Series location
+        self.data.loc[key] = np.atleast_1d(value)  # using the key to find the DataFrame location
+        self.matrix = self.data.to_numpy()
 
 
     def __delitem__(self, key: str) -> None:
@@ -62,6 +65,8 @@ class DictLikeMatrixWrapper():
         removes row associated with key
         """
         self.data = self.data.drop(index=[key])
+        self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
+        self._keys.remove(key)
 
     def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
         """
@@ -159,6 +164,8 @@ class DictLikeMatrixWrapper():
                 # the key
                 temp_df = DictLikeMatrixWrapper([key], {key: other.data.loc[key, 0]})
                 self.data = pd.concat([self.data, temp_df.data])
+        self._keys = self.data.index.to_list()
+        self.matrix = self.data.to_numpy()
 
     def __contains__(self, key: str) -> bool:
         """
