+++ tests/test_estimate_params.py
@@ -1137,6 +1137,7 @@ def run_tests():
     unittest.main()
     
 def main():
+
     l = unittest.TestLoader()
     runner = unittest.TextTestRunner()
     print("\n\nTesting EstimateParams Feature")
@@ -1146,4 +1147,18 @@ def main():
         raise Exception("Failed test")
 
 if __name__ == '__main__':
-    main()
+    import cProfile
+    cProfile.run('main()', "output.dat")
+
+    import pstats
+    from pstats import SortKey
+
+    with open("output_time.txt", 'w') as f:
+        p = pstats.Stats("output.dat", stream=f)
+        p.sort_stats("time").print_stats()
+
+    with open("output_calls.txt", 'w') as f:
+        p = pstats.Stats("output.dat", stream=f)
+        p.sort_stats("calls").print_stats()
+
+    # main()
