### 2023-04-13 3ee6fe8
Update src/prog_models/composite_model.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
