+++ tests/test_examples.py
@@ -52,4 +52,15 @@ def main():
         raise Exception("Failed test")
 
 if __name__ == '__main__':
-    main()
+    import cProfile
+    cProfile.run('main()', "output.dat")
+
+    import pstats
+
+    with open("output_time.txt", 'w') as f:
+        p = pstats.Stats("output.dat", stream=f)
+        p.sort_stats("time").print_stats()
+
+    with open("output_calls.txt", 'w') as f:
+        p = pstats.Stats("output.dat", stream=f)
+        p.sort_stats("calls").print_stats()
