+++ tests/test_estimate_params.py
@@ -31,7 +31,7 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-        # Checking to see if our estimated parameters are wihtin the bounds.
+        # Checking to see if our estimated parameters are within the bounds.
         self.assertLessEqual(m.parameters['thrower_height'], 4)
         self.assertGreaterEqual(m.parameters['thrower_height'], 0)
         self.assertLessEqual(m.parameters['throwing_speed'], 42)
@@ -62,20 +62,20 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-        # When our incorrecltly set parameters that are not within the bounds themselves
+        # When our incorrectly set parameters that are not within the bounds themselves
         m.parameters['thrower_height'] = 10
         m.parameters['throwing_speed'] = -10
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
         # These returned values are still within the bounds, they are just not close to the original value at all.
-        # This results in our estimate parameters to result in the upper extremas
+        # This results in our estimate parameters to result in the upper extremes
         self.assertEqual(m.parameters['thrower_height'], 4)
         self.assertEqual(m.parameters['throwing_speed'], 42)
 
         # Show casing results of having a local min/max
-        # Even though all our bounds accomodate for the original model params values,
-        # our estimate_parmams returns lower/upper bound values.
+        # Even though all our bounds accommodate for the original model params values,
+        # our estimate_params returns lower/upper bound values.
         m.parameters['thrower_height'] = 4
         m.parameters['throwing_speed'] = 24
         m.parameters['g'] = -20
@@ -225,7 +225,7 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(ValueError):
             m.estimate_params(data, keys, bounds=((30, 7), (40, 20), (0, -20)))
 
-        #Implememnt different variations of lists and tuples and see if they work as intended
+        # Implement different variations of lists and tuples and see if they work as intended
         # Try incomplete list:
         with self.assertRaises(ValueError):
             # Missing bound
@@ -274,7 +274,7 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(ValueError):
             m.estimate_params(data, keys, bounds=None)
         with self.assertRaises(ValueError):
-            # Tuple isn't of size 2, more specfically, tuple is size less than 2
+            # Tuple isn't of size 2, more specifically, tuple is size less than 2
             m.estimate_params(data, keys, bounds={'g': (7,)})
         with self.assertRaises(ValueError):
             # Tuple is a size greater than 2
@@ -349,11 +349,11 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(TypeError):
             m.estimate_params(data, keys, bounds=[[-1, 5], (20, 40), {-5, 15}])
 
-        # Lower Boung greater than Upper Bound
+        # Lower Bound greater than Upper Bound
         with self.assertRaises(ValueError):
             m.estimate_params(data, keys, bounds=[[4, 0], [-20, 20], [0, 40]])
 
-        # Incorrect lenght given for bounds
+        # Incorrect length given for bounds
         with self.assertRaises(ValueError):
[…]
