### 2023-04-13 4baf4ee
Update linear_model.py
path update
