+++ src/prog_models/models/dcmotor.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from prog_models.prognostics_model import PrognosticsModel
+from prog_models import PrognosticsModel
 
 RAD_TO_DEG = 180/np.pi
 PI2 = 2 * np.pi
