### 2023-04-13 b077a5f
Update battery_electrochem.py
path update
