### 2023-04-13 cf5284d
Update dcmotor_singlephase.py
path update
