+++ src/prog_models/models/esc.py
@@ -5,7 +5,7 @@ from math import floor
 import numpy as np
 import scipy.signal as signal
 
-from prog_models.prognostics_model import PrognosticsModel
+from prog_models import PrognosticsModel
 
 RAD_TO_DEG = 180/np.pi
 PI2 = 2 * np.pi
