### 2023-04-13 25909c7
Update propeller_load.py
path update
