+++ src/prog_models/models/propeller_load.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from prog_models.prognostics_model import PrognosticsModel
+from prog_models import PrognosticsModel
 
 
 def update_Cq(params):
