### 2023-04-13 a7fd81c
Update thrown_object.py
path update
