+++ src/prog_models/models/thrown_object.py
@@ -2,8 +2,7 @@
 
 import numpy as np
 
-from prog_models.prognostics_model import PrognosticsModel
-from prog_models.linear_model import LinearModel
+from prog_models import PrognosticsModel, LinearModel
 
 def calc_lumped_param(params):
     return {
