+++ tests/test_calc_error.py
@@ -1,0 +1,230 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the
+# National Aeronautics and Space Administration.  All Rights Reserved.
+
+import numpy as np
+import unittest
+
+from prog_models import *
+from prog_models.models import *
+
+
+class TestCalcError(unittest.TestCase):
+    """
+    Main Testing class for calc_error.
+
+    Validating that values are correctly being passed into the new calc_error calls and that we are receiving expected results!
+    """
+
+    # @unittest.skip
+    def test_calc_error(self):
+        # Note, lowering time steps or increasing simulate threshold may cause this model to not run (takes too long)
+        m = BatteryElectroChemEOD()
+
+        options = {
+            'save_freq': 200, # Frequency at which results are saved
+            'dt': 1, # Time step
+        }
+
+        def future_loading(t, x=None):
+            if (t < 600):
+                i = 2
+            elif (t < 900):
+                i = 1
+            elif (t < 1800):
+                i = 4
+            elif (t < 3000):
+                i = 2
+            else:
+                i = 3
+            return m.InputContainer({'i': i})
+    
+        simulated_results = m.simulate_to(2000, future_loading, **options)
+
+        # Running calc_error before setting incorrect parameters
+        m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
+
+        # Initializing parameters to very erroneous values       
+        m.parameters['qMax'] = 4000
+        keys = ['qMax']
+
+        # Before running estimate_params
+        with self.assertRaises(ValueError):
+            m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
+
+        m.estimate_params([(simulated_results.times, simulated_results.inputs, simulated_results.outputs)], keys, dt=0.5)
+
+        # After running estimate_params. Note that this would not change the outcome of the result
+        with self.assertRaises(ValueError):
+            m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)    
+
+        # Running through various dt values that could work
+        for i in np.arange(0, 1, 0.1):
+            with self.assertRaises(ValueError):
+                m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=i)
+
+        for i in range(2, 10):
+            with self.assertRaises(ValueError):
+                m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=i)
+ 
+        # Creating duplicate model
+        m1 = BatteryElectroChemEOD()
+
+        orig_params = m1.parameters.copy()
+
+        # Much bigger parameter initialization
+        m.parameters['kp'] = m1.parameters['kp'] = 10000
+        m.parameters['kn'] = m1.parameters['kn'] = 1000 
+        m.parameters['qpMax'] = m1.parameters['qpMax'] = 4500
+        m.parameters['qMax'] = m1.parameters['qMax'] = 9000
+        keys = ['kp', 'kn', 'qpMax','qMax']
+
+        change_params = m1.parameters.copy()
+
+        simulated_results = m.simulate_to(2000, future_loading, **options)
+        m1_sim_results = m1.simulate_to(2000, future_loading, **options)
+
+        data = [(simulated_results.times, simulated_results.inputs, simulated_results.outputs)]
+        data_m1 = [(m1_sim_results.times, m1_sim_results.inputs, m1_sim_results.outputs)]
+
+        # Check out the warnings that are occurring here...
+        # They are being spammed almost. Increasing save_frequency increases spam
+
[…]
