### 2023-04-13 d6eeca3
Update src/prog_models/prognostics_model.py
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
