### 2023-04-13 f84bf98
Apply suggestions from code review
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
