+++ tests/test_estimate_params.py
@@ -22,7 +22,7 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['throwing_speed'] = 25
         keys = ['thrower_height', 'throwing_speed']
         m.estimate_params(data, keys)
-        for key in keys: # using assert not equal also works.
+        for key in keys: 
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
         # Incorrectly set parameters, but provide a bounds for what said values should be.
@@ -55,14 +55,13 @@ class TestEstimateParams(unittest.TestCase):
         self.assertEqual(m.parameters['thrower_height'], 4)
         self.assertEqual(m.parameters['throwing_speed'], 41.99)
 
-        # Or two calls to estimate_params would resolve this issue
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-        # When our incorrecltly set parameters that are not within the bounds themselves
+        # Testing initial parameters that are not within the defined bounds; estimated parameters are not a good fit
         m.parameters['thrower_height'] = 10
         m.parameters['throwing_speed'] = -10
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
@@ -73,8 +72,8 @@ class TestEstimateParams(unittest.TestCase):
         self.assertEqual(m.parameters['thrower_height'], 4)
         self.assertEqual(m.parameters['throwing_speed'], 42)
 
-        # Show casing results of having a local min/max
-        # Even though all our bounds accomodate for the original model params values,
+        # Showcasing results of having a local min/max
+        # Even though all our bounds include the original model params values,
         # our estimate_parmams returns lower/upper bound values.
         m.parameters['thrower_height'] = 4
         m.parameters['throwing_speed'] = 24
@@ -156,12 +155,6 @@ class TestEstimateParams(unittest.TestCase):
             str(cm.exception)
         )
         with self.assertRaises(ValueError) as cm:
-            m.estimate_params(times='', inputs='', outputs='')
-        self.assertEqual(
-            'Times, inputs, and outputs must have at least one element',
-            str(cm.exception) 
-        )
-        with self.assertRaises(ValueError) as cm:
             m.estimate_params(times=[[]], inputs=[[]], outputs=[[]])
         self.assertEqual(
             'Times, inputs, and outputs for Run 0 must have at least one element',
@@ -203,7 +196,7 @@ class TestEstimateParams(unittest.TestCase):
             str(cm.exception)
         )
 
-        # Now with limits that dont include the true values
+        # Now with bounds that don't include the true values
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
@@ -211,7 +204,7 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1)
 
-        # Now with limits that do include the true values
+        # Now with bounds that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
@@ -225,7 +218,7 @@ class TestEstimateParams(unittest.TestCase):
         with self.assertRaises(ValueError):
             m.estimate_params(data, keys, bounds=((30, 7), (40, 20), (0, -20)))
 
[…]
