+++ tests/test_estimate_params.py
@@ -31,7 +31,7 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-        # Checking to see if our estimated parameters are wihtin the bounds.
+        # Checking to see if our estimated parameters are within the bounds.
         self.assertLessEqual(m.parameters['thrower_height'], 4)
         self.assertGreaterEqual(m.parameters['thrower_height'], 0)
         self.assertLessEqual(m.parameters['throwing_speed'], 42)
@@ -68,13 +68,13 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
         # These returned values are still within the bounds, they are just not close to the original value at all.
-        # This results in our estimate parameters to result in the upper extremas
+        # This results in our estimate parameters to result in the upper extremes
         self.assertEqual(m.parameters['thrower_height'], 4)
         self.assertEqual(m.parameters['throwing_speed'], 42)
 
         # Showcasing results of having a local min/max
         # Even though all our bounds include the original model params values,
-        # our estimate_parmams returns lower/upper bound values.
+        # our estimate_params returns lower/upper bound values.
         m.parameters['thrower_height'] = 4
         m.parameters['throwing_speed'] = 24
         m.parameters['g'] = -20
@@ -204,6 +204,8 @@ class TestEstimateParams(unittest.TestCase):
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1)
 
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         # Now with bounds that do include the true values
         m.estimate_params(data, keys, bounds=((0, 8), (20, 42), (-20, -5)))
         for key in keys:
@@ -227,23 +229,30 @@ class TestEstimateParams(unittest.TestCase):
             # Extra bound
             m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, 0), (-20, 10)))
 
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         # Dictionary bounds
         m.estimate_params(data, keys, bounds={'thrower_height': (0, 4), 'throwing_speed': (20, 42), 'g': (-20, 0)})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         # Dictionary bounds - missing
         # Will fill with (-inf, inf)
         m.estimate_params(data, keys, bounds={'thrower_height': (0, 4), 'throwing_speed': (20, 42)})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         # Dictionary bounds - extra
         m.estimate_params(data, keys, bounds={'thrower_height': (0, 4), 'throwing_speed': (20, 42), 'g': (-20, 0), 'dummy': (-50, 0)})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
-
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
         # Behavior, sets in default values and does not error with correct number of bounds.
         # User changes a value that exists and should not be changed
[…]
