### 2023-04-13 93c652f
removing note to self
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
