### 2023-04-13 a842584
Update test_battery.py
