+++ src/prog_models/visualize.py
@@ -3,25 +3,26 @@
 
 # Import packages
 # ============
-from typing import List
+from typing import List, Union
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 
 # Set default options
 # ====================
-mpl.rcParams['lines.linewidth']  = 3
+mpl.rcParams['lines.linewidth'] = 3
 mpl.rcParams['lines.markersize'] = 8
-mpl.rcParams['font.size']        = 14
-mpl.rcParams['axes.labelsize']   = 'x-large'
-mpl.rcParams['legend.fontsize']  = 'large'
+mpl.rcParams['font.size'] = 14
+mpl.rcParams['axes.labelsize'] = 'x-large'
+mpl.rcParams['legend.fontsize'] = 'large'
 mpl.rcParams['figure.titlesize'] = 'x-large'
-mpl.rcParams['figure.figsize']   = [10.0, 9.0]
-mpl.rcParams['figure.dpi']       = 100
-mpl.rcParams['savefig.dpi']      = 300
+mpl.rcParams['figure.figsize'] = [10.0, 9.0]
+mpl.rcParams['figure.dpi'] = 100
+mpl.rcParams['savefig.dpi'] = 300
+
 
 # VISUALIZE FUNCTIONS
 # ==========================
-def get_subplot_dim(num_subplots : int, rowfirst : bool = True) -> tuple:
+def get_subplot_dim(num_subplots: int, rowfirst: bool = True) -> tuple:
     """
     Compute the number of rows and columns (nrows, ncols) for a figure with multiple subplots.
     The function returns number of rows and columns given num_subplots.
@@ -60,18 +61,19 @@ def get_subplot_dim(num_subplots : int, rowfirst : bool = True) -> tuple:
     """
     nrows, ncols = 1, 1  # initialize number of rows and cols to 1.
     if rowfirst:
-        while nrows * ncols < num_subplots:         
+        while nrows * ncols < num_subplots:
             nrows += 1
-            if nrows * ncols < num_subplots:        
+            if nrows * ncols < num_subplots:
                 ncols += 1
     else:
-        while nrows * ncols < num_subplots:         
+        while nrows * ncols < num_subplots:
             ncols += 1
-            if nrows * ncols < num_subplots:        
+            if nrows * ncols < num_subplots:
                 nrows += 1
     return nrows, ncols
 
-def set_plot_options(opt : dict) -> dict:
+
+def set_plot_options(opt: dict) -> dict:
     """
     Set default plot options by integrating the options specified by the user in 'opt'
     The visualize library works with specific values to generate the plots.
@@ -110,7 +112,7 @@ def set_plot_options(opt : dict) -> dict:
         opt_list = list(opt.keys())
     except Exception:
         opt, opt_list = {}, []
-    
+
     # Fill out all options if not provided
     if 'figsize' not in opt_list:
         opt['figsize'] = (10, 9)
@@ -132,7 +134,7 @@ def set_plot_options(opt : dict) -> dict:
         opt['tight_layout'] = False
     if 'display_labels' not in opt_list:
         opt['display_labels'] = 'all'
-    
+
     # if title should be displayed but title fontsize is not specified, add it to the dictionary
     if opt['title'] and ('title_fontsize' not in opt_list or not opt['title_fontsize']):
         if 'fontsize' in opt_list:
@@ -168,7 +170,8 @@ def set_plot_options(opt : dict) -> dict:
 
     return opt
 
-def set_legend_options(leg_opt : dict, s_names : List[str]) -> dict:
+
+def set_legend_options(leg_opt: dict, s_names: List[str]) -> dict:
     """
[…]

[…]
