+++ output_time.txt
@@ -1,0 +1,552 @@
+Mon Apr 17 03:56:21 2023    output.dat
+
+         6617680 function calls (6609738 primitive calls) in 1.839 seconds
+
+   Ordered by: internal time
+
+   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
+    21587    0.252    0.000    0.453    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:506(output)
+   707270    0.215    0.000    0.281    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
+    21221    0.173    0.000    0.656    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:331(dx)
+324256/324094    0.147    0.000    0.576    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
+   212994    0.101    0.000    0.175    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
+    10000    0.097    0.000    0.480    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:391(apply_limits)
+   110537    0.074    0.000    0.074    0.000 {method 'reduce' of 'numpy.ufunc' objects}
+   110493    0.067    0.000    0.219    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
+   153665    0.056    0.000    0.056    0.000 {built-in method numpy.array}
+    84652    0.048    0.000    0.072    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:19(__init__)
+   212994    0.047    0.000    0.312    0.000 <__array_function__ internals>:177(atleast_1d)
+   591895    0.047    0.000    0.047    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
+   707270    0.043    0.000    0.043    0.000 {method 'index' of 'list' objects}
+    21221    0.035    0.000    0.713    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:324(next_state)
+   110059    0.033    0.000    0.251    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2307(any)
+   100000    0.033    0.000    0.111    0.000 {method 'any' of 'numpy.generic' objects}
+930895/930542    0.031    0.000    0.032    0.000 {built-in method builtins.len}
+    10000    0.031    0.000    0.061    0.000 {method 'normal' of 'numpy.random.mtrand.RandomState' objects}
+   213036    0.029    0.000    0.029    0.000 {built-in method numpy.asanyarray}
+   212979    0.028    0.000    0.028    0.000 {method 'reshape' of 'numpy.ndarray' objects}
+   110059    0.027    0.000    0.353    0.000 <__array_function__ internals>:177(any)
+   110493    0.020    0.000    0.020    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
+    10000    0.018    0.000    0.091    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/noise_functions.py:39(normal_process_noise)
+    43174    0.016    0.000    0.016    0.000 {built-in method builtins.sum}
[…]

[…]
