+++ output_calls.txt
@@ -1,119 +1,119 @@
-Mon Apr 17 03:56:21 2023    output.dat
+Mon Apr 17 11:57:50 2023    output.dat
 
-         6617680 function calls (6609738 primitive calls) in 1.839 seconds
+         7013735 function calls (7005605 primitive calls) in 1.928 seconds
 
    Ordered by: call count
 
    ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-930895/930542    0.031    0.000    0.032    0.000 {built-in method builtins.len}
-   707270    0.043    0.000    0.043    0.000 {method 'index' of 'list' objects}
-   707270    0.215    0.000    0.281    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-   591895    0.047    0.000    0.047    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-324256/324094    0.147    0.000    0.576    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-315041/314957    0.015    0.000    0.016    0.000 {built-in method builtins.isinstance}
-   215108    0.009    0.000    0.009    0.000 {method 'append' of 'list' objects}
-   213036    0.029    0.000    0.029    0.000 {built-in method numpy.asanyarray}
-   212994    0.010    0.000    0.010    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:19(_atleast_1d_dispatcher)
-   212994    0.101    0.000    0.175    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-   212994    0.047    0.000    0.312    0.000 <__array_function__ internals>:177(atleast_1d)
-   212979    0.028    0.000    0.028    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-   153665    0.056    0.000    0.056    0.000 {built-in method numpy.array}
-   123525    0.006    0.000    0.006    0.000 {method 'items' of 'dict' objects}
-   110537    0.074    0.000    0.074    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-   110493    0.020    0.000    0.020    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
-   110493    0.067    0.000    0.219    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
+999485/999132    0.034    0.000    0.034    0.000 {built-in method builtins.len}
+   757782    0.045    0.000    0.045    0.000 {method 'index' of 'list' objects}
+   757782    0.230    0.000    0.300    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
+   638827    0.052    0.000    0.052    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
+342285/342123    0.152    0.000    0.591    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
+335230/335146    0.016    0.000    0.017    0.000 {built-in method builtins.isinstance}
+   233299    0.010    0.000    0.010    0.000 {method 'append' of 'list' objects}
+   231056    0.031    0.000    0.031    0.000 {built-in method numpy.asanyarray}
[…]

[…]
