+++ tests/test_calc_error.py
@@ -43,6 +43,14 @@ class TestCalcError(unittest.TestCase):
         # Running calc_error before setting incorrect parameters
         m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
 
+        with self.assertWarns(UserWarning) as cm:
+            m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt = 1, stability_tol = 10)
+        self.assertEqual(
+            "configurable cutoff must be some float value in the domain (0, 1]."  
+            " Received 10. Resetting value to 0.95",
+            str(cm.warning)
+        )
+
         # Initializing parameters to very erroneous values       
         m.parameters['qMax'] = 4000
         keys = ['qMax']
