+++ output_calls.txt
@@ -1,90 +1,90 @@
-Mon Apr 17 11:57:50 2023    output.dat
+Mon Apr 17 12:05:35 2023    output.dat
 
-         7013735 function calls (7005605 primitive calls) in 1.928 seconds
+         7451103 function calls (7442973 primitive calls) in 2.058 seconds
 
    Ordered by: call count
 
    ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-999485/999132    0.034    0.000    0.034    0.000 {built-in method builtins.len}
-   757782    0.045    0.000    0.045    0.000 {method 'index' of 'list' objects}
-   757782    0.230    0.000    0.300    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-   638827    0.052    0.000    0.052    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-342285/342123    0.152    0.000    0.591    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-335230/335146    0.016    0.000    0.017    0.000 {built-in method builtins.isinstance}
-   233299    0.010    0.000    0.010    0.000 {method 'append' of 'list' objects}
-   231056    0.031    0.000    0.031    0.000 {built-in method numpy.asanyarray}
-   231014    0.011    0.000    0.011    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:19(_atleast_1d_dispatcher)
-   231014    0.110    0.000    0.189    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-   231014    0.052    0.000    0.337    0.000 <__array_function__ internals>:177(atleast_1d)
-   230999    0.030    0.000    0.030    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-   157323    0.058    0.000    0.058    0.000 {built-in method numpy.array}
-   123557    0.006    0.000    0.006    0.000 {method 'items' of 'dict' objects}
-   110546    0.072    0.000    0.072    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-   110502    0.020    0.000    0.020    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
-   110502    0.065    0.000    0.215    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
+1075496/1075143    0.036    0.000    0.037    0.000 {built-in method builtins.len}
+   813790    0.049    0.000    0.049    0.000 {method 'index' of 'list' objects}
+   813790    0.255    0.000    0.331    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
+   690834    0.056    0.000    0.056    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
+362298/362136    0.161    0.000    0.622    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
+357285/357201    0.017    0.000    0.018    0.000 {built-in method builtins.isinstance}
+   253302    0.011    0.000    0.011    0.000 {method 'append' of 'list' objects}
+   251058    0.034    0.000    0.034    0.000 {built-in method numpy.asanyarray}
[…]

[…]
