+++ src/prog_models/utils/calc_error.py
@@ -19,7 +19,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
     Keyword Args:
         x0 (StateContainer): Current State of the model
         dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
-        stability_tol (double, optional): Configurable parameter.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
             stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
@@ -102,7 +102,7 @@ def RMSE(m, times, inputs, outputs, **kwargs):
     Keyword Args:
         x0 (StateContainer): Current State of the model
         dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
-        stability_tol (double, optional): Configurable parameter.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
             stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
@@ -121,14 +121,15 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
     """Calculate Mean Squared Error (MSE) between simulated and observed
 
     Args:
-        times (list[float]): Array of times for each sample.
-        inputs (list[dict]): Array of input dictionaries where input[x] corresponds to time[x].
-        outputs (list[dict]): Array of output dictionaries where output[x] corresponds to time[x].
+        m (PrognosticsModel): Model to use for comparison
+        times (list[float]): array of times for each sample
+        inputs (list[dict]): array of input dictionaries where input[x] corresponds to time[x]
+        outputs (list[dict]): array of output dictionaries where output[x] corresponds to time[x]
 
     Keyword Args:
         x0 (dict, optional): Initial state.
-        dt (double, optional): Maximum time step.
-        stability_tol (double, optional): Configurable parameter.
+        dt (float, optional): Maximum time step.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
             stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
@@ -138,7 +139,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
             Else, model goes unstable after stability_tol is met, the mean squared error calculated from data up to the instability is returned.
 
     Returns:
-        double: Total error
+        float: Total error
     """
     if isinstance(times[0], Iterable):
         # Calculate error for each
@@ -205,7 +206,7 @@ def MAE(m, times, inputs, outputs, **kwargs):
     Keyword Args:
         x0 (StateContainer): Current State of the model
         dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
-        stability_tol (double, optional): Configurable parameter.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
[…]
