+++ tests/test_base_models.py
@@ -1263,24 +1263,26 @@ def run_tests():
     unittest.main()
     
 def main():
+    import cProfile, pstats
     l = unittest.TestLoader()
     runner = unittest.TextTestRunner()
     print("\n\nTesting Base Models")
+    profiler = cProfile.Profile()
+
+    profiler.enable()
     result = runner.run(l.loadTestsFromTestCase(TestModels)).wasSuccessful()
+    profiler.disable()
+
+    with open("output_time.txt", 'w') as f:
+        p = pstats.Stats(profiler, stream=f)
+        p.sort_stats("time").print_stats()
+
+    with open("output_calls.txt", 'w') as f:
+        p = pstats.Stats(profiler, stream=f)
+        p.sort_stats("calls").print_stats()
 
     if not result:
         raise Exception("Failed test")
 
 if __name__ == '__main__':
-    import cProfile
-    cProfile.run('main()', "output.dat")
-
-    import pstats
-
-    with open("output_time.txt", 'w') as f:
-        p = pstats.Stats("output.dat", stream=f)
-        p.sort_stats("time").print_stats()
-
-    with open("output_calls.txt", 'w') as f:
-        p = pstats.Stats("output.dat", stream=f)
-        p.sort_stats("calls").print_stats()
+    main()
