### 2023-04-18 5831ad5
Remove benchmarking time
