+++ tests/test_data_model.py
@@ -314,27 +314,26 @@ def run_tests():
     unittest.main()
 
 def main():
+    import cProfile, pstats
     l = unittest.TestLoader()
     runner = unittest.TextTestRunner()
-    print("\n\nTesting Data Models")
+    print("\n\nTesting Base Models")
+    profiler = cProfile.Profile()
 
-    with patch('matplotlib.pyplot.show'):
-        result = runner.run(l.loadTestsFromTestCase(TestDataModel)).wasSuccessful()
-    plt.close('all')
-    
+    profiler.enable()
+    result = runner.run(l.loadTestsFromTestCase(TestDataModel)).wasSuccessful()
+    profiler.disable()
+
+    with open("output_time.txt", 'w') as f:
+        p = pstats.Stats(profiler, stream=f)
+        p.sort_stats("time").print_stats()
+
+    with open("output_calls.txt", 'w') as f:
+        p = pstats.Stats(profiler, stream=f)
+        p.sort_stats("calls").print_stats()
+
     if not result:
         raise Exception("Failed test")
 
 if __name__ == '__main__':
-    import cProfile
-    cProfile.run('main()', "output.dat")
-
-    import pstats
-
-    with open("output_time.txt", 'w') as f:
-        p = pstats.Stats("output.dat", stream=f)
-        p.sort_stats("time").print_stats()
-
-    with open("output_calls.txt", 'w') as f:
-        p = pstats.Stats("output.dat", stream=f)
-        p.sort_stats("calls").print_stats()
+    main()
