### 2023-04-18 aa383b7
Fix spelling in print statement
Co-authored-by: Miryam S. <mstraut21@gmail.com>
