### 2023-04-18 8f88f2b
First changes to LQR
