+++ tests/test_estimate_params.py
@@ -85,19 +85,7 @@ class TestEstimateParams(unittest.TestCase):
 
         m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, -8)))
 
-        self.assertEqual(m.parameters['thrower_height'], 4)
-        self.assertEqual(m.parameters['throwing_speed'], 42)
-        self.assertEqual(m.parameters['g'], -20)
-
-        # However, this is a feature that occurs with bounds internally and is not an aspect of our original code.
-        m.parameters['thrower_height'] = 4
-        m.parameters['throwing_speed'] = 24
-        m.parameters['g'] = -20
-        m.estimate_params(data, keys)
-        for key in keys:
-            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
-
-        # Furthermore, notice that these changes are regardless if the lower/upper bounds equals our initially set parameters
+        # Note that with small changes to our bounds, our estimated values will change drastically.
         m.parameters['thrower_height'] = 4
         m.parameters['throwing_speed'] = 24
         m.parameters['g'] = -20
@@ -112,6 +100,14 @@ class TestEstimateParams(unittest.TestCase):
         self.assertNotEqual(m.parameters['throwing_speed'], 42)
         self.assertLess(m.parameters['throwing_speed'], 42)
         self.assertGreaterEqual(m.parameters['throwing_speed'], 20)
+
+        # Testing convergence with the same initial parameters but no given bounds were provided        
+        m.parameters['thrower_height'] = 4
+        m.parameters['throwing_speed'] = 24
+        m.parameters['g'] = -20
+        m.estimate_params(data, keys)
+        for key in keys:
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
 
     def test_estimate_params(self):
@@ -200,7 +196,7 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         m.estimate_params(data, keys, bounds=((0, 4), (20, 37), (-20, 0)))
-        # Not the most accurate way of seeing when these values are not accurate.
+        # Checking each key and seeing if it does not almost equal the original parameters
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 1)
 
@@ -246,15 +242,14 @@ class TestEstimateParams(unittest.TestCase):
         
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        # Dictionary bounds - extra
+        # Dictionary bounds - extra & garbage key
         m.estimate_params(data, keys, bounds={'thrower_height': (0, 4), 'throwing_speed': (20, 42), 'g': (-20, 0), 'dummy': (-50, 0)})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        # Behavior, sets in default values and does not error with correct number of bounds.
-        # User changes a value that exists and should not be changed
+        # Dictionary bounds - extra & not garbage key
         m.estimate_params(data, keys, bounds={'thrower_height': (0, 4), 'throwing_speed': (20, 42), 'g': (-20, 0), 'rho': (-100, 100)})
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
@@ -312,6 +307,8 @@ class TestEstimateParams(unittest.TestCase):
             m.estimate_params()
 
         # Testing with Arrays
+        m.parameters['thrower_height'] = 1.5
+        m.parameters['throwing_speed'] = 25
[…]
