+++ src/prog_models/prognostics_model.py
@@ -13,7 +13,7 @@ from warnings import warn
 
 from prog_models.exceptions import ProgModelInputException, ProgModelTypeError, ProgModelException, ProgModelStateLimitWarning
 from prog_models.sim_result import SimResult, LazySimResult
-from prog_models.utils import ProgressBar
+from prog_models.utils import ProgressBar, calc_error
 from prog_models.utils.containers import DictLikeMatrixWrapper
 from prog_models.utils.parameters import PrognosticsModelParameters
 from prog_models.utils.serialization import CustomEncoder, custom_decoder
@@ -1151,56 +1151,52 @@ class PrognosticsModel(ABC):
         """Calculate Mean Squared Error (MSE) between simulated and observed
 
         Args:
-            times (list[float]): Array of times for each sample.
-            inputs (list[dict]): Array of input dictionaries where input[x] corresponds to time[x].
-            outputs (list[dict]): Array of output dictionaries where output[x] corresponds to time[x].
-
+            times (list[float]): array of times for each sample
+            inputs (list[dict]): array of input dictionaries where input[x] corresponds to time[x]
+            outputs (list[dict]): array of output dictionaries where output[x] corresponds to time[x]
+        
         Keyword Args:
-            x0 (dict, optional): Initial state.
-            dt (float, optional): Maximum time step.
+            method (str, optional): Error method to use. Supported methods include:
+                * MSE (Mean Squared Error)
+                * RMSE (Root Mean Squared Error)
+                * MAX_E (Maximum Error)
+                * MAE (Mean Absolute Error)
+                * MAPE (Mean Absolute Percentage Error)
+            x0 (dict, optional): Initial state
+            dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
+            stability_tol (double, optional): Configurable parameter.
+                Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
+                In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
+                stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
+                before the model goes unstable in order to produce a valid estimate of mean squared error. 
+
+                If the model goes unstable before stability_tol is met, NaN is returned. 
+                Else, model goes unstable after stability_tol is met, the mean squared error calculated from data up to the instability is returned.
 
         Returns:
-            float: Total error
+            float: error
+
+        See Also:
+            :func:`calc_error.MSE`
         """
-        if isinstance(times[0], Iterable):
-            # Calculate error for each
-            error = [self.calc_error(t, i, z, **kwargs) for (t, i, z) in zip(times, inputs, outputs)]
-            return sum(error)/len(error)
+        method = kwargs.get('method', 'MSE')
 
-        x = kwargs.get('x0', self.initialize(inputs[0], outputs[0]))
-        dt = kwargs.get('dt', 1e99)
+        # Call appropriate error calculation method
+        if method.lower() == 'mse':
+            return calc_error.MSE(self, times, inputs, outputs, **kwargs)
+        if method.lower() == 'max_e':
+            return calc_error.MAX_E(self, times, inputs, outputs, **kwargs)
+        if method.lower() == 'rmse':
+            return calc_error.RMSE(self, times, inputs, outputs, **kwargs)
+        if method.lower() == 'mae':
+            return calc_error.MAE(self, times, inputs, outputs, **kwargs)
+        if method.lower() == 'mape':
+            return calc_error.MAPE(self, times, inputs, outputs, **kwargs)
 
[…]

[…]
