### 2023-04-19 c7ce495
Cleaning up old controller code
