+++ src/prog_models/models/uav_model/vehicles/AircraftModels.py
@@ -23,17 +23,6 @@ def build_model(**kwargs):
                   steadystate_input=None)
     params.update(kwargs)
     
-    # # Default control parameters (if not provided)
-    # if params['Q'] is None:
-    #     params['Q'] = np.diag([1000, 1000, 25000,  # x, y, z
-    #                            100.0, 100.0, 100.0,  # phi, theta, psi,
-    #                            1000, 1000, 5000,  # vx, vy, vz,
-    #                            1000, 1000, 1000])  # p, q, r
-    # if params['R'] is None:         params['R']     = np.diag([500, 4000, 4000, 4000])   # T, Mu, Mv, Mw
-    # if params['qi'] is None:        params['qi']    = np.array([100, 100, 1000])  # Integral error weights in position x, y, z
-    # if params['i_lag'] is None:     params['i_lag'] = 100                        # integral lag: how far back to use (in data points, so 1 point = 1 dt) for integral error
-
-
     # Generate UAV model
     # ------------------
     uav = Rotorcraft(model=params['model'], payload=params['payload'])
@@ -44,24 +33,6 @@ def build_model(**kwargs):
               steadystate_input = params['steadystate_input'],     # None assigns deault value (hover condition)
               dt                = params['dt']) # should be small enough to converge (RK4 allows larger dt)
 
-
-    # # Define controller and control strategy
-    # # ---------------------------------------
-    # # States for scheduled controller
-    # # use -2pi / 2pi because the rotation of the UAV may happen in either direction
-    # # to ensure that the minimum rotation necessary is achieved (weird thing). using 720 +1 points to ensure that yaw = 0 has a matrix.
-    # scheduled_states       = np.zeros((uav.dynamics['num_states'], 360*2+1))
-    # scheduled_states[5, :] = np.linspace(start=-2.0*np.pi, stop=2.0*np.pi, num=360*2+1)
-    
-    # # Set controller
-    # uav.set_controller(type_            = 'LQR_I',          # LQR (normal lqr control) or LQR_I (lqr with integral action)
-    #                    strategy         = 'scheduled',      # Compute control gain in real time or pre-compute and schedule it
-    #                    scheduled_states = scheduled_states, # Scheduled states to pre-compute control gains; valid only if control_strategy=='scheduled'
-    #                    scheduled_var    = 'psi',            # Variable that changes in scheduled-control; valid only if control_strategy=='scheduled'
-    #                    Q                = params['Q'],                # Control weights and integral lag
-    #                    R                = params['R'],             
-    #                    qi               = params['qi'],
-    #                    int_lag          = params['i_lag'])    
     return uav
 
 
@@ -76,7 +47,6 @@ class Rotorcraft():
                  air_density=1.225,
                  **kwargs):
         self.model = model
-        # self.controller = None
         self.gravity = gravity
         self.state = None
         self.state_names = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'xdot', 'ydot', 'zdot', 'p', 'q', 'r']
@@ -104,7 +74,6 @@ class Rotorcraft():
 
     def set_dt(self, dt):
         self.dt = dt
-        # self.controller.dt = dt
     
     def build(self, initial_state=None, steadystate_input=None, dt=0.01):
 
[…]
