### 2023-04-21 94af910
updated import with numpy array to handle dictionaries with values that contain numpy arrays.
