### 2023-04-21 a446932
LQR controller changes
