+++ src/prog_models/utils/calc_error.py
@@ -2,6 +2,7 @@
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
 from collections.abc import Iterable
+from fastdtw import *
 from warnings import warn
 import math
 import numpy as np
@@ -52,7 +53,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
     # Checks stability_tol is within bounds
     # Throwing a default after the warning.
     if stability_tol >= 1 or stability_tol < 0:
-        warn(f"configurable cutoff must be some float value in the domain (0, 1]. "
+        warn(f"Configurable cutoff must be some float value in the domain (0, 1]. "
              f"Received {stability_tol}. Resetting value to 0.95")
         stability_tol = 0.95
 
@@ -60,7 +61,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
     t_last = times[0]
     err_max = 0
     z_obs = m.output(x)  # Initialize
-    cutoffThreshold = math.floor(stability_tol * len(times))
+    cutoffThreshold = stability_tol * times[-1]
 
     for t, u, z in zip(times, inputs, outputs):
         while t_last < t:
@@ -72,7 +73,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
                 z_obs = m.output(x)
         if not (None in z_obs.matrix or None in z.matrix):
             if any(np.isnan(z_obs.matrix)):
-                if counter < cutoffThreshold:
+                if t <= cutoffThreshold:
                     raise ValueError(f"Model unstable- NAN reached in simulation (t={t}) before cutoff threshold. "
                                      f"Cutoff threshold is {cutoffThreshold}, or roughly {stability_tol * 100}% of the data")                 
                 else: 
@@ -162,7 +163,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
     # Checks stability_tol is within bounds
     # Throwing a default after the warning.
     if stability_tol >= 1 or stability_tol < 0:
-        warn(f"configurable cutoff must be some float value in the domain (0, 1]. "  
+        warn(f"Configurable cutoff must be some float value in the domain (0, 1]. "  
              f"Received {stability_tol}. Resetting value to 0.95")
         stability_tol = 0.95
 
@@ -170,7 +171,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
     t_last = times[0]
     err_total = 0
     z_obs = m.output(x)
-    cutoffThreshold = math.floor(stability_tol * len(times))
+    cutoffThreshold = stability_tol * times[-1]
 
     for t, u, z in zip(times, inputs, outputs):
         while t_last < t:
@@ -182,7 +183,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
                 z_obs = m.output(x)
         if not (None in z_obs.matrix or None in z.matrix):
             if any (np.isnan(z_obs.matrix)):
-                if counter < cutoffThreshold:
+                if t <= cutoffThreshold:
                     raise ValueError(f"Model unstable- NAN reached in simulation (t={t}) before cutoff threshold. "
                                      f"Cutoff threshold is {cutoffThreshold}, or roughly {stability_tol * 100}% of the data")     
                 else:
@@ -190,7 +191,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
                     break
             err_total += np.sum(np.square(z.matrix - z_obs.matrix), where= ~np.isnan(z.matrix))
             counter += 1
-
+    
     return err_total/counter
 
 def MAE(m, times, inputs, outputs, **kwargs):
@@ -239,7 +240,7 @@ def MAE(m, times, inputs, outputs, **kwargs):
[…]
