+++ tests/test_estimate_params.py
@@ -15,7 +15,6 @@ class TestEstimateParams(unittest.TestCase):
         """
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
-        data = [(results.times, results.inputs, results.outputs)]
         gt = m.parameters.copy()
 
         self.assertEqual(m.parameters, gt)
@@ -24,14 +23,14 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
         keys = ['thrower_height', 'throwing_speed']
-        m.estimate_params(data, keys)
+        m.estimate_params(times=results.times, inputs=results.inputs, outputs=results.outputs, keys=keys)
         for key in keys: 
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         
         # Incorrectly set parameters, but provide a bounds for what said values should be.
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
+        m.estimate_params(times=results.times, inputs=results.inputs, outputs=results.outputs, keys=keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 2)
         # Checking to see if our estimated parameters are within the bounds.
@@ -42,7 +41,7 @@ class TestEstimateParams(unittest.TestCase):
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
-        m.estimate_params(data, keys, bounds=((0, 4), (40, 40)))
+        m.estimate_params(times=results.times, inputs=results.inputs, outputs=results.outputs, keys=keys, bounds=((0, 4), (40, 40)))
         for key in keys:
             self.assertAlmostEqual(m.parameters[key], gt[key], 1)
         self.assertLessEqual(m.parameters['thrower_height'], 4)
@@ -52,7 +51,7 @@ class TestEstimateParams(unittest.TestCase):
         # Testing initial parameters that are not within the defined bounds; estimated parameters are not a good fit
         m.parameters['thrower_height'] = 10
         m.parameters['throwing_speed'] = -10
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42)))
+        m.estimate_params(times=results.times, inputs=results.inputs, outputs=results.outputs, keys=keys, bounds=((0, 4), (20, 42)))
         for key in keys:
             self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
         # These returned values are still within the bounds, they are just not close to the original value at all.
@@ -60,32 +59,12 @@ class TestEstimateParams(unittest.TestCase):
         self.assertAlmostEqual(m.parameters['thrower_height'], 4)
         self.assertAlmostEqual(m.parameters['throwing_speed'], 42)
 
-        # Showcasing results of having a local min/max
-        # Even though all our bounds include the original model params values,
-        # our estimate_params returns lower/upper bound values.
-        m.parameters['thrower_height'] = 4
-        m.parameters['throwing_speed'] = 24
-        m.parameters['g'] = -20
-        keys = ['thrower_height', 'throwing_speed', 'g']
-        m.estimate_params(data, keys, bounds=((0, 4), (20, 42), (-20, -8)))
-        for key in keys:
-            self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
-
-        # Note that with small changes to our bounds, our estimated values will change drastically.
[…]
