+++ src/prog_models/utils/calc_error.py
@@ -24,7 +24,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
     Keyword Args:
         x0 (StateContainer): Current State of the model
         dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
-        stability_tol (double, optional): Configurable parameter.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
             stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
@@ -57,10 +57,11 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
     # Checks stability_tol is within bounds
     # Throwing a default after the warning.
     if stability_tol >= 1 or stability_tol < 0:
-        warn(f"configurable cutoff must be some float value in the domain (0, 1]. Received {stability_tol}. Resetting value to 0.95")
+        warn(f"configurable cutoff must be some float value in the domain (0, 1]. "
+             f"Received {stability_tol}. Resetting value to 0.95")
         stability_tol = 0.95
 
-    counter = 0
+    counter = 0 
     t_last = times[0]
     err_max = 0
     z_obs = m.output(x)  # Initialize
@@ -82,7 +83,8 @@ def MAX_E(m, times, inputs, outputs, **kwargs):
             # This is true for any window-based model
             if any(np.isnan(z_obs.matrix)):
                 if counter < cutoffThreshold:
-                    raise ValueError(f"Model unstable- NAN reached in simulation (t={t}) before cutoff threshold. Cutoff threshold is {cutoffThreshold}, or roughly {stability_tol * 100}% of the data")                
+                    raise ValueError(f"Model unstable- NAN reached in simulation (t={t}) before cutoff threshold. "
+                                     f"Cutoff threshold is {cutoffThreshold}, or roughly {stability_tol * 100}% of the data")                 
                 else: 
                     warn(f"Model unstable- NaN reached in simulation (t={t})")
                     break
@@ -110,7 +112,7 @@ def RMSE(m, times, inputs, outputs, **kwargs):
     Keyword Args:
         x0 (StateContainer): Current State of the model
         dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
-        stability_tol (double, optional): Configurable parameter.
+        stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
             stability_tol represents the fraction of the provided argument `times` that are required to be met in simulation, 
@@ -125,18 +127,19 @@ def RMSE(m, times, inputs, outputs, **kwargs):
     return np.sqrt(MSE(m, times, inputs, outputs, **kwargs))
 
 
-def MSE(self, times, inputs, outputs, **kwargs) -> float:
+def MSE(m, times, inputs, outputs, **kwargs) -> float:
     """Calculate Mean Squared Error (MSE) between simulated and observed
 
     Args:
-        times (list[float]): Array of times for each sample.
-        inputs (list[dict]): Array of input dictionaries where input[x] corresponds to time[x].
-        outputs (list[dict]): Array of output dictionaries where output[x] corresponds to time[x].
+        m (PrognosticsModel): Model to use for comparison
+        times (list[float]): array of times for each sample
+        inputs (list[dict]): array of input dictionaries where input[x] corresponds to time[x]
[…]
