### 2023-04-27 7182981
Update tests/test_estimate_params.py
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
