### 2023-04-27 cdd6258
Apply suggestions from code review
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
