+++ src/prog_models/utils/containers.py
@@ -1,10 +1,8 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
-from numpy import array
-import numpy as np
+from numpy import float64, matrix, ndarray, array, newaxis, nan, delete, atleast_1d, vstack
 import pandas as pd
-
 from prog_models.exceptions import ProgModelTypeError
 from typing import Union
 
@@ -18,31 +16,31 @@ class DictLikeMatrixWrapper():
         data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
 
-    def __init__(self, keys: list, data: Union[dict, np.array]):
+    def __init__(self, keys: list, data: Union[dict, array]):
         """
         Initializes the container
         """
         if not isinstance(keys, list):
             keys = list(keys)  # creates list with keys
         self._keys = keys.copy()
-        if isinstance(data, np.matrix):
-            self.data = pd.DataFrame(np.array(data, dtype=np.float64), self._keys, dtype=np.float64)
-            self.matrix = self.data.to_numpy(dtype=np.float64)
-        elif isinstance(data, np.ndarray):
+        if isinstance(data, matrix):
+            self.data = pd.DataFrame(array(data, dtype=float64), self._keys, dtype=float64)
+            self.matrix = self.data.to_numpy(dtype=float64)
+        elif isinstance(data, ndarray):
             if data.ndim == 1:
-                data = data[np.newaxis].T
+                data = data[newaxis].T
                 self.data = pd.DataFrame(data, self._keys)
             self.data = pd.DataFrame(data, self._keys).T
             self.matrix = data
         elif isinstance(data, (dict, DictLikeMatrixWrapper)):
-            if data and not isinstance(list(data.values())[0], np.ndarray):  # len(self.matrix[0]) == 1:
+            if data and not isinstance(list(data.values())[0], ndarray):  # len(self.matrix[0]) == 1:
                 if isinstance(data, DictLikeMatrixWrapper):
                     data = dict(data.copy())
-                self.data = pd.DataFrame(data, columns=self._keys, index=[0], dtype=np.float64).astype(object).replace(
-                    np.nan, None)
+                self.data = pd.DataFrame(data, columns=self._keys, index=[0], dtype=float64).astype(object).replace(
+                    nan, None)
             else:
                 self.data = pd.DataFrame(data, columns=self._keys)
-            self.matrix = self.data.to_numpy(dtype=np.float64).T if len(data) > 0 else np.array([])
+            self.matrix = self.data.to_numpy(dtype=float64).T if len(data) > 0 else array([])
         else:
             raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
@@ -50,16 +48,16 @@ class DictLikeMatrixWrapper():
         """
         reduce is overridden for pickles
         """
-        return (DictLikeMatrixWrapper, (self._keys, self.matrix))
+        return DictLikeMatrixWrapper, (self._keys, self.matrix)
 
-    def getMatrix(self) -> np.ndarray:
+    def getmatrix(self) -> ndarray:
         """
         creates a numpy array corresponding to the pd.DataFrame data
 
         Returns: np.ndarray
         """
-        matrix = self.data.to_numpy().T
-        return matrix
+        matrix_np = self.data.to_numpy().T
+        return matrix_np
 
     def __getitem__(self, key: str) -> int:
         """
@@ -77,13 +75,13 @@ class DictLikeMatrixWrapper():
         sets a row at the key given
         """
         index = self._keys.index(key)  # the int value index for the key given
-        self.matrix[index] = np.atleast_1d(value)
[…]
