+++ tests/test_calc_error.py
@@ -44,24 +44,21 @@ class TestCalcError(unittest.TestCase):
         m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
 
         # Initializing parameters to very erroneous values       
-        m.parameters['qMax'] = 4000
+        m.parameters['qMax'] = 8000
         keys = ['qMax']
 
-        # Before running estimate_params
+        # Before running estimate_paramss
         with self.assertRaises(ValueError):
             m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)
 
-        m.estimate_params([(simulated_results.times, simulated_results.inputs, simulated_results.outputs)], keys, dt=0.5)
-
+        m.estimate_params([(simulated_results.times, simulated_results.inputs, simulated_results.outputs)], keys, dt=1)
         # After running estimate_params. Note that this would not change the outcome of the result
         with self.assertRaises(ValueError):
             m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=1)    
 
-        # Running through various dt values that could work
-        # Expecting an error in cutoff threshold.
 
         # By changing the 
-        for i in np.arange(0, 1, 0.1):
+        for i in np.arange(0.1, 1, 0.1):
             with self.assertRaises(ValueError):
                 m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt=i)
 
@@ -71,17 +68,12 @@ class TestCalcError(unittest.TestCase):
  
         # Creating duplicate model
         m1 = BatteryElectroChemEOD()
-
-        orig_params = m1.parameters.copy()
-
         # Much bigger parameter initialization
         m.parameters['kp'] = m1.parameters['kp'] = 10000
         m.parameters['kn'] = m1.parameters['kn'] = 1000 
         m.parameters['qpMax'] = m1.parameters['qpMax'] = 4500
         m.parameters['qMax'] = m1.parameters['qMax'] = 9000
         keys = ['kp', 'kn', 'qpMax','qMax']
-
-        change_params = m1.parameters.copy()
 
         simulated_results = m.simulate_to(2000, future_loading, **options)
         m1_sim_results = m1.simulate_to(2000, future_loading, **options)
@@ -94,14 +86,8 @@ class TestCalcError(unittest.TestCase):
 
         # Calling estimate_params does not change any of the parameters here because we are always accounting for exceptions...
 
-        m.estimate_params(data, keys, method='Powell')
-        m1.estimate_params(data_m1, keys, method='CG')
-
-        updated_params = m1.parameters.copy()
-
-        # Checking to make sure estimate_params actually changed values away from the original and to something else
-        self.assertEqual(change_params, updated_params)
-        self.assertNotEqual(orig_params, updated_params)
+        m.estimate_params(times = simulated_results.times, inputs = simulated_results.inputs, outputs = simulated_results.outputs, keys = keys, method='Powell')
+        m1.estimate_params(times = m1_sim_results.times, inputs = m1_sim_results.inputs, outputs = m1_sim_results.outputs, keys = keys, method='CG')
         
         with self.assertRaises(ValueError) as cm:
             m.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, dt = 1)
@@ -122,14 +108,6 @@ class TestCalcError(unittest.TestCase):
             str(cm.warning)
         )
 
-
-        # with self.assertWarns(UserWarning) as cm:
-        #     m1.calc_error(simulated_results.times, simulated_results.inputs, simulated_results.outputs, 
[…]
