+++ src/prog_models/utils/containers_df.py
@@ -1,202 +1,0 @@
-# Copyright © 2021 United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration.  All Rights Reserved.
-
-import numpy as np
-from typing import Union
-import pandas as pd
-
-from prog_models.exceptions import ProgModelTypeError
-
-
-class DictLikeMatrixWrapper():
-    """
-    A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-
-    Arguments:
-        keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
-        data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
-    """
-
-    def __init__(self, keys: list, data: Union[dict, np.array, pd.Series]):
-        """
-        Initializes the container
-        """
-        if not isinstance(keys, list):
-            keys = list(keys)  # creates list with keys
-        self._keys = keys.copy()
-        if isinstance(data, np.matrix):
-            self.data = pd.DataFrame(np.array(data, dtype=np.float64), self._keys)
-            self.matrix = self.data.to_numpy(dtype=np.float64)
-        elif isinstance(data, np.ndarray):  # data is a multidimensional array, in column vector form
-            self.data = pd.DataFrame(data, self._keys)
-            self.matrix = self.data.to_numpy()
-        elif isinstance(data, (dict, DictLikeMatrixWrapper)):  # data is not in column vector form
-            self._keys = list(data.keys())
-            for key in keys:
-                if key not in data:
-                    self._keys.append(key)
-            arr_num = list(data.values())[0]
-            if isinstance(arr_num, int):
-                self.data = pd.DataFrame(data, columns=self._keys, index=[0]).astype(object).replace(np.nan, None)
-            else:
-                self.data = pd.DataFrame(data, columns=self._keys).astype(object).replace(np.nan, None)
-            self.matrix = self.data.T.to_numpy(dtype=np.float64)
-        else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
-
-    def __reduce__(self):
-        """
-        reduce is overridden for pickles
-        """
-        keys = self.data.index.to_list()
-        matrix = self.data.to_numpy()
-        return (DictLikeMatrixWrapper, (keys, matrix))
-
-    def __getitem__(self, key: str) -> int:
-        """
-        get all values associated with a key, ex: all values of 'i'
-        """
-        row = self.data.loc[key].to_list()  # creates list from a row of the DataFrame data
-        if len(self.data.loc[key]) == 1:  # list contains 1 value, returns that value (non-vectorized)
-            return self.data.loc[key, 0]
-        return row  # returns entire row/list (vectorized case)
-
-    def __setitem__(self, key: str, value: int) -> None:
-        """
-        sets a row at the key given
-        """
-        self.data.loc[key] = np.atleast_1d(value)  # using the key to find the DataFrame location
-        self.matrix = self.data.to_numpy()
-
-    def __delitem__(self, key: str) -> None:
-        """
-        removes row associated with key
-        """
-        self.data = self.data.drop(index=[key])
-        self.matrix = np.delete(self.matrix, self._keys.index(key), axis=0)
-        self._keys.remove(key)
-
-    def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
-        """
-        add 'other' matrix to the existing matrix
-        """
-        df_summed = self.data.add(other.data)  # the values in self and other summed in new series
-        key_list = self.data.index.to_list()
[…]
