### 2023-05-02 9380d6c
updating sim result, housekeeping - test_sim_result_df.py
