+++ tests/test_sim_result_df.py
@@ -1,0 +1,702 @@
+# Copyright © 2021 United States Government as represented by the Administrator of the National Aeronautics and Space Administration.  All Rights Reserved.
+
+from io import StringIO
+import numpy as np
+import pickle
+import sys
+import unittest
+import pandas as pd
+
+from prog_models.models import BatteryElectroChemEOD
+from prog_models.sim_result import SimResult, LazySimResult
+from prog_models.utils.containers import DictLikeMatrixWrapper
+
+
+class TestSimResult(unittest.TestCase):
+    """def setUp(self):
+        set stdout (so it won't print)
+        sys.stdout = StringIO()
+
+    def tearDown(self):
+        sys.stdout = sys.__stdout__"""
+
+    def test_sim_result(self):
+        # Variables
+        time = list(range(5))
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
+        frame = pd.DataFrame(state)
+        frame.insert(0, "time", time)
+        result = SimResult(time, state)
+        # Checks values from SimResult object and static variables
+        self.assertListEqual(list(result), state)
+        self.assertListEqual(result.times, time)
+        for i in range(5):
+            self.assertEqual(result.time(i), time[i])
+            self.assertEqual(result[i], state[i])
+        self.assertTrue(frame.equals(result.frame))
+        try:
+            tmp = result[5]
+            self.fail("Should be out of range error")
+        except IndexError:
+            pass
+        try:
+            tmp = result.time(5)
+            self.fail("Should be out of range error")
+        except IndexError:
+            pass
+
+    def test_pickle(self):
+        # Variables
+        time = list(range(5))   # list of int, 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
+        result = SimResult(time, state)
+        pickle.dump(result, open('model_test.pkl', 'wb'))
+        result2 = pickle.load(open('model_test.pkl', 'rb'))
+        self.assertEqual(result, result2)
+
+    def test_extend(self):
+        # Variables
+        time = list(range(5))  # list of int from 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(5)]
+        result = SimResult(time, state)
+        time2 = list(range(10))  # list of int from 0 to 9
+        state2 = [{'a': i * 5, 'b': i * 5} for i in range(10)]
+        result2 = SimResult(time2, state2)
+        time_extended = time + time2
+        state_extended = state + state2
+
+        self.assertEqual(result.times, time)
+        self.assertEqual(result2.times, time2)
+        self.assertEqual(result.data, state)  # Assert data is correct before extending
+        self.assertEqual(result2.data, state2)
+
+        result.extend(result2)  # Extend result with result2
+        self.assertEqual(result.times, time_extended)
+        self.assertEqual(result.data, state_extended)
+
+        self.assertRaises(ValueError, result.extend, 0)  # Passing non-LazySimResult types to extend method
+        self.assertRaises(ValueError, result.extend, [0, 1])
+        self.assertRaises(ValueError, result.extend, {})
+        self.assertRaises(ValueError, result.extend, set())
+        self.assertRaises(ValueError, result.extend, 1.5)
+
+    def test_extended_by_lazy(self):
+        NUM_ELEMENTS = 5
+        time = list(range(NUM_ELEMENTS))
+        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(NUM_ELEMENTS)]
[…]
