+++ tests/test_calc_error.py
@@ -14,7 +14,7 @@ class TestCalcError(unittest.TestCase):
 
     Validating that values are correctly being passed into the new calc_error calls and that we are receiving expected results!
     """
-
+    @unittest.skip
     def test_calc_error(self):
         # Note, lowering time steps or increasing simulate threshold may cause this model to not run (takes too long)
         m = BatteryElectroChemEOD()
@@ -204,7 +204,7 @@ class TestCalcError(unittest.TestCase):
         with self.assertRaises(TypeError) as cm:
             m.calc_error(results.times, results.inputs, results.outputs, x0 = 1)
         self.assertEqual(
-            "Keyword argument 'x0' must be initialized to a StateContainer, not a int.",
+            "Keyword argument 'x0' must be initialized to a Dict or StateContainer, not a int.",
             str(cm.exception)
         )
 
@@ -249,6 +249,7 @@ class TestCalcError(unittest.TestCase):
         
         m.calc_error(times, inputs, outputs)
 
+        # Will be testing length errors, will also be providing a 'runs' feedback whenever there are multiple give?
         incorrectTimes = [[0, 1, 2, 4, 5, 6, 7, 8, 9]]
 
         with self.assertRaises(ValueError) as cm:
@@ -266,6 +267,56 @@ class TestCalcError(unittest.TestCase):
             "Times, inputs, and outputs must all be the same length. Current lengths are: times = 8, inputs = 9, outputs = 9",
             str(cm.exception)
         )
+
+    def test_type_errors(self):
+        m = ThrownObject()
+        results = m.simulate_to_threshold(save_freq=0.5)
+        gt = m.parameters.copy()
+
+        with self.assertRaises(TypeError) as cm:
+            m.calc_error()
+        self.assertEqual(
+            "PrognosticsModel.calc_error() missing 3 required positional arguments: 'times', 'inputs', and 'outputs'",
+            str(cm.exception)
+        )
+
+        # With Strings
+        with self.assertRaises(TypeError):
+            m.calc_error("1, 2, 3", "1, 2, 3", "1, 2, 3")
+
+        # With a list in a string
+        with self.assertRaises(TypeError):
+            m.calc_error(['1', '2', '3'], ['1', '2', '3'], ['1', '2', '3'])
+        
+        # Passing in bool values
+        with self.assertRaises(TypeError) as cm:
+            m.calc_error([False, False, True], [False, False, True], [False, False, True])
+        self.assertEqual(
+            "Data must be a dictionary or numpy array, not <class 'bool'>",
+            str(cm.exception)
+        )
+
+        # Passing in tuples where, where inputs and outputs do not contain dicts or StateContainers, rather contain ints
+        with self.assertRaises(TypeError) as cm:
+            m.calc_error((1, 2, 3), (2, 3, 4), (3, 4, 5))
+        self.assertEqual(
+            "Data must be a dictionary or numpy array, not <class 'int'>",
+            str(cm.exception)
+        )
+
+        try:
+            m.calc_error((1, 2, 3), ({'1': 1}, {'2': 2}, {'3': 3}), ({'1': 1}, {'2': 2}, {'3': 3}))
+        except:
+            self.fail("Test should have passed")
+        
+        
+        # m.calc_error(np.array(results.times), np.array(results.inputs), np.array(results.outputs))
+
+        # try:
+        #     m.calc_error(np.ndarray(results.times), np.ndarray(results.inputs), np.ndarray(results.outputs))
+        # except:
+        #     self.fail("Test should be passing")
+
 
         
     def test_RMSE(self):
