+++ tests/test_sim_result_df.py
@@ -31,7 +31,7 @@ class TestSimResult(unittest.TestCase):
         self.assertListEqual(list(result), state)
         self.assertListEqual(result.times, time)
         for i in range(5):
-            self.assertEqual(result.time(i), time[i])
+            self.assertEqual(result.times[i], time[i])
             self.assertEqual(result[i], state[i])
         self.assertTrue(frame.equals(result.frame))
         try:
@@ -40,14 +40,14 @@ class TestSimResult(unittest.TestCase):
         except IndexError:
             pass
         try:
-            tmp = result.time(5)
+            tmp = result.times[5]
             self.fail("Should be out of range error")
         except IndexError:
             pass
 
     def test_pickle(self):
         # Variables
-        time = list(range(5))   # list of int, 0 to 4
+        time = list(range(5))  # list of int, 0 to 4
         state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
         result = SimResult(time, state)
         pickle.dump(result, open('model_test.pkl', 'wb'))
@@ -81,41 +81,34 @@ class TestSimResult(unittest.TestCase):
         self.assertRaises(ValueError, result.extend, 1.5)
 
     def test_extended_by_lazy(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))  # list of int, 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(5)]
         result = SimResult(time, state)  # Creating one SimResult object
 
         def f(x):
             return {k: v * 2 for k, v in x.items()}
 
-        NUM_ELEMENTS = 10
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
-        result2 = LazySimResult(f, time, state)  # Creating one LazySimResult object
-
-        self.assertEqual(result.times, [0, 1, 2, 3, 4])
-        self.assertEqual(result2.times, [0, 1, 2, 3, 4, 5, 6, 7, 8, 9])
-        self.assertEqual(result.data,
-                         [{'a': 0.0, 'b': 0.0}, {'a': 2.5, 'b': 2.5}, {'a': 5.0, 'b': 5.0}, {'a': 7.5, 'b': 7.5},
-                          {'a': 10.0, 'b': 10.0}])  # Assert data is correct before extending
-        self.assertEqual(result2.data, [{'a': 0, 'b': 0}, {'a': 10, 'b': 10}, {'a': 20, 'b': 20}, {'a': 30, 'b': 30},
-                                        {'a': 40, 'b': 40}, {'a': 50, 'b': 50}, {'a': 60, 'b': 60}, {'a': 70, 'b': 70},
-                                        {'a': 80, 'b': 80}, {'a': 90, 'b': 90}])
+        time2 = list(range(10))  # list of int, 0 to 9
+        state2 = [{'a': i * 5, 'b': i * 5} for i in range(10)]
+        data2 = [{'a': i * 10, 'b': i * 10} for i in range(10)]
+        result2 = LazySimResult(f, time2, state2)  # Creating one LazySimResult object
+        # confirming the data in result and result2 are correct
+        self.assertEqual(result.times, time)
+        self.assertEqual(result2.times, time2)
+        self.assertEqual(result.data, state)  # Assert data is correct before extending
+        self.assertEqual(result2.data, data2)
         result.extend(result2)  # Extend result with result2
[…]
