+++ src/prog_models/sim_result.py
@@ -54,8 +54,8 @@ class SimResult(UserList):
         Returns:
             bool: If the two SimResults are equal
         """
-        time_check = self.times == other.times
-        data_check = self.data == other.data
+        time_check = np.array_equal(self.times, other.times)
+        data_check = np.array_equal(self.data, other.data)
         frame_check = self.frame.equals(other.frame)
         return time_check and data_check and frame_check
 
@@ -95,8 +95,8 @@ class SimResult(UserList):
         Returns:
             dict: Element Removed
         """
-        if index == -1:
-            index_df = len(self.frame.index)-1
+        if index is -1:
+            index_df = len(self.frame.index) - 1
         else:
             index_df = index
         self.times.pop(index)
@@ -116,18 +116,15 @@ class SimResult(UserList):
         if sum([i is None for i in (d, t)]) != 1:
             raise ValueError("ValueError: Only one named argument (d, t) can be specified.")
 
-        if (t is not None):
+        if t is not None:
             # removes the index of the timestamp meant for removal
-            num = self.frame[self.frame['time'] == t].index[0]
-            if num == self.times.index(t):
-                self.pop(self.times.index(t))
+            index = self.times.index(t)
+            self.pop(index)
         else:
-            temp_df = self.frame.drop(['time'], axis=1)
-            # finds index of dict meant to be removed
+            # finds index of dictionary meant to be removed
             if isinstance(d, dict) and len(d) > 0:
-                d_dict = {key: [val] for key, val in d.items()}
-                num = temp_df[temp_df.isin(d_dict).iloc[:, 0] == True].index[0]
-                self.pop(num)
+                index = self.data.index(d)
+                self.pop(index)
             else:
                 raise ValueError
 
@@ -146,10 +143,10 @@ class SimResult(UserList):
         Returns:
             float: Time for which the data point at index `index` corresponds
         """
-        #return self.frame['time'].iat[index]
+        # return self.frame['time'].iat[index]
         return self.times[index]
 
-    def to_numpy(self, keys=None) -> np.ndarray:
+    def to_numpy(self, keys: list = None) -> np.ndarray:
         """
         Convert from simresult to numpy array
         Args:
@@ -157,15 +154,12 @@ class SimResult(UserList):
         Returns:
             np.ndarray: numpy array representing simresult
         """
-        if len(self.data) == 0:
+        if len(self.data) is 0:
             return np.array([[]], dtype=np.float64)
-        if len(self.data[0]) == 0:
-            return np.array([[] for _ in self.data], dtype=np.float64)
-        if isinstance(self.data[0], DictLikeMatrixWrapper) and keys is None:
-            return np.array([u_i.matrix[:, 0] for u_i in self.data], dtype=np.float64)
-        if keys is None:
-            keys = self.data[0].keys()
-        return np.array([[u_i[key] for key in keys] for u_i in self.data], dtype=np.float64)
+        if keys is not None:
+            with_keys_numpy = self.frame.drop(['time'], axis=1)[keys].to_numpy(dtype=np.float64)
+            return with_keys_numpy
+        return self.frame.drop(['time'], axis=1).to_numpy(dtype=np.float64)
 
     def plot(self, **kwargs) -> figure:
         """
@@ -203,7 +197,6 @@ class SimResult(UserList):
         https://www.sciencedirect.com/science/article/pii/S0004370222000078
 
         Args:
-            None
 
         Returns:
             float: Value between [0, 1] indicating monotonicity of a given event for the Prediction.
@@ -241,7 +234,7 @@ class LazySimResult(SimResult):  # lgtm [py/missing-equals]
     Used to store the result of a simulation, which is only calculated on first request
     """
 
[…]
