+++ examples/sim.py
@@ -59,7 +59,7 @@ def run_example():
     # Alternately, you can set a max step size and allow step size to be adjusted automatically
     options['dt'] = ('auto', 2)  # set step size automatically, with a max of 2 seconds
     options['save_freq'] = 201  # Save every 201 seconds
-    options['save_pts'] = [250, 772, 1023]  # Special points we sould like to see reported
+    options['save_pts'] = [250, 772, 1023]  # Special points we should like to see reported
     simulated_results = batt.simulate_to_threshold(future_loading, **options)
     # Note that even though the step size is 2, the odd points in the save frequency are met perfectly, dt is adjusted automatically to capture the save points
 

[…]
