+++ src/prog_models/utils/calc_error.py
@@ -6,12 +6,13 @@ This file contains functions for calculating error given a model and some data (
 """
 
 from collections.abc import Iterable
+from typing import List
 from warnings import warn
 import math
 import numpy as np
 
 
-def MAX_E(m, times, inputs, outputs, **kwargs) -> float:
+def MAX_E(m, times: List[float], inputs: List[dict], outputs: List[dict], **kwargs) -> float:
     """
     Calculate the Maximum Error between model behavior and some collected data.
 
@@ -23,7 +24,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs) -> float:
 
     Keyword Args:
         x0 (StateContainer): Current State of the model
-        dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
+        dt (float, optional): Maximum time step in simulation. Time step used in simulation is lower of dt and time between samples. Defaults to use time between samples.
         stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
@@ -42,7 +43,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs) -> float:
         return max(error)
 
     x = kwargs.get('x0', m.initialize(inputs[0], outputs[0]))
-    dt = kwargs.get('dt', 1)
+    dt = kwargs.get('dt', 1e99)
     stability_tol = kwargs.get('stability_tol', 0.95)
 
     if not isinstance(x, m.StateContainer):
@@ -99,7 +100,7 @@ def MAX_E(m, times, inputs, outputs, **kwargs) -> float:
     return err_max
 
 
-def RMSE(m, times, inputs, outputs, **kwargs) -> float:
+def RMSE(m, times: List[float], inputs: List[dict], outputs: List[dict], **kwargs) -> float:
     """
     Calculate the Root Mean Squared Error between model behavior and some collected data.
 
@@ -111,7 +112,7 @@ def RMSE(m, times, inputs, outputs, **kwargs) -> float:
 
     Keyword Args:
         x0 (StateContainer): Current State of the model
-        dt (float, optional): Minimum time step in simulation. Defaults to 1e99.
+        dt (float, optional): Maximum time step in simulation. Time step used in simulation is lower of dt and time between samples. Defaults to use time between samples.
         stability_tol (float, optional): Configurable parameter.
             Configurable cutoff value, between 0 and 1, that determines the fraction of the data points for which the model must be stable.
             In some cases, a prognostics model will become unstable under certain conditions, after which point the model can no longer represent behavior. 
@@ -127,7 +128,7 @@ def RMSE(m, times, inputs, outputs, **kwargs) -> float:
     return np.sqrt(MSE(m, times, inputs, outputs, **kwargs))
 
 
-def MSE(m, times, inputs, outputs, **kwargs) -> float:
+def MSE(m, times: List[float], inputs: List[dict], outputs: List[dict], **kwargs) -> float:
     """Calculate Mean Squared Error (MSE) between simulated and observed
 
     Args:
@@ -138,7 +139,7 @@ def MSE(m, times, inputs, outputs, **kwargs) -> float:
 
     Keyword Args:
         x0 (dict, optional): Initial state.
-        dt (float, optional): Maximum time step.
+        dt (float, optional): Maximum time step in simulation. Time step used in simulation is lower of dt and time between samples. Defaults to use time between samples.
         stability_tol (float, optional): Configurable parameter.
[…]
