### 2023-05-05 59ed47b
Type Hints to calc_error and estimate_params
