### 2023-05-06 3549253
Further development of UAV model
LQR and LQRI controller work, adding to example, beginning tests, cleaning up code/moving around parameters, etc.
