+++ src/prog_models/models/uav_model/uav_model.py
@@ -141,7 +141,7 @@ class UAVGen(PrognosticsModel):
     -------------
      .. [0] M. Corbetta et al., "Real-time UAV trajectory prediction for safely monitoring in low-altitude airspace," AIAA Aviation 2019 Forum,  2019. https://arc.aiaa.org/doi/pdf/10.2514/6.2019-3514
     """
-    events = [] # ['TrajectoryComplete']
+    events = ['TrajectoryComplete']
     inputs = ['T','mx','my','mz']
     states = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r','t']
     outputs = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r']
@@ -152,19 +152,16 @@ class UAVGen(PrognosticsModel):
         # Simulation parameters:
         'dt': 0.1, 
         'gravity': 9.81,
-        'cruise_speed': None, 
-        'ascent_speed': None, 
-        'descent_speed': None,  
-        'landing_speed': None, 
-        'hovering_time': 0.0,
-        'takeoff_time': 0.0, 
-        'landing_time': 0.0, 
         'final_time_buffer_sec': 30, 
         'final_space_buffer_m': 2, 
 
         # Vehicle parameters:
         'vehicle_model': 'tarot18', 
         'vehicle_payload': 0.0,
+        'vehicle_max_speed': 15.0, #### These are needed for traj_gen, are these okay defaults?
+        'vehicle_max_roll': 0.7853981633974483, ##### fix this 
+        'vehicle_max_pitch': 0.7853981633974483, #### fix this 
+        'ref_traj': None
     }
 
     def __init__(self, **kwargs):
@@ -278,6 +275,13 @@ class UAVGen(PrognosticsModel):
         return self.StateContainer(np.array([np.atleast_1d(item) for item in dxdt]))
     
     def event_state(self, x : dict) -> dict:
+
+        # Based on time 
+        return {
+                'TrajectoryComplete': x['t']/self.parameters['ref_traj']['t'][-1]
+        }
+
+        # Based on reference trajectory
         # # Extract next waypoint information 
         # num_wypts = len(self.parameters['waypoints']['waypoints_time']) - 1 # Don't include initial waypoint
         # index_next = self.parameters['waypoints']['next_waypoint']
@@ -334,7 +338,6 @@ class UAVGen(PrognosticsModel):
         #     return {
         #             'TrajectoryComplete': (num_wypts - index_next)/num_wypts
         #         }
-        pass
  
     def output(self, x : dict):
         # Output is the same as the state vector, without time 
@@ -342,31 +345,32 @@ class UAVGen(PrognosticsModel):
 
 
     def threshold_met(self, x : dict) -> dict:
-        # t_lower_bound = self.parameters['waypoints']['waypoints_time'][-1] - (self.parameters['waypoints']['waypoints_time'][-1] - self.parameters['waypoints']['waypoints_time'][-2])/2
-        # t_upper_bound = self.parameters['waypoints']['waypoints_time'][-1] + self.parameters['final_time_buffer_sec']
-        # if x['t'] < t_lower_bound:
-        #     # Trajectory hasn't reached final ETA
-        #     return {
-        #         'TrajectoryComplete': False
-        #     }
-        # elif t_lower_bound <= x['t'] <= t_upper_bound:
-        #     # Trajectory is within bounds of final ETA
[…]
