+++ src/prog_models/loading_fcns/controllers.py
@@ -68,7 +68,6 @@ def lqr_fn(A, B, Q, R):
 # # ==============================================================
 class LQR():
     """ Linear Quadratic Regulator"""
-    # def __init__(self, n_states, n_inputs, Q=None, R=None) -> None:
     def __init__(self, x_ref, vehicle, **kwargs):
 
         self.type      = 'LQR'                # type of controller
@@ -88,7 +87,6 @@ class LQR():
         self.parameters = dict(int_lag=100,             # error integral length: how far back in time to compute integral error
                                Q=np.diag([1000, 1000, 25000, 100.0, 100.0, 100.0, 1000, 1000, 5000, 1000, 1000, 1000]), # state error penalty matrix: how 'bad' is an error in the state vector w.r.t. the reference state vector
                                R=np.diag([500, 4000, 4000, 4000]), # input penalty matrix: how 'hard' it is to produce the desired input (thrust and three moments along three axes)
-                               # qi=np.array([100, 100, 1000]),   # integral error penalty: how much the error history is contributing to the overall error
                                scheduled_var='psi',     # variable used to create the scheduled controller gains (only psi allowed for now)
                                index_scheduled_var=5)   # index corresponding to the scheduled_var (psi) in the state vector x; i.e., x[5] = psi
         self.parameters.update(kwargs)                  # update control parameters according to user
@@ -145,12 +143,12 @@ class LQR():
 
         n, m = state_vector_vals.shape
         assert n == self.n_states, "number of states set at initialization and size of state_vector_vals mismatch."
-        if self.type == 'LQR':
-            self.control_gains = np.zeros((self.n_inputs, self.n_states, m))
-        elif self.type == 'LQR_I':
-            self.control_gains = np.zeros((self.n_inputs, self.n_states + self.n_outputs, m))
-        else: 
-            raise ProgModelInputException("Controller {} is not supported. Supported controllers: {}".format(self.type, ['LQR', 'LQR_I']))
+        # if self.type == 'LQR':
+        self.control_gains = np.zeros((self.n_inputs, self.n_states, m))
+        # elif self.type == 'LQR_I':
+        #     self.control_gains = np.zeros((self.n_inputs, self.n_states + self.n_outputs, m))
+        # else: 
+        #     raise ProgModelInputException("Controller {} is not supported. Supported controllers: {}".format(self.type, ['LQR', 'LQR_I']))
         self.scheduled_states = state_vector_vals
 
         for j in range(m):
@@ -161,22 +159,14 @@ class LQR():
 
         print('Control gain matrices complete.')
 
-class LQR_I():
+class LQR_I(LQR):
     """ Linear Quadratic Regulator with Integral Effect"""
 
     def __init__(self, x_ref, vehicle, **kwargs):
 
+        super().__init__(x_ref, vehicle, **kwargs)
+
         self.type      = 'LQR_I'                # type of controller
-        self.states    = vehicle.states         # state variables of the system to be controlled (x, y, z, phi, theta, psi)
-        self.n_states  = len(self.states) - 1   # number of states (minus one to remove time)
-        self.outputs   = vehicle.outputs[:3]    # output variables of the system to be controlled (x, y, z only)
-        self.n_outputs = 3                      # number of outputs
-        self.inputs    = vehicle.inputs         # input variables of the system to be controlled ()
-        self.n_inputs  = len(self.inputs)       # number of inputs
-        self.err_hist  = []                     # error history (integral)
-        self.ref_traj  = x_ref                  # reference state to follow during simulation (x_ref, y_ref, z_ref, phi_ref, theta_ref, psi_ref, ...)
[…]
