+++ tests/__main__.py
@@ -17,6 +17,7 @@ from tests.test_linear_model import main as linear_main
 from tests.test_composite import main as composite_main
 from tests.test_serialization import main as serialization_main
 from tests.test_ensemble import main as ensemble_main
+from tests.test_uav_model import main as uav_main
 
 if __name__ == '__main__':
     was_successful = True
@@ -107,6 +108,11 @@ if __name__ == '__main__':
 
     try:
         ensemble_main()
+    except Exception:
+        was_successful = False
+
+    try:
+        uav_main()
     except Exception:
         was_successful = False
 

[…]
