+++ tests/test_sim_result.py
@@ -5,6 +5,7 @@ import numpy as np
 import pickle
 import sys
 import unittest
+import pandas as pd
 
 from prog_models.models import BatteryElectroChemEOD
 from prog_models.sim_result import SimResult, LazySimResult
@@ -13,160 +14,211 @@ from prog_models.utils.containers import DictLikeMatrixWrapper
 
 class TestSimResult(unittest.TestCase):
     """def setUp(self):
-        # set stdout (so it wont print)
+        set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
         sys.stdout = sys.__stdout__"""
-    
+
     def test_sim_result(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
+        frame = pd.DataFrame(state)
+        frame.insert(0, "time", time)
         result = SimResult(time, state)
+        # Checks values from SimResult object and static variables
         self.assertListEqual(list(result), state)
         self.assertListEqual(result.times, time)
+        self.assertTrue(frame.equals(result.frame))
         for i in range(5):
             self.assertEqual(result.get_time(i), time[i])
             self.assertEqual(result[i], state[i])
-
+        self.assertTrue(frame.equals(result.frame))
         try:
-            tmp = result[NUM_ELEMENTS]
+            tmp = result[5]
+            self.fail("Should be out of range error")
+        except IndexError:
+            pass
+        try:
+            tmp = result.times[5]
             self.fail("Should be out of range error")
         except IndexError:
             pass
 
-        try:
-            tmp = result.get_time(NUM_ELEMENTS)
-            self.fail("Should be out of range error")
-        except IndexError:
-            pass
-    
     def test_pickle(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))  # list of int, 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
         result = SimResult(time, state)
         pickle.dump(result, open('model_test.pkl', 'wb'))
         result2 = pickle.load(open('model_test.pkl', 'rb'))
         self.assertEqual(result, result2)
+        self.assertTrue(result.frame.equals(result2.frame))
 
     def test_extend(self):
-        NUM_ELEMENTS = 5  # Creating two result objects
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))  # list of int from 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(5)]
         result = SimResult(time, state)
-        NUM_ELEMENTS = 10
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
-        result2 = SimResult(time, state)
-        self.assertEqual(result.times, [0, 1, 2, 3, 4])
[…]

[…]
