+++ tests/test_base_models.py
@@ -12,10 +12,9 @@ import pandas as pd
 # This ensures that the directory containing ProgModelTemplate is in the python search directory
 sys.path.append(join(dirname(__file__), ".."))
 
-from prog_models import ProgModelTypeError, ProgModelInputException, ProgModelException, PrognosticsModel, CompositeModel
-from prog_models.models import ThrownObject, BatteryElectroChemEOD
+from prog_models import *
+from prog_models.models import *
 from prog_models.models.test_models.linear_models import (OneInputNoOutputNoEventLM, OneInputOneOutputNoEventLM, OneInputNoOutputOneEventLM, OneInputOneOutputNoEventLMPM)
-from prog_models.models.test_models.linear_thrown_object import (LinearThrownObject, LinearThrownDiffThrowingSpeed, LinearThrownObjectUpdatedInitalizedMethod, LinearThrownObjectDiffDefaultParams)
 
 
 class MockModel():
@@ -42,7 +41,7 @@ class MockModel():
         return self.OutputContainer({'o1': x['a'] + x['b'] + x['c']})
 
 
-class MockProgModel(MockModel, PrognosticsModel):
+class MockProgModel(MockModel, prognostics_model.PrognosticsModel):
     events = ['e1', 'e2']
 
     def event_state(self, x):
@@ -69,6 +68,42 @@ def derived_callback3(config):
     return {  # Testing 2nd chained update 
         'p4': -2 * config['p2'], 
     }
+
+class LinearThrownObject(LinearModel):
+    inputs = [] 
+    states = ['x', 'v']
+    outputs = ['x']
+    events = ['impact']
+
+    A = np.array([[0, 1], [0, 0]])
+    E = np.array([[0], [-9.81]])
+    C = np.array([[1, 0]])
+    F = None # Will override method
+
+    default_parameters = {
+        'thrower_height': 1.83,  # m
+        'throwing_speed': 40,  # m/s
+        'g': -9.81  # Acceleration due to gravity in m/s^2
+    }
+
+    def initialize(self, u=None, z=None):
+        return self.StateContainer({
+            'x': self.parameters['thrower_height'],  # Thrown, so initial altitude is height of thrower
+            'v': self.parameters['throwing_speed']  # Velocity at which the ball is thrown - this guy is a professional baseball pitcher
+            })
+    
+    def threshold_met(self, x):
+        return {
+            'falling': x['v'] < 0,
+            'impact': x['x'] <= 0
+        }
+
+    def event_state(self, x): 
+        x_max = x['x'] + np.square(x['v'])/(-self.parameters['g']*2) # Use speed and position to estimate maximum height
+        return {
+            'falling': np.maximum(x['v']/self.parameters['throwing_speed'],0),  # Throwing speed is max speed
+            'impact': np.maximum(x['x']/x_max,0) if x['v'] < 0 else 1  # 1 until falling begins, then it's fraction of height
+        }
 
 class MockModelWithDerived(MockProgModel):
     param_callbacks = {
@@ -97,10 +132,7 @@ class TestModels(unittest.TestCase):
                 }
                 return ns_ret
         
-        m = MockProgModelStateDict(
-            process_noise_dist='none',
-            measurement_noise_dist='none')
-        
+        m = MockProgModelStateDict(process_noise_dist='none', measurement_noise_dist='none')
         def load(t, x=None):
             return {'i1': 1, 'i2': 2.1}
 
@@ -121,13 +153,11 @@ class TestModels(unittest.TestCase):
                     [x['t'] + dt]]
                 )
 
-        m = MockProgModelStateNdarray(
-            process_noise_dist='none', 
-            measurement_noise_dist='none')
+        m = MockProgModelStateNdarray(process_noise_dist='none', measurement_noise_dist='none')
 
         # Any event, default
-
[…]
