+++ src/prog_models/utils/next_state.py
@@ -1,146 +1,0 @@
-# Copyright © 2021 United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration.  All Rights Reserved.
-
-def euler_next_state(model, x, u, dt: float):
-    """
-    State transition equation using simple euler integration: Calls next_state(), calculating the next state, and then adds noise and applies limits
-
-    Parameters
-    ----------
-    x : StateContainer
-        state, with keys defined by model.states \n
-        e.g., x = m.StateContainer({'abc': 332.1, 'def': 221.003}) given states = ['abc', 'def']
-    u : InputContainer
-        Inputs, with keys defined by model.inputs \n
-        e.g., u = m.InputContainer({'i':3.2}) given inputs = ['i']
-    dt : float
-        Timestep size in seconds (≥ 0) \n
-        e.g., dt = 0.1
-
-    Returns
-    -------
-    x : StateContainer
-        Next state, with keys defined by model.states
-        e.g., x = m.StateContainer({'abc': 332.1, 'def': 221.003}) given states = ['abc', 'def']
-
-    See Also
-    --------
-    PrognosticsModel.next_state
-    """
-    # Calculate next state and add process noise
-    next_state = model.apply_process_noise(model.next_state(x, u, dt), dt)
-
-    # Apply Limits
-    return model.apply_limits(next_state)
-
-def euler_next_state_wrapper(model, x, u, dt: float):
-    """
-    State transition equation using simple euler integration: Calls next_state(), calculating the next state, and then adds noise and applies limits
-
-    Parameters
-    ----------
-    x : StateContainer
-        state, with keys defined by model.states \n
-        e.g., x = m.StateContainer({'abc': 332.1, 'def': 221.003}) given states = ['abc', 'def']
-    u : InputContainer
-        Inputs, with keys defined by model.inputs \n
-        e.g., u = m.InputContainer({'i':3.2}) given inputs = ['i']
-    dt : float
-        Timestep size in seconds (≥ 0) \n
-        e.g., dt = 0.1
-
-    Returns
-    -------
-    x : StateContainer
-        Next state, with keys defined by model.states
-        e.g., x = m.StateContainer({'abc': 332.1, 'def': 221.003}) given states = ['abc', 'def']
-
-    See Also
-    --------
-    PrognosticsModel.next_state
-    """
-    # Calculate next state and add process noise
-    next_state = model.StateContainer(model.next_state(x, u, dt))
-    next_state = model.apply_process_noise(next_state, dt)
-
-    # Apply Limits
-    return model.apply_limits(next_state)
-
-def rk4_next_state(model, x, u, dt: float):
-    """
-    State transition equation using rungekutta4 integration: Calls next_state(), calculating the next state, and then adds noise and applies limits
-
-    Parameters
-    ----------
-    x : StateContainer
-        state, with keys defined by model.states \n
-        e.g., x = m.StateContainer({'abc': 332.1, 'def': 221.003}) given states = ['abc', 'def']
-    u : InputContainer
-        Inputs, with keys defined by model.inputs \n
-        e.g., u = m.InputContainer({'i':3.2}) given inputs = ['i']
-    dt : float
-        Timestep size in seconds (≥ 0) \n
-        e.g., dt = 0.1
-
-    Returns
-    -------
-    x : StateContainer
-        Next state, with keys defined by model.states
[…]
