+++ src/prog_models/models/__init__.py
@@ -1,15 +1,14 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
-from prog_models.models.battery_circuit import BatteryCircuit
-from prog_models.models.battery_electrochem import BatteryElectroChem, BatteryElectroChemEOD, BatteryElectroChemEOL, BatteryElectroChemEODEOL
-from prog_models.models.centrifugal_pump import CentrifugalPump, CentrifugalPumpBase, CentrifugalPumpWithWear
-from prog_models.models.pneumatic_valve import PneumaticValve, PneumaticValveBase, PneumaticValveWithWear
-from prog_models.models.dcmotor import DCMotor
-from prog_models.models.dcmotor_singlephase import DCMotorSP
-from prog_models.models.esc import ESC
-from prog_models.models.powertrain import Powertrain
-from prog_models.models.propeller_load import PropellerLoad
-from prog_models.models.thrown_object import ThrownObject
-from prog_models.models.experimental.paris_law import ParisLawCrackGrowth
-from prog_models.models.thrown_object import LinearThrownObject
+from .battery_circuit import BatteryCircuit
+from .battery_electrochem import BatteryElectroChem, BatteryElectroChemEOD, BatteryElectroChemEOL, BatteryElectroChemEODEOL
+from .centrifugal_pump import CentrifugalPump, CentrifugalPumpBase, CentrifugalPumpWithWear
+from .pneumatic_valve import PneumaticValve, PneumaticValveBase, PneumaticValveWithWear
+from .dcmotor import DCMotor
+from .dcmotor_singlephase import DCMotorSP
+from .esc import ESC
+from .powertrain import Powertrain
+from .propeller_load import PropellerLoad
+from .thrown_object import ThrownObject
+from .experimental.paris_law import ParisLawCrackGrowth
