### 2023-05-08 e250352
Update prognostics_model.py
