### 2023-05-08 d487e0c
Cleaned up fastdtw and added to workflow
