+++ tests/test_estimate_params.py
@@ -7,6 +7,7 @@ import unittest
 from prog_models import *
 from prog_models.models import *
 
+
 class TestEstimateParams(unittest.TestCase):
     def test_estimate_params_works(self):
         """
@@ -65,11 +66,7 @@ class TestEstimateParams(unittest.TestCase):
         keys = ['thrower_height', 'throwing_speed', 'g']
         m.estimate_params(times=results.times, inputs=results.inputs, outputs=results.outputs, keys=keys)
         for key in keys:
-            self.assertNotAlmostEqual(m.parameters[key], gt[key], 2)
-        # These returned values are still within the bounds, they are just not close to the original value at all.
-        # This results in our estimate parameters to result in the upper extremes
-        self.assertAlmostEqual(m.parameters['thrower_height'], 4)
-        self.assertAlmostEqual(m.parameters['throwing_speed'], 42)
+            self.assertAlmostEqual(m.parameters[key], gt[key], 2)
 
 
     def test_estimate_params(self):
@@ -121,13 +118,6 @@ class TestEstimateParams(unittest.TestCase):
         self.assertEqual(
              'Times, inputs, and outputs must be same length. Length of times: 0, Length of inputs: 0, Length of outputs: 9',
              str(cm.exception)
-        )
-        
-        with self.assertRaises(ValueError) as cm:
-            m.estimate_params(times=results.times, inputs=None, outputs=None)
-        self.assertEqual(
-            'Missing keyword arguments inputs, outputs',
-            str(cm.exception)
         )
         
         with self.assertRaises(ValueError) as cm:
@@ -682,35 +672,37 @@ class TestEstimateParams(unittest.TestCase):
         """
         m = ThrownObject()
 
-        # Default Values 
-        m.parameters['thrower_height'] = 3.1
-        m.parameters['throwing_speed'] = 29
-        m.parameters['g'] = 10
         # The value of time1, time2, inputs, and outputs are arbitrary values
 
         time1 = [0, 1, 2, 4, 5, 6, 7, 8, 9]
         time2 = [0, 1, 2, 3]
 
-        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = 1e-9)
-        track2 = m.calc_error(results.times, results.inputs, results.outputs)
+        inputs = [[{}]*9, [{}]*4]
+        outputs = [[{'x': 1.83},
+            {'x': 36.95},
+            {'x': 62.36},
+            {'x': 77.81},
+            {'x': 83.45},
+            {'x': 79.28},
+            {'x': 65.3},
+            {'x': 41.51},
+            {'x': 7.91},], 
+            [
+                {'x': 1.83},
+                {'x': 36.95},
+                {'x': 62.36},
+                {'x': 77.81},
+            ]]
         
-        self.assertNotAlmostEqual(track1, track2)
         
         # Checking to see if multiple runs can exist.
         # time1 and time2 are explicitly being passed in into a parent wrapper list.
         # See definitions of variables to understand format.
         m.estimate_params(times=[time1, time2], inputs=inputs, outputs=outputs)
 
-        # Reset parameters
-        m.parameters['thrower_height'] = 3.1
-        m.parameters['throwing_speed'] = 29
-        m.parameters['g'] = 10
         # Checking to see if wrapping in tuple works.
         m.estimate_params(times=(time1, time2), inputs=inputs, outputs=outputs)
 
-        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, 
-                          bounds=bound, keys=keys, method = 'TNC', tol = 1e-4)
-        track1 = m.calc_error(results.times, results.inputs, results.outputs)
[…]
