+++ tests/test_uav_model.py
@@ -11,8 +11,8 @@ from prog_models.aux_fcns.traj_gen import trajectory_gen_fcn as traj_gen
 from prog_models.models.uav_model import UAVGen 
 from prog_models.loading_fcns.controllers import LQR_I, LQR
 from prog_models.exceptions import ProgModelInputException
+from prog_models.aux_fcns.traj_gen_utils import geometry as geom
 
-# TODO: adjust tests for new trajectory generation structure 
 class TestUAVGen(unittest.TestCase):
     
     def setUp(self):
@@ -34,6 +34,11 @@ class TestUAVGen(unittest.TestCase):
         waypoints['alt_ft']    = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
         waypoints['time_unix'] = np.array([1544188336, 1544188358, 1544188360, 1544188377, 1544188394, 1544188411, 1544188428, 1544188496, 1544188539, 1544188584, 1544188601, 1544188635, 1544188652, 1544188672, 1544188692])
 
+        waypoints_no_time = {}
+        waypoints_no_time['lat_deg']   = np.array([37.09776, 37.09776, 37.09776, 37.09798, 37.09748, 37.09665, 37.09703, 37.09719, 37.09719, 37.09719, 37.09719, 37.09748, 37.09798, 37.09776, 37.09776])
+        waypoints_no_time['lon_deg']   = np.array([-76.38631, -76.38629, -76.38629, -76.38589, -76.3848, -76.38569, -76.38658, -76.38628, -76.38628, -76.38628, -76.38628, -76.3848, -76.38589, -76.38629, -76.38629])
+        waypoints_no_time['alt_ft']    = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
+       
         # Incorrect number of inputs to traj_gen
         with self.assertRaises(ProgModelInputException):
             # No waypoint information provided
@@ -46,143 +51,107 @@ class TestUAVGen(unittest.TestCase):
             ref_traj = traj_gen(waypoints=waypoints)   
 
         # Waypoints defined incorrectly 
+        # Wrong type for waypoints 
         with self.assertRaises(ProgModelInputException):
-            # Wrong waypoint format
-            ref_traj = traj_gen(waypoints={'flight_plan': [1, 2, 3]}, vehicle=vehicle)
+            # Waypoints defined incorrectly; must be dict or string specifying filename
+            ref_traj = traj_gen(waypoints=True, vehicle=vehicle)
         with self.assertRaises(ProgModelInputException):
-            # Wrong waypoint format
-            ref_traj = traj_gen(waypoints={'flight_plan': np.array([1, 2, 3])}, vehicle=vehicle)
+            # Waypoints defined incorrectly; must be dict or string specifying filename
+            ref_traj = traj_gen(waypoints=np.array([1,2,3]), vehicle=vehicle)
         with self.assertRaises(ProgModelInputException):
-            # Wrong format of waypoint information provided
-            ref_traj = traj_gen(waypoints={'flight_plan': {'lat_deg': [1, 2, 3], 'lon_deg': [1, 2, 3], 'alt_m': [1, 2, 3]}}, vehicle=vehicle)
+            # Waypoints defined with dictionary but values in dictionary are not numpy arrays, as required
[…]

[…]
