### 2023-05-09 584268a
First draft next_state rewrite
