+++ src/prog_models/prognostics_model.py
@@ -15,7 +15,7 @@ from prog_models.exceptions import ProgModelInputException, ProgModelTypeError, 
 from prog_models.sim_result import SimResult, LazySimResult
 from prog_models.utils import ProgressBar, calc_error
 from prog_models.utils.containers import DictLikeMatrixWrapper
-from prog_models.utils.next_state import euler_next_state, rk4_next_state, euler_next_state_wrapper, rk4_next_state_wrapper
+from prog_models.utils.next_state import next_state_functions
 from prog_models.utils.parameters import PrognosticsModelParameters
 from prog_models.utils.serialization import CustomEncoder, custom_decoder
 from prog_models.utils.size import getsizeof
@@ -359,14 +359,8 @@ class PrognosticsModel(ABC):
         ----
         A model should overwrite either `next_state` or `dx`. Override `dx` for continuous models, and `next_state` for discrete, where the behavior cannot be described by the first derivative
         """
-        # Note: Default is to use the dx method (continuous model) - overwrite next_state for continuous
         dx = self.dx(x, u)
-        if isinstance(x, DictLikeMatrixWrapper) and isinstance(dx, DictLikeMatrixWrapper):
-            return self.StateContainer(x.matrix + dx.matrix * dt)
-        elif isinstance(dx, dict) or isinstance(x, dict):
-            return self.StateContainer({key: x[key] + dx[key]*dt for key in dx.keys()})
-        else:
-            raise ValueError(f"ValueError: dx return must be of type StateContainer, was {type(dx)}")
+        return self.StateContainer({key: x[key] + dx[key]*dt for key in dx.keys()})
 
     @property
     def is_continuous(self):
@@ -583,7 +577,9 @@ class PrognosticsModel(ABC):
         Returns:
             bool: if the model is a state transition model
         """
-        has_overridden_transition = type(self).next_state != PrognosticsModel.next_state or type(self).dx != PrognosticsModel.dx
+        has_default_next_state = type(self).next_state == PrognosticsModel.next_state
+        has_integrator_next_state = type(self).next_state in next_state_functions.values()
+        has_overridden_transition = not (has_default_next_state or has_integrator_next_state) or type(self).dx != PrognosticsModel.dx
         return has_overridden_transition and len(self.states) > 0
 
     @property
@@ -835,7 +831,6 @@ class PrognosticsModel(ABC):
         config = { # Defaults
             't0': 0.0,
             'dt': ('auto', 1.0),
-            'integration_method': 'euler',
             'save_pts': [],
             'save_freq': 10.0,
             'horizon': 1e100, # Default horizon (in s), essentially inf
@@ -882,11 +877,14 @@ class PrognosticsModel(ABC):
             x = self.StateContainer(x)
         
         # Optimization
-        output = self.__output
-        threshold_met_eqn = self.threshold_met
         event_state = self.event_state
         load_eqn = future_loading_eqn
+        next_state = self.next_state
+        apply_noise = self.apply_process_noise
+        apply_limits = self.apply_limits
+        output = self.__output
         progress = config['progress']
+        threshold_met_eqn = self.threshold_met
 
         # Threshold Met Equations
         def check_thresholds(thresholds_met):
@@ -985,6 +983,12 @@ class PrognosticsModel(ABC):
                 u = future_loading_eqn(t, x)
                 return self.InputContainer(u)
 
+        if not isinstance(next_state(x, u, dt0), DictLikeMatrixWrapper):
+            # Wrapper around the next state equation
+            def next_state(x, u, dt):
+                x = self.next_state(x, u, dt)
+                return self.StateContainer(x)
+
         if not isinstance(self.output(x), DictLikeMatrixWrapper):
             # Wrapper around the output equation
             def output(x):
@@ -1001,29 +1005,10 @@ class PrognosticsModel(ABC):
[…]
