+++ src/prog_models/utils/parameters.py
@@ -9,13 +9,14 @@ import numpy as np
 import types
 from typing import Callable
 
+from prog_models.utils.next_state import next_state_functions
 from prog_models.utils.noise_functions import measurement_noise_functions, process_noise_functions
 from prog_models.utils.serialization import *
 from prog_models.utils.size import getsizeof
 from prog_models.exceptions import ProgModelTypeError
 
 from typing import TYPE_CHECKING
-if TYPE_CHECKING: # Fix circular import issue in PrognosticsModelParameters init
+if TYPE_CHECKING:  # Fix circular import issue in PrognosticsModelParameters init
     from prog_models.prognostics_model import PrognosticsModel
 
 
@@ -90,7 +91,19 @@ class PrognosticsModelParameters(UserDict):
         if key in self.callbacks:
             for callback in self.callbacks[key]:
                 changes = callback(self)
-                self.update(changes) # Merge in changes
+                self.update(changes)  # Merge in changes
+
+        # Handle setting integration_method. This will override the next_state method
+        if key == 'integration_method':
+            if self._m.is_discrete and self._m.is_state_transition_model:
+                raise ProgModelTypeError(
+                    "Cannot set integration method for discrete model (where next_state is overridden)")
+            if value.lower() not in next_state_functions.keys():
+                raise ProgModelTypeError(
+                    f"Unsupported integration method {value.lower()}")
+            self._m.next_state = types.MethodType(
+                next_state_functions[value.lower()],
+                self._m)
         
         if key == 'process_noise' or key == 'process_noise_dist':
             if callable(self['process_noise']):  # Provided a function
