+++ src/prog_models/data_models/data_model.py
@@ -44,6 +44,7 @@ class DataModel(PrognosticsModel, ABC):
                 List of :term:`output` keys
             event_keys (list[str]): 
                 List of :term:`event` keys
+            frame -- List[pd.DataFrame] of all data
         
         See specific data class for more additional keyword arguments
 
@@ -160,6 +161,7 @@ class DataModel(PrognosticsModel, ABC):
         outputs = [d.outputs for d in data]
         states = [d.states for d in data]
         event_states = [d.event_states for d in data]
+
         t_met = [[list(m.threshold_met(x).values()) for x in state] for state in states]
 
         return cls.from_data(times = times, inputs = inputs, states = states, outputs = outputs, event_states = event_states, t_met= t_met, **config)
