+++ tests/__init__.py
@@ -1,4 +1,7 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
-__all__ = ['test_base_models', 'test_battery', 'test_examples', 'test_pneumatic_valve', 'test_dict_like_matrix_wrapper', 'test_tutorials', 'test_datasets', 'test_manual', 'test_linear_model', 'test_composite']
+# __all__ = ['test_base_models', 'test_battery', 'test_examples', 'test_pneumatic_valve', 'test_dict_like_matrix_wrapper', 'test_tutorials', 'test_datasets', 'test_manual', 'test_linear_model', 'test_composite']
+
+# without DictLike wrapper...
+__all__ = ['test_base_models', 'test_battery', 'test_examples', 'test_pneumatic_valve', 'test_tutorials', 'test_datasets', 'test_manual', 'test_linear_model', 'test_composite']
