+++ src/prog_models/utils/containers_df.py
@@ -1,6 +1,6 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
-
+import numpy as np
 from numpy import float64, matrix, ndarray, array, newaxis, nan, delete, atleast_1d, array_equal
 import pandas as pd
 from prog_models.exceptions import ProgModelTypeError
@@ -25,23 +25,24 @@ class DictLikeMatrixWrapper():
         self._keys = keys.copy()
 
         if isinstance(data, matrix):
-            self.data = pd.DataFrame(array(data, dtype=float64), self._keys, dtype=float64)
-            self.matrix = self.data.to_numpy(dtype=float64)
+            self.data = pd.DataFrame(array(data, dtype=float64), index=self._keys, dtype=float64).T
+            self._matrix = self.data.T.to_numpy(dtype=float64)
         elif isinstance(data, ndarray):
             if data.ndim == 1:
                 data = data[newaxis].T
                 self.data = pd.DataFrame(data, self._keys)
-            self.data = pd.DataFrame(data, self._keys).T
-            self.matrix = data
+            self.data = pd.DataFrame(dict(zip(self._keys, data)), dtype=float64)
+            self._matrix = self.data.T.to_numpy(dtype=float64)
         elif isinstance(data, (dict, DictLikeMatrixWrapper)):
             if data and not isinstance(list(data.values())[0], ndarray):  # len(self.matrix[0]) == 1:
                 if isinstance(data, DictLikeMatrixWrapper):
                     data = dict(data.copy())
-                self.data = pd.DataFrame(data, columns=self._keys, index=[0]).replace(
+                self.data = pd.DataFrame(data, columns=self._keys, index=[0], dtype=float64).replace(
                     nan, None)
             else:
                 self.data = pd.DataFrame(data, columns=self._keys)
-            self.matrix = self.data.to_numpy().T if len(data) > 0 else array([])
+            self._matrix = self.data.to_numpy(dtype=float64).T if len(data) > 0 else array([])
+
         else:
             raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
@@ -49,7 +50,7 @@ class DictLikeMatrixWrapper():
         """
         reduce is overridden for pickles
         """
-        return DictLikeMatrixWrapper, (self._keys, self.matrix)
+        return DictLikeMatrixWrapper, (self._keys, self._matrix)
 
     def __getitem__(self, key: str) -> int:
         """
@@ -66,7 +67,8 @@ class DictLikeMatrixWrapper():
         sets a row at the key given
         """
         index = self._keys.index(key)  # the int value index for the key given
-        self.matrix[index] = atleast_1d(value)
+        self._matrix[index] = atleast_1d(value)
+        self.data.loc[:, key] = float64(value)
 
     def __delitem__(self, key: str) -> None:
         """
@@ -75,13 +77,13 @@ class DictLikeMatrixWrapper():
         # self.matrix = delete(self.matrix, self._keys.index(key), axis=0)
         self._keys.remove(key)
         self.data = self.data.drop(columns=[key], axis=1)
-        self.matrix = self.data.T.to_numpy()
+        self._matrix = self.data.T.to_numpy(dtype=float64)
 
     def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
         """
         add another matrix to the existing matrix
         """
-        rowadded = self.data.add(other.data).T.to_numpy()
+        rowadded = self.data.add(other.data).T.to_numpy(dtype=float64)
[…]
