+++ src/prog_models/utils/containers_df.py
@@ -1,237 +1,0 @@
-# Copyright © 2021 United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration.  All Rights Reserved.
-import numpy as np
-from numpy import float64, matrix, ndarray, array, newaxis, nan, delete, atleast_1d, array_equal
-import pandas as pd
-from prog_models.exceptions import ProgModelTypeError
-from typing import Union
-
-
-class DictLikeMatrixWrapper():
-    """
-    A container that uses pandas dictionary like data structure, but is backed by a numpy array, which is itself directly accessible. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-
-    Arguments:
-        keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
-        data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
-    """
-
-    def __init__(self, keys: list, data: Union[dict, array]):
-        """
-        Initializes the container
-        """
-        if not isinstance(keys, list):
-            keys = list(keys)  # creates list with keys
-        self._keys = keys.copy()
-
-        if isinstance(data, matrix):
-            self.data = pd.DataFrame(array(data, dtype=float64), index=self._keys, dtype=float64).T
-            self._matrix = self.data.T.to_numpy(dtype=float64)
-        elif isinstance(data, ndarray):
-            if data.ndim == 1:
-                data = data[newaxis].T
-                self.data = pd.DataFrame(data, self._keys)
-            self.data = pd.DataFrame(dict(zip(self._keys, data)), dtype=float64)
-            self._matrix = self.data.T.to_numpy(dtype=float64)
-        elif isinstance(data, (dict, DictLikeMatrixWrapper)):
-            if data and not isinstance(list(data.values())[0], ndarray):  # len(self.matrix[0]) == 1:
-                if isinstance(data, DictLikeMatrixWrapper):
-                    data = dict(data.copy())
-                self.data = pd.DataFrame(data, columns=self._keys, index=[0], dtype=float64).replace(
-                    nan, None)
-            else:
-                self.data = pd.DataFrame(data, columns=self._keys)
-            self._matrix = self.data.to_numpy(dtype=float64).T if len(data) > 0 else array([])
-
-        else:
-            raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
-
-    def __reduce__(self):
-        """
-        reduce is overridden for pickles
-        """
-        return DictLikeMatrixWrapper, (self._keys, self._matrix)
-
-    def __getitem__(self, key: str) -> int:
-        """
-        get all values associated with a key, ex: all values of 'i'
-        """
-        row = self.data.loc[:, key].to_list()  # creates list from a column of pandas DF
-        if len(row) == 1:  # list contains 1 value, returns that value (non-vectorized)
-            return row[0]
-        else:
-            return row  # returns entire row/list (vectorized case)
-
-    def __setitem__(self, key: str, value: int) -> None:
-        """
-        sets a row at the key given
-        """
-        index = self._keys.index(key)  # the int value index for the key given
-        self._matrix[index] = atleast_1d(value)
-        self.data.loc[:, key] = float64(value)
-
-    def __delitem__(self, key: str) -> None:
-        """
-        removes row associated with key
-        """
-        # self.matrix = delete(self.matrix, self._keys.index(key), axis=0)
-        self._keys.remove(key)
-        self.data = self.data.drop(columns=[key], axis=1)
-        self._matrix = self.data.T.to_numpy(dtype=float64)
-
-    def __add__(self, other: "DictLikeMatrixWrapper") -> "DictLikeMatrixWrapper":
-        """
[…]
