+++ tests/test_battery.py
@@ -23,7 +23,7 @@ def future_loading(t, x=None):
 
 class TestBattery(unittest.TestCase):
     def setUp(self):
-        # set stdout (so it wont print)
+        # set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
