+++ tests/test_powertrain.py
@@ -9,7 +9,7 @@ from prog_models.models import ESC, DCMotor, Powertrain, PropellerLoad
 
 class TestPowertrain(unittest.TestCase):
     def setUp(self):
-        # set stdout (so it wont print)
+        # set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
