+++ src/prog_models/utils/containers.py
@@ -1,136 +1,119 @@
 # Copyright © 2021 United States Government as represented by the Administrator of the
 # National Aeronautics and Space Administration.  All Rights Reserved.
+
 import numpy as np
-from numpy import float64, matrix, ndarray, array, newaxis, nan, delete, atleast_1d, array_equal
 import pandas as pd
-from prog_models.exceptions import ProgModelTypeError
 from typing import Union
+
+from prog_models.exceptions import ProgModelTypeError
 
 
 class DictLikeMatrixWrapper():
     """
-    A container that uses pandas dictionary like data structure, but is backed by a numpy array, which is itself directly accessible. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
+    A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
 
     Arguments:
         keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
         data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
     """
 
-    def __init__(self, keys: list, data: Union[dict, array]):
-        """
-        Initializes the container
+    def __init__(self, keys: list, data: Union[dict, np.array]):
+        """ Initializes the container
         """
         if not isinstance(keys, list):
             keys = list(keys)  # creates list with keys
         self._keys = keys.copy()
-        if isinstance(data, matrix):
-            self.data = pd.DataFrame(array(data, dtype=float64), index=self._keys, dtype=float64).T
-        elif isinstance(data, ndarray):
+        if isinstance(data, np.matrix):
+            self.matrix = np.array(data, dtype=np.float64)
+        elif isinstance(data, np.ndarray):
             if data.ndim == 1:
-                data = data[newaxis].T
-                self.data = pd.DataFrame(data, self._keys)
-            self.data = pd.DataFrame(dict(zip(self._keys, data)), dtype=float64)
+                data = data[np.newaxis].T
+            self.matrix = data
         elif isinstance(data, (dict, DictLikeMatrixWrapper)):
-            if data and not isinstance(list(data.values())[0], ndarray):  # len(self.matrix[0]) == 1:
-                if isinstance(data, DictLikeMatrixWrapper):
-                    data = dict(data.copy())
-                self.data = pd.DataFrame(data, columns=self._keys, index=[0], dtype=float64).replace(
-                    nan, None)
-            else:
-                self.data = pd.DataFrame(data, columns=self._keys)
+            # ravel is used to prevent vectorized case, where data[key] returns multiple values,  from resulting in a 3D matrix
+            self.matrix = np.array(
+                [
+                    np.ravel([data[key]]) if key in data else [None] for key in keys
+                ], dtype=np.float64)
         else:
             raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
 
     @property
-    def matrix(self) -> pd.DataFrame:
+    def frame(self) -> pd.DataFrame:
         """
-        Getter -- creating matrix from self.data
+            frame -- Getter for pd.DataFrame
 
-            Return: ndarray, of all the data
+            Returns: pd.DataFrame
         """
-        return self.data.values.T
-
-    @matrix.setter
-    def matrix(self, new_value: Union[float, int, ndarray]) -> None:
-        """
-        Setter -- changes values in self.data based on index
-
-            Return: None
-        """
-        self.data.loc[:, :] = new_value.T
-
+        self._frame = pd.DataFrame(self.matrix.T, columns=self._keys)
+        return self._frame
 
     def __reduce__(self):
         """
[…]
