+++ src/prog_models/models/dcmotor.py
@@ -194,7 +194,6 @@ class DCMotor(PrognosticsModel):
         dxdt = np.dot(Ac, x.matrix) + np.dot(self.parameters['Bc'], u.matrix)
         x.matrix += dxdt * dt  # Update inplace
         x.matrix[4] %= PI2  # Wrap angle
-
         return x
 
     def output(self, x):
