+++ tests/test_direct.py
@@ -12,7 +12,7 @@ from prog_models.models import ThrownObject
 
 class TestDirect(unittest.TestCase):
     def setUp(self):
-        # set stdout (so it wont print)
+        # set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
