+++ tests/test_pneumatic_valve.py
@@ -10,7 +10,7 @@ from prog_models.models.pneumatic_valve import PneumaticValve, PneumaticValveWit
 
 class TestPneumaticValve(unittest.TestCase):
     def setUp(self):
-        # set stdout (so it wont print)
+        # set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
