+++ tests/test_sim_result.py
@@ -5,6 +5,7 @@ import numpy as np
 import pickle
 import sys
 import unittest
+import pandas as pd
 
 from prog_models.models import BatteryElectroChemEOD
 from prog_models.sim_result import SimResult, LazySimResult
@@ -12,70 +13,63 @@ from prog_models.utils.containers import DictLikeMatrixWrapper
 
 
 class TestSimResult(unittest.TestCase):
-    def setUp(self):
-        # set stdout (so it won't print)
+    """def setUp(self):
+        set stdout (so it won't print)
         sys.stdout = StringIO()
 
     def tearDown(self):
-        sys.stdout = sys.__stdout__
+        sys.stdout = sys.__stdout__"""
 
     def test_sim_result(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
         result = SimResult(time, state)
+        # Checks values from SimResult object and static variables
         self.assertListEqual(list(result), state)
         self.assertListEqual(result.times, time)
         for i in range(5):
-            self.assertEqual(result.time(i), time[i])
+            self.assertEqual(result.get_time(i), time[i])
             self.assertEqual(result[i], state[i])
-
         try:
-            tmp = result[NUM_ELEMENTS]
+            tmp = result[5]
             self.fail("Should be out of range error")
         except IndexError:
             pass
-
         try:
-            tmp = result.time(NUM_ELEMENTS)
+            tmp = result.times[5]
             self.fail("Should be out of range error")
         except IndexError:
             pass
 
     def test_pickle(self):
-        NUM_ELEMENTS = 5
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))  # list of int, 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 5} for i in range(5)]
         result = SimResult(time, state)
         pickle.dump(result, open('model_test.pkl', 'wb'))
         result2 = pickle.load(open('model_test.pkl', 'rb'))
         self.assertEqual(result, result2)
 
     def test_extend(self):
-        NUM_ELEMENTS = 5  # Creating two result objects
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(NUM_ELEMENTS)]
+        # Variables
+        time = list(range(5))  # list of int from 0 to 4
+        state = [{'a': i * 2.5, 'b': i * 2.5} for i in range(5)]
         result = SimResult(time, state)
-        NUM_ELEMENTS = 10
-        time = list(range(NUM_ELEMENTS))
-        state = [{'a': i * 5, 'b': i * 5} for i in range(NUM_ELEMENTS)]
-        result2 = SimResult(time, state)
-        self.assertEqual(result.times, [0, 1, 2, 3, 4])
-        self.assertEqual(result2.times, [0, 1, 2, 3, 4, 5, 6, 7, 8, 9])
-        self.assertEqual(result.data,
-                         [{'a': 0.0, 'b': 0.0}, {'a': 2.5, 'b': 2.5}, {'a': 5.0, 'b': 5.0}, {'a': 7.5, 'b': 7.5},
-                          {'a': 10.0, 'b': 10.0}])  # Assert data is correct before extending
[…]
