+++ src/prog_models/aux_fcns/traj_gen_utils/loadsave_utils.py
@@ -10,8 +10,6 @@ Loading utility functions
 import numpy as np
 import datetime as dt
 
-from prog_models.exceptions import ProgModelInputException
-
 # AUXILIARY CONVERSION
 # ====================
 DEG2RAD  = np.pi / 180.0
@@ -28,21 +26,21 @@ def load_traj_from_txt(fname, skiprows=1, comments='#', max_rows=None):
     # Check units and return warnings if incorrect:
     data_id = np.loadtxt(fname=fname, dtype='str', max_rows=1)
     if 'lat_deg' not in data_id and 'lat_rad' not in data_id:
-        raise ProgModelInputException("Waypoints latitude must be defined in degrees (with 'lat_deg') or radians (with 'lat_rad').")
+        raise TypeError("Waypoints latitude must be defined in degrees (with 'lat_deg') or radians (with 'lat_rad').")
     if 'lon_deg' not in data_id and 'lon_rad' not in data_id:
-        raise ProgModelInputException("Waypoints longitude must be defined in degrees (with 'lon_deg') or radians (with 'lon_rad').")
+        raise TypeError("Waypoints longitude must be defined in degrees (with 'lon_deg') or radians (with 'lon_rad').")
     if 'alt_ft' not in data_id and 'alt_m' not in data_id:
-        raise ProgModelInputException("Waypoints altitude must be defined in feet (with 'alt_ft') or meters (with 'alt_m').")    
+        raise TypeError("Waypoints altitude must be defined in feet (with 'alt_ft') or meters (with 'alt_m').")    
     if not (data_id[0] == 'lat_deg' or data_id[0] == 'lat_rad'):
-        raise ProgModelInputException("Waypoint latitudes must be the first column in text file.")
+        raise TypeError("Waypoint latitudes must be the first column in text file.")
     if not (data_id[1] == 'lon_deg' or data_id[1] == 'lon_rad'):
-        raise ProgModelInputException("Waypoint longitudes must be the second column in text file.") 
+        raise TypeError("Waypoint longitudes must be the second column in text file.") 
     if not (data_id[2] == 'alt_m' or data_id[2] == 'alt_ft'):
-        raise ProgModelInputException("Waypoint altitudes must be the third column in text file.")
+        raise TypeError("Waypoint altitudes must be the third column in text file.")
     if data_id.shape[0] > 3 and data_id[3] != 'time_unix':
-        raise ProgModelInputException("ETAs must be defined in unix time (with 'time_unix').")
+        raise TypeError("ETAs must be defined in unix time (with 'time_unix').")
     if data_id.shape[0] > 4:
-        raise ProgModelInputException("Too much waypoint information provided. Only latitude, longitude, altitude, and time is accepted.")
+        raise TypeError("Too much waypoint information provided. Only latitude, longitude, altitude, and time is accepted.")
 
     # Convert, if necessary
     d = np.loadtxt(fname=fname, skiprows=skiprows, comments=comments, max_rows=max_rows)
