+++ tests/test_uav_model.py
@@ -6,12 +6,12 @@ import unittest
 import numpy as np
 from scipy.interpolate import interp1d
 
-# from prog_models.models.uav_model.uav_model import UAVGen 
 from prog_models.aux_fcns.traj_gen import trajectory_gen_fcn as traj_gen
 from prog_models.models.uav_model import UAVGen 
 from prog_models.loading_fcns.controllers import LQR_I, LQR
-from prog_models.exceptions import ProgModelInputException
+from prog_models.exceptions import ProgModelInputException, ProgModelException
 from prog_models.aux_fcns.traj_gen_utils import geometry as geom
+from prog_models.loading_fcns.controllers import LQR_I, LQR
 
 class TestUAVGen(unittest.TestCase):
     
@@ -40,81 +40,81 @@ class TestUAVGen(unittest.TestCase):
         waypoints_no_time['alt_ft']    = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
        
         # Incorrect number of inputs to traj_gen
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # No waypoint information provided
             ref_traj = traj_gen()
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Only vehicle model (no waypoints) provided
             ref_traj = traj_gen(vehicle=vehicle)
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Only waypoints (no vehicle model) provided
             ref_traj = traj_gen(waypoints=waypoints)   
 
         # Waypoints defined incorrectly 
         # Wrong type for waypoints 
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Waypoints defined incorrectly; must be dict or string specifying filename
             ref_traj = traj_gen(waypoints=True, vehicle=vehicle)
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Waypoints defined incorrectly; must be dict or string specifying filename
             ref_traj = traj_gen(waypoints=np.array([1,2,3]), vehicle=vehicle)
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Waypoints defined with dictionary but values in dictionary are not numpy arrays, as required
             ref_traj = traj_gen(waypoints={'lat_deg': [1, 2, 3], 'lon_deg': [1, 2, 3]}, vehicle=vehicle)
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Waypoints defined with dictionary but values in dictionary are not numpy arrays, as required
             ref_traj = traj_gen(waypoints={'lat_deg': '1'}, vehicle=vehicle)
-        with self.assertRaises(ProgModelInputException):
+        with self.assertRaises(TypeError):
             # Waypoints defined with dictionary but values in dictionary are not numpy arrays, as required
             ref_traj = traj_gen(waypoints={'lat_deg': np.array([1, 2, 3]), 'lon_deg': [1, 2, 3]}, vehicle=vehicle)
         
         # Missing information
[…]

[…]
