### 2023-05-11 af1fc7e
Update src/prog_models/models/battery_electrochem.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
