### 2023-05-11 4acf165
Update dmd.py
edited based on Chris Teubert suggestion
