+++ src/prog_models/models/pneumatic_valve.py
@@ -6,7 +6,6 @@ import numpy as np
 import warnings
 
 from prog_models import PrognosticsModel
-from prog_models.utils.containers import DictLikeMatrixWrapper
 
 
 def calc_x(x: float, forces: float, Ls: float, new_x: float) -> float:
