### 2023-05-11 af55cb5
Update sim_result.py
it didn't allow me to accept the suggestion
