### 2023-05-11 6879cb5
Update sim_result.py
I committed a hybrid of what I had originally and what Chris Teubert suggested. I tested both his and mine and my if statement was faster than the for loop, but his editing of the df was better than mine.
