### 2023-05-11 015696c
Update tests/test_surrogates.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
