### 2023-05-11 64bd638
Update tests/test_linear_model.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
