### 2023-05-11 082955d
Update tests/test_pneumatic_valve.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
