### 2023-05-11 decd7e2
Update tests/test_centrifugal_pump.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
