+++ src/prog_models/utils/containers.py
@@ -2,14 +2,17 @@
 # National Aeronautics and Space Administration.  All Rights Reserved.
 
 import numpy as np
+import pandas as pd
 from typing import Union
+from warnings import warn
 
 from prog_models.exceptions import ProgModelTypeError
 
+
 class DictLikeMatrixWrapper():
     """
-    A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessible. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
-    
+    A container that behaves like a dictionary, but is backed by a numpy array, which is itself directly accessable. This is used for model states, inputs, and outputs- and enables efficient matrix operations.
+
     Arguments:
         keys -- list: The keys of the dictionary. e.g., model.states or model.inputs
         data -- dict or numpy array: The contained data (e.g., :term:`input`, :term:`state`, :term:`output`). If numpy array should be column vector in same order as keys
@@ -19,7 +22,7 @@ class DictLikeMatrixWrapper():
         """ Initializes the container
         """
         if not isinstance(keys, list):
-            keys = list(keys)   # creates list with keys
+            keys = list(keys)  # creates list with keys
         self._keys = keys.copy()
         if isinstance(data, np.matrix):
             self.matrix = np.array(data, dtype=np.float64)
@@ -28,13 +31,24 @@ class DictLikeMatrixWrapper():
                 data = data[np.newaxis].T
             self.matrix = data
         elif isinstance(data, (dict, DictLikeMatrixWrapper)):
-          # ravel is used to prevent vectorized case, where data[key] returns multiple values,  from resulting in a 3D matrix
-          self.matrix = np.array(
-             [
-                  np.ravel([data[key]]) if key in data else [None] for key in keys
-              ], dtype=np.float64)
+            # ravel is used to prevent vectorized case, where data[key] returns multiple values,  from resulting in a 3D matrix
+            self.matrix = np.array(
+                [
+                    np.ravel([data[key]]) if key in data else [None] for key in keys
+                ], dtype=np.float64)
         else:
             raise ProgModelTypeError(f"Data must be a dictionary or numpy array, not {type(data)}")
+
+    @property
+    def frame(self) -> pd.DataFrame:
+        """
+            frame -- Getter for pd.DataFrame
+
+            Returns: pd.DataFrame
+        """
+        warn('frame will be deprecated after version 1.5 of ProgPy.', DeprecationWarning, stacklevel=2)
+        self._frame = pd.DataFrame(self.matrix.T, columns=self._keys)
+        return self._frame
 
     def __reduce__(self):
         """
@@ -46,23 +60,23 @@ class DictLikeMatrixWrapper():
         """
         get all values associated with a key, ex: all values of 'i'
         """
-        row = self.matrix[self._keys.index(key)]    #creates list from a row of matrix
-        if len(row) == 1:   #list contains 1 value, returns that value (non-vectorized)
+        row = self.matrix[self._keys.index(key)]  # creates list from a row of matrix
+        if len(row) == 1:  # list contains 1 value, returns that value (non-vectorized)
             return row[0]
-        return row #returns entire row/list (vectorized case)
+        return row  # returns entire row/list (vectorized case)
 
     def __setitem__(self, key: str, value: int) -> None:
         """
         sets a row at the key given
         """
-        index = self._keys.index(key)   #the int value index for the key given
+        index = self._keys.index(key)  # the int value index for the key given
         self.matrix[index] = np.atleast_1d(value)
 
-    
     def __delitem__(self, key: str) -> None:
         """
         removes row associated with key
         """
[…]
