+++ setup.py
@@ -16,6 +16,7 @@ INSTALL_REQS = [
         'matplotlib',
         'requests',  # For data downloading features
         'chaospy',  # For PCE
+        'fastdtw', # For DTW error calculation
         "tensorflow; platform_system!='Darwin' or platform_machine!='arm64'", 
         "tensorflow-macos; platform_system=='Darwin' and platform_machine=='arm64'",
     ]
