+++ src/prog_models/sim_result.py
@@ -120,7 +120,7 @@ class SimResult(UserList):
         self.data = []
         self._frame = None
 
-    def get_time(self, index: int) -> float:
+    def time(self, index: int) -> float:
         """Get time for data point at index `index`
 
         Args:
