### 2023-05-12 4a41ae3
Update src/prog_models/sim_result.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
