### 2023-05-12 139ca0e
Update src/prog_models/sim_result.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
