+++ src/prog_models/sim_result.py
@@ -2,6 +2,8 @@
 
 from collections import UserList, defaultdict
 from copy import deepcopy
+from warnings import warn
+
 from matplotlib.pyplot import figure
 import numpy as np
 import pandas as pd
