### 2023-05-13 da8bb89
Added additional Deprecation warnings and dot()
