### 2023-05-13 5d8d861
Remove | (not supported in 3.7)
