### 2023-05-16 c605b77
Created test, dtw alg much greater error
