+++ src/prog_models/utils/containers.py
@@ -139,7 +139,7 @@ class DictLikeMatrixWrapper():
         """
         Compares two DictLikeMatrixWrappers (i.e. *Containers) or a DictLikeMatrixWrapper and a dictionary
         """
-        warn("Behavior of '==' operator will change after version 1.5 of ProgPy. New behavior will return element wise equality as a new series. To check if two data frames are equals use pd.DataFrame.equals and pd.Series.equals.", DeprecationWarning, stacklevel=2)
+        warn("Behavior of '==' operator will change after version 1.5 of ProgPy. New behavior will return element wise equality as a new series. To check if two Containers are equal use container.equal(other).", DeprecationWarning)
         return self.equals(other)
 
     def __hash__(self):
