### 2023-05-16 221ae68
requested changes for warnings and functions
