### 2023-05-16 ffb4382
requested changes for warnings and functions, 2nd
