### 2023-05-17 ceeb7d0
requested changes for warnings and functions, 3rd
