### 2023-05-17 a38ecc4
Fix issue with next_state
