### 2023-05-19 6c29829
Update noise functions
