+++ src/prog_models/utils/noise_functions.py
@@ -6,16 +6,22 @@ import numpy as np
 # ---------------------------
 # Measurement Noise Functions
 # ---------------------------
-def uniform_measurement_noise(self, z : dict):
-    return self.OutputContainer(z.matrix + np.random.uniform(-1*self.parameters['measurement_noise'].matrix, self.parameters['measurement_noise'].matrix, size=z.matrix.shape))
+def uniform_measurement_noise(self, z: dict):
+    noise_mat = self.parameters['measurement_noise'].matrix
+    z.matrix = z.matrix + np.random.uniform(-1*noise_mat, noise_mat, size=z.matrix.shape)
+    return z
 
-def triangular_measurement_noise(self, z : dict):
-    return self.OutputContainer(z.matrix + np.random.triangular(-1*self.parameters['measurement_noise'].matrix, 0, self.parameters['measurement_noise'].matrix, size=z.matrix.shape))
+def triangular_measurement_noise(self, z: dict):
+    noise_mat = self.parameters['measurement_noise'].matrix
+    z.matrix = z.matrix + np.random.triangular(-1*noise_mat, 0, noise_mat, size=z.matrix.shape)
+    return z
 
-def normal_measurement_noise(self, z : dict):
-    return self.OutputContainer(z.matrix + np.random.normal(0, self.parameters['measurement_noise'].matrix, size=z.matrix.shape))
+def normal_measurement_noise(self, z: dict):
+    noise_mat = self.parameters['measurement_noise'].matrix
+    z.matrix = z.matrix + np.random.normal(0, noise_mat, size=z.matrix.shape)
+    return z
 
-def no_measurement_noise(self, z : dict) -> dict:
+def no_measurement_noise(self, z: dict) -> dict:
     return z
 
 measurement_noise_functions = {
@@ -30,16 +36,25 @@ measurement_noise_functions = {
 # Process Noise Functions
 # ---------------------------
 
-def triangular_process_noise(self, x : dict, dt : int =1):
-    return self.StateContainer(x.matrix + dt*np.random.triangular(-1*self.parameters['process_noise'].matrix, 0, self.parameters['process_noise'].matrix, size=x.matrix.shape))
+def triangular_process_noise(self, x: dict, dt: int = 1):
+    noise_mat = self.parameters['process_noise'].matrix
+    noise = np.random.triangular(-1*noise_mat, 0, noise_mat, size=x.matrix.shape)
+    x.matrix = x.matrix + dt*noise
+    return x
 
-def uniform_process_noise(self, x : dict, dt : int =1):
-    return self.StateContainer(x.matrix + dt*np.random.uniform(-1*self.parameters['process_noise'].matrix, self.parameters['process_noise'].matrix, size=x.matrix.shape))
+def uniform_process_noise(self, x: dict, dt: int = 1):
+    noise_mat = self.parameters['process_noise'].matrix
+    noise = np.random.uniform(-1*noise_mat, noise_mat, size=x.matrix.shape)
+    x.matrix = x.matrix + dt*noise
+    return x
 
-def normal_process_noise(self, x : dict, dt : int =1):
-    return self.StateContainer(x.matrix + dt*np.random.normal(0, self.parameters['process_noise'].matrix, size=x.matrix.shape))
+def normal_process_noise(self, x: dict, dt: int = 1):
+    noise_mat = self.parameters['process_noise'].matrix
+    noise = np.random.normal(0, noise_mat, size=x.matrix.shape)
+    x.matrix = x.matrix + dt*noise
+    return x
 
-def no_process_noise(self, x : dict, dt :int =1) -> dict:
+def no_process_noise(self, x: dict, dt: int = 1) -> dict:
     return x
 
 process_noise_functions = {
