### 2023-05-19 0e92d88
Apply suggestions from code review
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
