+++ .github/workflows/python-package.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   test:
-    timeout-minutes: 45
+    timeout-minutes: 30
     runs-on: ubuntu-latest
     strategy:
       matrix:
