+++ paper.bib
@@ -19,7 +19,8 @@
   journal={International Journal of Prognostics and Health Management},
   volume={8},
   number={2},
-  year={2017}
+  year={2017},
+  doi={10.36001/ijphm.2017.v8i2.2618}
 }
 
 @misc{2016_nasa_prog_model_library,
@@ -58,7 +59,8 @@
   booktitle={Annual Conference of the PHM Society},
   volume={14},
   number={1},
-  year={2022}
+  year={2022},
+  doi={10.36001/phmconf.2022.v14i1.3238}
 }
 
 @misc{Dataset_RWBattery,
