### 2023-05-22 nasa/prog_models#530 @teubert
Thank you @nkrusch for your comments. I've taken all of your suggestions. The changes to the paper can be found in here: https://github.com/nasa/prog_models/commit/bccff282fb7b58ebf937809a1f212be3f33b42f6 and I added some instructions for running unit tests here: https://github.com/nasa/progpy/commit/820fcd627b9acc404a3e98eaee67bde25953d11c

Re: your issues with the tests. Tensorflow has been a real headache for me! I'm actually also on an M1 mac and am not running into the described issue, but some other M1 mac users have described similar behavior. I've found it challenging to identify exactly what configuration will lead to that issue.

Let me know if you think my changes resolve your review comments.
