### 2023-05-22 041f410
Improvements on controller functions: removed a for loop in lqr_fn by using index-based extraction of eigenvectors. Removed duplicate function build_scheduled_control from LQR_I by initializing control_gains attribute differently for LQR and LQR_I; dimensions are now correct. Removed err_hist and outputs attributes from LQR and added them in LQR_I, since they are only relevant for the integral part of the controller.
