+++ src/prog_models/loading_fcns/controllers.py
@@ -45,12 +45,7 @@ def lqr_fn(A, B, Q, R):
 
     # --------- Extract eigevectors whose eigenvalues have real part < 0 ------------ #
     eig_val, eig_vec = np.linalg.eig(HM)
-    e = len(eig_val)
-    V_ = np.zeros((e, 1))
-    for ii in range(e):
-        if np.real(eig_val[ii]) < 0:
-            V_ = np.concatenate((V_, eig_vec[:, ii].reshape((e,1))), axis=1)
-    V_ = V_[:, 1:]
+    V_ = eig_vec[:, np.real(eig_val)<0.0]
 
     # --------- Partition matrix of resulting eigenvectors ------------ #
     X = V_[:n, :n] # first half (row-wise)
@@ -81,15 +76,12 @@ class LQR():
         self.type      = 'LQR'                  # type of controller
         self.states    = vehicle.states         # state variables of the system to be controlled (x, y, z, phi, theta, psi)
         self.n_states  = len(self.states) - 1   # number of states (minus one to remove time)
-        self.outputs   = vehicle.outputs[:3]    # output variables of the system to be controlled (x, y, z only)
-        self.n_outputs = 3                      # number of outputs
         self.inputs    = vehicle.inputs         # input variables of the system to be controlled ()
         self.n_inputs  = len(self.inputs)       # number of inputs
-        self.err_hist  = []                     # error history (integral)
         self.ref_traj  = x_ref                  # reference state to follow during simulation (x_ref, y_ref, z_ref, phi_ref, theta_ref, psi_ref, ...)
         self.ss_input  = vehicle.vehicle_model.steadystate_input
         self.vehicle_max_thrust = vehicle.vehicle_model.dynamics['max_thrust']
-
+        
         # Default control parameters
         # --------------------------------
         self.parameters = dict(int_lag=100,             # error integral length: how far back in time to compute integral error
@@ -104,8 +96,12 @@ class LQR():
 
         # Initialize other controller-related variables
         # ---------------------------------------------
-        self.dt        = vehicle.parameters['dt']
-    
+        self.dt = vehicle.parameters['dt']
+
+        # Initialize control gain storage matrix:
+        # ---------------------------------------
+        self.control_gains = np.zeros((self.n_inputs, self.n_states, 1))
+
     def __call__(self, t, x=None):
 
         # Check that build_scheduled_control has been called
@@ -156,7 +152,7 @@ class LQR():
 
         n, m = state_vector_vals.shape
         assert n == self.n_states, "number of states set at initialization and size of state_vector_vals mismatch."
-        self.control_gains = np.zeros((self.n_inputs, self.n_states, m))
+        self.control_gains = np.zeros((self.control_gains.shape[0], self.control_gains.shape[1], m))
         self.scheduled_states = state_vector_vals
 
         for j in range(m):
@@ -175,6 +171,9 @@ class LQR_I(LQR):
         super().__init__(x_ref, vehicle, **kwargs)
 
         self.type      = 'LQR_I'                # type of controller
+        self.outputs   = vehicle.outputs[:3]    # output variables of the system to be controlled (x, y, z only)
+        self.n_outputs = 3                      # number of outputs
+        self.err_hist  = []                     # error history (integral)
 
         # Default control parameters
         # --------------------------------
@@ -188,7 +187,7 @@ class LQR_I(LQR):
 
         # Initialize other controller-related variables
         # ---------------------------------------------
-        self.C         = np.zeros((self.n_outputs, self.n_states))
+        self.C = np.zeros((self.n_outputs, self.n_states))
[…]
