### 2023-05-22 27f09e2
Added a new class that is supposed to replace UAVGen (if accepted) and Rotorcraft class from AircraftModels.py. The reason for this chance is that the Rotorcraft class in AircraftModels was initialized but basically never used. A lot of overhead for using, at the end, UAVGen for basically everything. I left the UAVGen because I haven't verified if other pieces of the code still use it.
