+++ src/prog_models/models/uav_model/uav_model.py
@@ -470,3 +470,390 @@ class UAVGen(PrognosticsModel):
         ax2.set_ylabel('z', fontsize=params['fontsize'])
 
         return fig
+
+
+
+
+# Small Rotorcraft Model
+# ======================
+class SmallRotorcraft(PrognosticsModel):
+
+    events = ['TrajectoryComplete']
+    inputs = ['T','mx','my','mz']
+    n_inputs = 4
+    states = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r','t']
+    n_states = 12
+    # states = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r','t', 'mission_complete']
+    outputs = ['x', 'y', 'z', 'phi', 'theta', 'psi', 'vx', 'vy', 'vz', 'p', 'q', 'r']
+    n_outputs = 12
+    
+    is_vectorized = True
+
+    default_parameters = {  # Set to defaults
+
+        # Simulation parameters:
+        'dt': 0.1, 
+        'gravity': 9.81,
+        'air_density': 1.225,
+        'steadystate_input': None,
+        'final_time_buffer_sec': 30, 
+        'final_space_buffer_m': 2, 
+
+        # Vehicle parameters:
+        'vehicle_model': 'tarot18', 
+        'vehicle_payload': 0.0,
+        'vehicle_max_speed': 15.0, #### These are needed for traj_gen, are these okay defaults?
+        'vehicle_max_roll': 0.7853981633974483, ##### fix this 
+        'vehicle_max_pitch': 0.7853981633974483, #### fix this 
+        'ref_traj': None
+    }
+
+    def __init__(self, payload=0.0, gravity=9.81, air_density=1.225, **kwargs):
+        super().__init__(**kwargs)
+
+        # Update simulation-specific parameters
+        self.parameters['vehicle_payload'] = payload
+        self.parameters['gravity'] = gravity
+        self.parameters['air_density'] = air_density
+
+        # Select model
+        # ------------
+        if self.parameters['vehicle_model'].lower() == 'djis1000':
+            self.mass, self.geom, self.dynamics = AircraftModels.vehicles.DJIS1000(self.parameters['vehicle_payload'], gravity)
+        elif self.parameters['vehicle_model'].lower() == 'tarot18':   
+            self.mass, self.geom, self.dynamics = AircraftModels.vehicles.TAROT18(self.parameters['vehicle_payload'], gravity)
+        else:                                                         
+            self.mass, self.geom, self.dynamics = dict(), dict(), dict()
+        
+        # Steady-state input value: hover, [weight, 0, 0, 0]
+        if self.parameters['steadystate_input'] is None:
+            self.parameters['steadystate_input'] = self.mass['total'] * self.parameters['gravity']
+        
+        # Introduction of Aerodynamic effects:
+        self.aero = dict(drag=AircraftModels.aero.DragModel(bodyarea=self.dynamics['aero']['ad'],
+                                                            Cd=self.dynamics['aero']['cd'],
+                                                            air_density=self.parameters['air_density']),
+                        lift=None)
+        
+    def initialize(self, u=None, z=None): 
+      # Extract initial state from reference trajectory    
+      # TODO: needs to be more general, but don't have ref_traj here 
+      if u is None:
+        self.InputContainer({'T': self.parameters['steadystate_input'], 'mx': 0.0, 'my': 0.0, 'mz': 0.0})
+      else:
+        self.InputContainer(u)
+
+      return self.StateContainer({'x': 0, # self.ref_traj.cartesian_pos[0, 0],
[…]
