+++ tests/test_estimate_params.py
@@ -928,8 +928,15 @@ class TestEstimateParams(unittest.TestCase):
                           bounds=bound, keys = keys, tol = 10)
         if not all(abs(m.parameters[key] - gt[key]) > 0.02 for key in keys):
             raise ValueError("m.parameter shouldn't be too close to the original parameters")
+        
+        # When tolerance is 0, estimate_params converges to some minimum error
+        m.parameters['thrower_height'] = 3.1
+        m.parameters['throwing_speed'] = 29
+        m.parameters['g'] = 10
+        m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = 0)
+        self.assertNotEqual(m.calc_error(results.times, results.inputs, results.outputs), 0)
 
-        # Note that tolerance does not convert here
+        # Note that tolerance does not convert here.
         with self.assertRaises(TypeError):
             m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, 
                               bounds=bound, keys = keys , tol = "12")    
@@ -939,38 +946,32 @@ class TestEstimateParams(unittest.TestCase):
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = [])
+        hold1 = m.calc_error(results.times, results.inputs, results.outputs)
 
-
+        # Tolerance works as intended as long as it is within a sequence of length 1
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10  
-
-        # Tolerance works as intended as long as it is within a sequence of length 1
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = [15])
         hold2 = m.calc_error(results.times, results.inputs, results.outputs)
         # Making sure that it has a different value.
         self.assertNotEqual(hold1, hold2)
 
-
+        # Confirming that passing tolerance as a sequence of length 1 outputs the same error as setting tolerance to the value in the original sequence.
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
-
-        # Confirming that passing tolerance as a sequence of length 1 outputs the same error as setting tolerance to the value in the original sequence.
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = 15)
         hold1 = m.calc_error(results.times, results.inputs, results.outputs)
         self.assertEqual(hold1, hold2)
 
-
+        # Now testing with a different type of sequence
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
-
-        # Now testing with a different type of sequence
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys = keys, tol = (15))
         hold2 = m.calc_error(results.times, results.inputs, results.outputs)
         self.assertEqual(hold1, hold2)
-
 
         # Cannot pass Sets into tolerance
         with self.assertRaises(TypeError):
@@ -1003,45 +1004,40 @@ class TestEstimateParams(unittest.TestCase):
         m.estimate_params(times = results.times, inputs = results.inputs, outputs = results.outputs, keys=keys, method = 'TNC', tol = 1e-9)
         track1 = m.calc_error(results.times, results.inputs, results.outputs)
 
-
+        # Using different methods should result in different errors.
         m.parameters['thrower_height'] = 3.1
         m.parameters['throwing_speed'] = 29
         m.parameters['g'] = 10
-
[…]
