### 2023-05-23 e9806d4
removed tol = 0 test. diff behavior depending on version
