### 2023-05-24 0f8d9f7
cleaned up SmallRotorcraft class and removed AircraftModels, which is not necessary anymore now.
