### 2023-05-24 e2ef386
Final adjustments for deleting AircraftModels and class
