+++ src/prog_models/models/uav_model/uav_model.py
@@ -7,476 +7,10 @@ from warnings import warn
 from typing import Callable
 
 from prog_models.prognostics_model import PrognosticsModel
-# from .vehicles import AircraftModels
 from .vehicles.aero import aerodynamics as aero
 from .vehicles import vehicles
 from prog_models.aux_fcns.traj_gen_utils import geometry as geom
 from prog_models.exceptions import ProgModelInputException, ProgModelException
-
-class UAVGen(PrognosticsModel):
-    """
-    Vectorized prognostics :term:`model` to generate a predicted trajectory for a UAV using a n=6 degrees-of-freedom dynamic model
-    with feedback control loop. The model follows the form:
-    
-    u     = h(x, x_{ref})
-    dx/dt = f(x, \theta, u)
-    
-    where:
-      x is a 2n state vector containing position, attitude and corresponding derivatives
-      \theta is a vector of model parameters including UAV mass, inertia moment, aerodynamic coefficients, etc
-      u is the input vector: thrust along the body vertical axis, and three moments along the UAV body axis to follow the desired trajectory.
-      x_{ref} is the desired state vector at that specific time step, with dimension 2n
-      f(.) is growth rate function of all vechile state
-      h(.) is the feedback-loop control function that returns the necessary thrust and moments (u vector) to cover the error between desired state x_{ref} and current state x
-      dx/dt is the state-increment per unit time.
-
-    
-    Model generates cartesian positions and velocities, pitch, roll, and yaw, and angular velocities throughout time to satisfy some user-define waypoints. 
-
-    See [0]_ for modeling details. 
-
-    :term:`Events<event>`: (1)
-        TrajectoryComplete: All waypoints are completed 
-    
-    :term:`Inputs/Loading<input>`: (0)
-        | User-defined inputs: waypoints plus ETAs or speeds
-        | Model-defined inputs: 
-            | T: thrust
-            | mx: moment in x 
-            | my: moment in y
-            | mz: moment in z
-
-    :term:`States<state>`: (13)
-        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame, center is at first waypoint
-        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame, center is at first waypoint
-        | z: third position in cartesian reference frame East-North-Up (ENU), i.e., Up in fixed inertia frame, center is at first waypoint
-        | phi: Euler's first attitude angle
-        | theta: Euler's second attitude angle
-        | psi: Euler's third attitude angle
-        | vx: velocity along x-axis, i.e., velocity along East in fixed inertia frame
-        | vy: velocity along y-axis, i.e., velocity along North in fixed inertia frame
-        | vz: velocity along z-axis, i.e., velocity Up in fixed inertia frame
-        | p: angular velocity around UAV body x-axis
-        | q: angular velocity around UAV body y-axis 
-        | r: angular velocity around UAV body z-axis 
-        | t: time 
-
-    :term:`Outputs<output>`: (12)
-        | x: first position in cartesian reference frame East-North-Up (ENU), i.e., East in fixed inertia frame, center is at first waypoint
-        | y: second position in cartesian reference frame East-North-Up (ENU), i.e., North in fixed inertia frame, center is at first waypoint 
[…]
