### 2023-05-24 790bc77
First changes for mission_complete in inputs/ctrl
