+++ src/prog_models/loading_fcns/controllers.py
@@ -76,7 +76,8 @@ class LQR():
         
         self.type      = 'LQR'                  # type of controller
         self.states    = vehicle.states         # state variables of the system to be controlled (x, y, z, phi, theta, psi)
-        self.n_states  = len(self.states) - 1   # number of states (minus one to remove time)
+        # self.n_states  = len(self.states) - 1   # number of states (minus one to remove time)
+        self.n_states  = len(self.states) - 2   # number of states (minus one to remove time)
         self.inputs    = vehicle.inputs         # input variables of the system to be controlled ()
         self.n_inputs  = len(self.inputs)       # number of inputs
         self.ref_traj  = x_ref                  # reference state to follow during simulation (x_ref, y_ref, z_ref, phi_ref, theta_ref, psi_ref, ...)
@@ -112,8 +113,8 @@ class LQR():
         if x is None:
             x_k = np.zeros((self.n_states, 1))
         else:
-            x_k = np.array([x.matrix[ii][0] for ii in range(len(x.matrix)-1)])
-            # x_k = np.array([x.matrix[ii][0] for ii in range(len(x.matrix)-2)])
+            # x_k = np.array([x.matrix[ii][0] for ii in range(len(x.matrix)-1)])
+            x_k = np.array([x.matrix[ii][0] for ii in range(len(x.matrix)-2)])
         
         # Identify reference state (desired state) at t
         t_k = np.round(t + self.dt/2.0, 1)  # current time step
@@ -132,8 +133,8 @@ class LQR():
         u             = self.compute_input(K, error)                 # compute input u given the gain matrix K and the error between current and reference state
         u[0]         += self.ss_input
         u[0]  = min(max([0, u[0]]), self.vehicle_max_thrust)
-        return {'T': u[0], 'mx': u[1], 'my': u[2], 'mz': u[3]}
-        # return {'T': u[0], 'mx': u[1], 'my': u[2], 'mz': u[3], 'mission_complete': t/self.ref_traj['t'][-1]}
+        # return {'T': u[0], 'mx': u[1], 'my': u[2], 'mz': u[3]}
+        return {'T': u[0], 'mx': u[1], 'my': u[2], 'mz': u[3], 'mission_complete': t/self.ref_traj['t'][-1]}
 
     def compute_gain(self, A, B):
         """ Compute controller gain given state of the system described by linear model A, B"""
