+++ src/prog_models/models/powertrain.py
@@ -93,7 +93,7 @@ class Powertrain(PrognosticsModel):
     }
 
     default_parameters = {
-        # Load parameters 
+        # Load parameters
         'c_q': 5.42e-7,  # coefficient of torque (APC data, derived) [dimensionless]
         'rho': 1.225,  # (Kg/m^3)
         'D': 0.381,  # (m)
