+++ src/prog_models/data_models/lstm_model.py
@@ -7,7 +7,6 @@ from numbers import Number
 import numpy as np
 import sys
 from tensorflow import keras
-from tensorflow.keras import layers
 from warnings import warn
 
 from prog_models.data_models import DataModel
@@ -209,7 +208,7 @@ class LSTMStateTransitionModel(DataModel):
         print("Window_size: ", self.parameters['window'], file=file)
         if self.parameters['state_model'] is not None:
             print('\nState Model: ', file=file)
-            self.parameters['state_model'].summary(print_fn= file.write, expand_nested=expand_nested, show_trainable=show_trainable)
+            self.parameters['state_model'].summary(print_fn=file.write, expand_nested=expand_nested, show_trainable=show_trainable)
         
         print('\nOutput Model: ', file=file)
         self.parameters['output_model'].summary(print_fn=file.write, expand_nested=expand_nested, show_trainable=show_trainable)
@@ -269,9 +268,9 @@ class LSTMStateTransitionModel(DataModel):
                     u_i = [[[u[i+j]] for j in range(1, window+1)] for i in range(len(u)-window)]
                 elif isinstance(u[0], (list, np.ndarray)):
                     # Input is d-d array
-                    # Note: 1 is added to account for current time (current input used to predict output at time i) 
+                    # Note: 1 is added to account for current time (current input used to predict output at time i)
                     n_inputs = len(u[0])
-                    u_i = [[[u[i+j][k] for k in range(n_inputs)] for j in range(1,window+1)] for i in range(len(u)-window)]
+                    u_i = [[[u[i+j][k] for k in range(n_inputs)] for j in range(1, window+1)] for i in range(len(u)-window)]
                 else:
                     raise TypeError(f"Unsupported input type: {type(u)} for internal element (data[0][i]")  
             else:
@@ -285,11 +284,12 @@ class LSTMStateTransitionModel(DataModel):
                     # No outputs
                     z = []
                 else:
-                    z = np.array([z_i.matrix[:,0] for z_i in z])
+                    z = np.array([z_i.matrix[:, 0] for z_i in z])
 
             if isinstance(z, (list, np.ndarray)):
                 if len(z) != len(u) and len(u) != 0 and len(z) != 0:
-                    # Checked here to avoid SimResults from accidentially triggering this check
+                    # Checked here to avoid SimResults from accidentially
+                    # triggering this check
                     raise IndexError(f"Number of outputs ({len(z)}) does not match number of inputs ({len(u)})")
 
                 if len(z) == 0:
@@ -303,7 +303,7 @@ class LSTMStateTransitionModel(DataModel):
                     n_outputs = len(z[0])
                     z_i = [[z[i][k] for k in range(n_outputs)] for i in range(window, len(z))]
                 else:
-                    raise TypeError(f"Unsupported input type: {type(z)} for internal element (output[i])")  
+                    raise TypeError(f"Unsupported input type: {type(z)} for internal element (output[i])")
 
                 # Also add to input (past outputs are part of input)
                 z_ii = [[z[i+j] for j in range(window)] for i in range(len(z_i))]
@@ -328,7 +328,8 @@ class LSTMStateTransitionModel(DataModel):
 
                 if isinstance(es, (list, np.ndarray)):
                     if len(es) != len(u) and len(u) != 0 and len(es) != 0:
-                        # Checked here to avoid SimResults from accidentially triggering this check
+                        # Checked here to avoid SimResults from accidentially
+                        # triggering this check
                         raise IndexError(f"Number of event_states ({len(es)}) does not match number of inputs ({len(u)})")
 
[…]
