+++ src/prog_models/ensemble_model.py
@@ -15,7 +15,7 @@ class EnsembleModel(PrognosticsModel):
         :language: python
         :class: highlight
     
-    Ensembe Models are constructed from a set of other models (e.g., :python:`m = EnsembleModel((m1, m2, m3))`). The models then operate functionally as one prognostic model. 
+    Ensembe Models are constructed from a set of other models (e.g., :python:`m = EnsembleModel((m1, m2, m3))`). The models then operate functionally as one prognostic model.
 
     See example :download:`examples.ensemble <../../../../prog_models/examples/ensemble.py>`
 
