+++ tests/test_calc_error.py
@@ -20,7 +20,6 @@ class TestCalcError(unittest.TestCase):
         Results from calc_error of DTW work as intended.
         """
         m = LinearThrownObject()
-        results = m.simulate_to_threshold(save_freq = 0.5)
 
         # Consistent results simulated.
         times = [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0, 7.5, 8.0, 8.5, 9.0]
