+++ tests/test_calc_error.py
@@ -32,6 +32,24 @@ class TestCalcError(unittest.TestCase):
         # Has been verified to result in desired values from other DTW algorithms.
         self.assertEqual(m.calc_error(times, inputs, outputs, method = 'dtw'), 4.8146507570483195)
 
+        # Testing case of inconsistent simulation of data (values from times and outputs have been deleted to substantially increase error)
+        times = [0.0, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 6.5, 7.5, 8.0, 8.5, 9.0]
+        inputs = [{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
+        outputs = [{'x': 2.83}, {'x': 40.422881456712254}, {'x': 55.51861687290881}, {'x': 68.06865643702567}, 
+                   {'x': 78.16641111234323}, {'x': 85.89550327176332}, {'x': 91.18868647545982}, {'x': 94.01376508127296}, 
+                   {'x': 94.31711597903195}, {'x': 87.67210201789473}, {'x': 44.79567793740186}, 
+                   {'x': 27.97245305860176}, {'x': 8.736607826437163}, {'x': -12.879687324031048}]
+        
+        DTW_err = m.calc_error(times, inputs, outputs, method='dtw')
+        self.assertEqual(DTW_err, 97.6842960920543)
+
+        # Since we have deleted a few values such that the results from times and outputs may not necessarily match,
+        # DTW would match simulated and observed data to each other's closest counterparts. 
+        # As such, DTW would have a naturally lower error than a standard error calculation method like 'mse'.
+        MSE_err = m.calc_error(times, inputs, outputs)
+        self.assertLess(DTW_err, MSE_err)
+
+
 
 
 def run_tests():
