+++ tests/test_estimate_params.py
@@ -75,8 +75,6 @@ class TestEstimateParams(unittest.TestCase):
         """
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
-        data = [(results.times, results.inputs, results.outputs)]
-        gt = m.parameters.copy()
 
         # Reset some parameters
         m.parameters['thrower_height'] = 1.5
@@ -514,8 +512,6 @@ class TestEstimateParams(unittest.TestCase):
         """
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
-        data = [(results.times, results.inputs, results.outputs)]
-        gt = m.parameters.copy()
 
         m.parameters['thrower_height'] = 1.5
         m.parameters['throwing_speed'] = 25
@@ -591,7 +587,6 @@ class TestEstimateParams(unittest.TestCase):
 
         m = ThrownObject()
         results = m.simulate_to_threshold(save_freq=0.5)
-        data = [(results.times, results.inputs, results.outputs)]
         gt = m.parameters.copy()
 
         # Now lets reset some parameters with a method call
