### 2023-05-27 c30cfce
Change coarse waypoints format to DataFrames
