+++ examples/uav_dynamics_model.py
@@ -4,8 +4,9 @@
 """
 Example of generating a trajectory for a rotorcraft through a set of coarse waypoints 
 """
-import numpy as np
 import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd 
 
 from prog_models.aux_fcns.traj_gen import trajectory_gen_fcn as traj_gen
 from prog_models.models.uav_model import SmallRotorcraft
@@ -24,14 +25,20 @@ def run_example():
     # Initialize vehicle 
     vehicle = SmallRotorcraft(**vehicle_params)
 
-    # Define coarse waypoints: waypoints must be defined with a dictionary of numpy arrays or as columns in a text file 
+    # EXAMPLE 1: 
+    # Define coarse waypoints: waypoints must be defined as a Pandas DataFrame
     # See documentation for specific information on inputting waypoints 
-    # Latitude, longitude, and altitude values are required; ETAs are optional 
+    # Latitude, longitude, and altitude values are required; ETAs are optional (see Example 3)
+
+    # Here, we specify waypoints in a dictionary and then convert it to a Pandas DataFrame
     waypoints = {}
     waypoints['lat_deg']   = np.array([37.09776, 37.09776, 37.09776, 37.09798, 37.09748, 37.09665, 37.09703, 37.09719, 37.09719, 37.09719, 37.09719, 37.09748, 37.09798, 37.09776, 37.09776])
     waypoints['lon_deg']   = np.array([-76.38631, -76.38629, -76.38629, -76.38589, -76.3848, -76.38569, -76.38658, -76.38628, -76.38628, -76.38628, -76.38628, -76.3848, -76.38589, -76.38629, -76.38629])
     waypoints['alt_ft']    = np.array([-1.9682394, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 164.01995, 0.0, 0.0, 164.01995, 164.01995, 164.01995, 164.01995, 0.0])
     waypoints['time_unix'] = np.array([1544188336, 1544188358, 1544188360, 1544188377, 1544188394, 1544188411, 1544188428, 1544188496, 1544188539, 1544188584, 1544188601, 1544188635, 1544188652, 1544188672, 1544188692])
+
+    # Convert waypoints to pandas dataframe
+    waypoints_pd = pd.DataFrame(waypoints)
 
     # Define reference trajectory parameters, if desired
     ref_params = {
@@ -39,8 +46,7 @@ def run_example():
     }
 
     # Calculate reference trajectory 
-    ref_traj = traj_gen(waypoints=waypoints, vehicle=vehicle, **ref_params)
-    vehicle.parameters['ref_traj'] = ref_traj
+    ref_traj = traj_gen(waypoints=waypoints_pd, vehicle=vehicle, **ref_params)
 
     # Define controller and build scheduled control 
     ctrl = LQR(ref_traj, vehicle)
@@ -53,14 +59,21 @@ def run_example():
     }
 
     # Simulate vehicle to fly trajectory 
-    # traj_results = vehicle.simulate_to_threshold(ctrl, **options)
-    traj_results = vehicle.simulate_to(350, ctrl, **options)
+    traj_results = vehicle.simulate_to_threshold(ctrl, **options)
 
     # Visualize Results
-    vehicle.visualize_traj(traj_results)
+    vehicle.visualize_traj(pred=traj_results, ref=ref_traj)
 
-    # Now define another trajectory through the same waypoints but with speeds defined instead of ETAs
+    # Note: waypoints can be specified in a variety of ways and then converted to a Pandas DataFrame for use 
[…]
