### 2023-05-27 776c476
Add back in uav_model test to __main__
