+++ tests/__main__.py
@@ -117,5 +117,10 @@ if __name__ == '__main__':
     except Exception:
         was_successful = False
 
+    try:
+        uav_main()
+    except Exception:
+        was_successful = False
+
     if not was_successful:
         raise Exception("Failed test")
