### 2023-05-27 3a9272e
Integrate into next_state
