+++ src/prog_models/utils/parameters.py
@@ -101,17 +101,18 @@ class PrognosticsModelParameters(UserDict):
             if self._m.is_discrete and self._m.is_state_transition_model:
                 raise ProgModelTypeError(
                     "Cannot set integration method for discrete model (where next_state is overridden)")
-            if issubclass(value, OdeSolver):
+            if isinstance(value, type) and issubclass(value, OdeSolver):
+                # the integration_method is a SciPy Integrator
                 fcn = SciPyIntegrateNextState(self._m, value)
                 self._m.next_state = types.MethodType(
-                fcn,
-                self._m)
+                    fcn,
+                    self._m)
                 return
             method = value.lower()
             if method in next_state_functions.keys():
                 self._m.next_state = types.MethodType(
-                next_state_functions[method],
-                self._m)
+                    next_state_functions[method],
+                    self._m)
                 return
             raise ProgModelTypeError(
                     f"Unsupported integration method {method}")
