### 2023-05-30 3f7e9c1
FIxed data location and added more comprehensive tests
