### 2023-05-30 59959d8
documenting acceptable types in calc_error
