### 2023-05-30 ea6d7a1
Wrote function defintion for check_iterable
