+++ output_calls.txt
@@ -1,573 +1,0 @@
-Fri May 19 10:22:53 2023    output.dat
-
-         11997312 function calls (11960206 primitive calls) in 3.321 seconds
-
-   Ordered by: call count
-
-   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-1817922/1814427    0.062    0.000    0.064    0.000 {built-in method builtins.len}
-  1356249    0.082    0.000    0.082    0.000 {method 'index' of 'list' objects}
-  1356249    0.433    0.000    0.558    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-  1192554    0.098    0.000    0.098    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-651193/625893    0.033    0.000    0.045    0.000 {built-in method builtins.isinstance}
-572945/572711    0.241    0.000    0.897    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-   460154    0.065    0.000    0.065    0.000 {built-in method numpy.asanyarray}
-   447070    0.019    0.000    0.019    0.000 {method 'append' of 'list' objects}
-   444426    0.021    0.000    0.021    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:19(_atleast_1d_dispatcher)
-   444426    0.222    0.000    0.379    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-   444426    0.100    0.000    0.670    0.000 <__array_function__ internals>:177(atleast_1d)
-   444408    0.061    0.000    0.061    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-   201608    0.098    0.000    0.098    0.000 {built-in method numpy.array}
-   155669    0.009    0.000    0.009    0.000 {method 'copy' of 'list' objects}
-   155637    0.084    0.000    0.149    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:18(__init__)
-   124945    0.006    0.000    0.006    0.000 {method 'items' of 'dict' objects}
-   111919    0.077    0.000    0.077    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-   111857    0.021    0.000    0.021    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
-   111857    0.067    0.000    0.226    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
-   110062    0.010    0.000    0.010    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2302(_any_dispatcher)
-   110062    0.034    0.000    0.255    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2307(any)
-   110062    0.028    0.000    0.360    0.000 <__array_function__ internals>:177(any)
-   104099    0.008    0.000    0.008    0.000 {built-in method builtins.getattr}
-   100010    0.015    0.000    0.080    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/_methods.py:54(_any)
-   100000    0.034    0.000    0.114    0.000 {method 'any' of 'numpy.generic' objects}
[…]
