+++ output_time.txt
@@ -1,573 +1,0 @@
-Fri May 19 10:22:53 2023    output.dat
-
-         11997312 function calls (11960206 primitive calls) in 3.321 seconds
-
-   Ordered by: internal time
-
-   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-    44612    0.525    0.000    0.947    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:506(output)
-  1356249    0.433    0.000    0.558    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-    44245    0.372    0.000    1.411    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:331(dx)
-572945/572711    0.241    0.000    0.897    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-   444426    0.222    0.000    0.379    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-    10024    0.100    0.000    0.493    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:392(apply_limits)
-   444426    0.100    0.000    0.670    0.000 <__array_function__ internals>:177(atleast_1d)
-  1192554    0.098    0.000    0.098    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-   201608    0.098    0.000    0.098    0.000 {built-in method numpy.array}
-   155637    0.084    0.000    0.149    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:18(__init__)
-  1356249    0.082    0.000    0.082    0.000 {method 'index' of 'list' objects}
-   111919    0.077    0.000    0.077    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-    44245    0.071    0.000    1.527    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:325(next_state)
-   111857    0.067    0.000    0.226    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
-   460154    0.065    0.000    0.065    0.000 {built-in method numpy.asanyarray}
-1817922/1814427    0.062    0.000    0.064    0.000 {built-in method builtins.len}
-   444408    0.061    0.000    0.061    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-      381    0.059    0.000    1.992    0.005 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/calc_error.py:126(MSE)
-    90396    0.034    0.000    0.055    0.000 {built-in method builtins.sum}
-   100000    0.034    0.000    0.114    0.000 {method 'any' of 'numpy.generic' objects}
-   110062    0.034    0.000    0.255    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2307(any)
-651193/625893    0.033    0.000    0.045    0.000 {built-in method builtins.isinstance}
-    10000    0.031    0.000    0.063    0.000 {method 'normal' of 'numpy.random.mtrand.RandomState' objects}
-   110062    0.028    0.000    0.360    0.000 <__array_function__ internals>:177(any)
[…]
