+++ output_calls.txt
@@ -1,562 +1,0 @@
-Mon Apr 17 12:05:35 2023    output.dat
-
-         7451103 function calls (7442973 primitive calls) in 2.058 seconds
-
-   Ordered by: call count
-
-   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-1075496/1075143    0.036    0.000    0.037    0.000 {built-in method builtins.len}
-   813790    0.049    0.000    0.049    0.000 {method 'index' of 'list' objects}
-   813790    0.255    0.000    0.331    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-   690834    0.056    0.000    0.056    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-362298/362136    0.161    0.000    0.622    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-357285/357201    0.017    0.000    0.018    0.000 {built-in method builtins.isinstance}
-   253302    0.011    0.000    0.011    0.000 {method 'append' of 'list' objects}
-   251058    0.034    0.000    0.034    0.000 {built-in method numpy.asanyarray}
-   251016    0.012    0.000    0.012    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:19(_atleast_1d_dispatcher)
-   251016    0.121    0.000    0.208    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-   251016    0.056    0.000    0.371    0.000 <__array_function__ internals>:177(atleast_1d)
-   251001    0.033    0.000    0.033    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-   161325    0.062    0.000    0.062    0.000 {built-in method numpy.array}
-   123569    0.006    0.000    0.006    0.000 {method 'items' of 'dict' objects}
-   110557    0.073    0.000    0.073    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-   110513    0.020    0.000    0.020    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
-   110513    0.066    0.000    0.218    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
-   110059    0.008    0.000    0.008    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2302(_any_dispatcher)
-   110059    0.033    0.000    0.250    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2307(any)
-   110059    0.027    0.000    0.351    0.000 <__array_function__ internals>:177(any)
-   102838    0.007    0.000    0.008    0.000 {built-in method builtins.getattr}
-   100009    0.015    0.000    0.078    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/_methods.py:54(_any)
-   100000    0.033    0.000    0.111    0.000 {method 'any' of 'numpy.generic' objects}
-    96154    0.005    0.000    0.005    0.000 {method 'copy' of 'list' objects}
-    96130    0.053    0.000    0.079    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:19(__init__)
[…]
