+++ output_time.txt
@@ -1,562 +1,0 @@
-Mon Apr 17 12:05:35 2023    output.dat
-
-         7451103 function calls (7442973 primitive calls) in 2.058 seconds
-
-   Ordered by: internal time
-
-   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
-    25389    0.294    0.000    0.532    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:506(output)
-   813790    0.255    0.000    0.331    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:45(__getitem__)
-    25021    0.204    0.000    0.780    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/models/battery_electrochem.py:331(dx)
-362298/362136    0.161    0.000    0.622    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
-   251016    0.121    0.000    0.208    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/shape_base.py:23(atleast_1d)
-    10008    0.097    0.000    0.482    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:391(apply_limits)
-   110557    0.073    0.000    0.073    0.000 {method 'reduce' of 'numpy.ufunc' objects}
-   110513    0.066    0.000    0.218    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:69(_wrapreduction)
-   161325    0.062    0.000    0.062    0.000 {built-in method numpy.array}
-   251016    0.056    0.000    0.371    0.000 <__array_function__ internals>:177(atleast_1d)
-   690834    0.056    0.000    0.056    0.000 /opt/homebrew/Cellar/python@3.10/3.10.8/Frameworks/Python.framework/Versions/3.10/lib/python3.10/collections/__init__.py:1101(__getitem__)
-    96130    0.053    0.000    0.079    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/containers.py:19(__init__)
-   813790    0.049    0.000    0.049    0.000 {method 'index' of 'list' objects}
-    25021    0.040    0.000    0.845    0.000 /Users/aditummala/Desktop/prog_models/src/prog_models/prognostics_model.py:324(next_state)
-1075496/1075143    0.036    0.000    0.037    0.000 {built-in method builtins.len}
-   251058    0.034    0.000    0.034    0.000 {built-in method numpy.asanyarray}
-   100000    0.033    0.000    0.111    0.000 {method 'any' of 'numpy.generic' objects}
-   110059    0.033    0.000    0.250    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:2307(any)
-   251001    0.033    0.000    0.033    0.000 {method 'reshape' of 'numpy.ndarray' objects}
-    10000    0.031    0.000    0.061    0.000 {method 'normal' of 'numpy.random.mtrand.RandomState' objects}
-   110059    0.027    0.000    0.351    0.000 <__array_function__ internals>:177(any)
-   110513    0.020    0.000    0.020    0.000 /opt/homebrew/lib/python3.10/site-packages/numpy/core/fromnumeric.py:70(<dictcomp>)
-    50778    0.019    0.000    0.019    0.000 {built-in method builtins.sum}
-      313    0.018    0.000    0.835    0.003 /Users/aditummala/Desktop/prog_models/src/prog_models/utils/calc_error.py:120(MSE)
[…]
