### 2023-05-30 30f36de
added documentation for a bunch of functions, although this pass is not complete.
