+++ src/prog_models/aux_fcns/traj_gen_utils/route.py
@@ -14,6 +14,20 @@ from prog_models.exceptions import ProgModelInputException
 # FUNCTIONS
 # ==========
 def check_and_adjust_eta_feasibility(lat, lon, alt, eta, vehicle_max_speed, vehicle_max_speed_vert, distance_method='greatcircle'):
+    """
+    Check that ETAs are feasible by verifying that the average speed necessary to 
+    reach the way-point is not beyond the maximum speed of the vehicle.
+    If so, the ETA for each way-point is increased progressively until the criterion is met.
+
+    :param lat:                         rad, n x 1, latitude of way-points
+    :param lon:                         rad, n x 1, longitude of way-points
+    :param alt:                         m, n x 1, altitude of way-points
+    :param eta:                         s, n x 1, ETAs of way-points
+    :param vehicle_max_speed:           m/s, scalar, vehicle max horizontal speed 
+    :param vehicle_max_speed_vert:      m/s, scalar, vehicle max speed ascending or descending
+    :param distance_method:             string, method to calculate the distance between two points, either greatcircle (default), or vicenti.
+    :return:                            s, n x 1, ETAs of way-points corrected to ensure feasibility.
+    """
     # Check if the speeds required for ETAs are beyond the max vehicle speeds.
     n = len(lat)-1
     d_eta = np.diff(eta)
@@ -27,7 +41,8 @@ def check_and_adjust_eta_feasibility(lat, lon, alt, eta, vehicle_max_speed, vehi
         while dh / d_eta[point] > vehicle_max_speed or dv / d_eta[point] > vehicle_max_speed_vert:
             d_eta[point] += 1.0
     return np.asarray(np.cumsum(np.insert(d_eta, 0, 0.0)))
-    
+
+
 def build(lat, lon, alt, departure_time, parameters: dict = dict(), etas=None, vehicle_max_speed=None):
     """
     Generate route given waypoints (lat, lon, alt), departure time, 
@@ -86,6 +101,14 @@ def build(lat, lon, alt, departure_time, parameters: dict = dict(), etas=None, v
 
 
 def reshape_route_attribute(x, dim=None, msk=None):
+    """
+    Reshape route attribute to ensure right length
+
+    :param x:           attribute to reshape
+    :param dim:         scalar, int or None, desired length of attribute. Used only if x has no attribute "len"
+    :param msk:         list or array of int, or None, mask to reshape attribute x. The function return x by repeating x[msk] at msk points.
+    :return:            reshaped attribute
+    """
     if not hasattr(x, "__len__"):   
         x = x * np.ones((dim, ))
     elif msk:
@@ -95,6 +118,9 @@ def reshape_route_attribute(x, dim=None, msk=None):
 # ROUTE CLASS
 # ============    
 class Route():
+    """
+    This class is used to generate a route with waypoints, corresponding etas, and the desired speed in between way-points.
+    """
     def __init__(self, 
                  departure_time=None, 
                  cruise_speed=None, 
@@ -102,6 +128,16 @@ class Route():
                  descent_speed=None, 
                  landing_speed=None, 
                  landing_alt=10.5):
+        """
+        Initialize class
+
+        :param departure_time:      dt.datetime object, departure time according to plan.
+        :param cruise_speed:        m/s, scalar, cruise speed in between waypoints
+        :param ascent_speed:        m/s, scalar, ascent speed in between waypoints
+        :param descent_speed:        m/s, scalar, descent speed in between waypoints
+        :param landing_speed:        m/s, scalar, landing speed, either ascending or descending speed when altitude is lower than landing_altitude
[…]
