### 2023-05-31 79da7cc
Rename loading_fcns to loading and documentation edits/typos
