+++ src/prog_models/aux_fcns/traj_gen_utils/geometry.py
@@ -159,8 +159,8 @@ def geodetic_distance(lats, lons, alts, method='greatcircle', return_surf_vert=F
 # ================
 def rot_eart2body_fast(sphi, cphi, stheta, ctheta, spsi, cpsi):
     """
-    Return the rotation matrix R to transform coordinates from a Earth-fixed (inertial) reference frame to a body reference frame.
-    The function is called "fast" because takes the pre-computed since, cosine, and other trigonomentric functions of Euler's angles.
+    Return the rotation matrix R to transform coordinates from an Earth-fixed (inertial) reference frame to a body reference frame.
+    The function is called "fast" because takes the pre-computed sine, cosine, and other trigonometric functions of Euler's angles.
 
     :param sphi:            scalar, sine of phi
     :param cphi:            scalar, cosine of phi
@@ -176,8 +176,8 @@ def rot_eart2body_fast(sphi, cphi, stheta, ctheta, spsi, cpsi):
 
 def rot_body2earth_fast(sphi, cphi, stheta, ctheta, spsi, cpsi):
     """
-    Return the rotation matrix R to transform coordinates from a body (non-inertial) reference frame to a Earth-fixed (inertial) reference frame.
-    The function is called "fast" because takes the pre-computed since, cosine, and other trigonomentric functions of Euler's angles.
+    Return the rotation matrix R to transform coordinates from a body (non-inertial) reference frame to an Earth-fixed (inertial) reference frame.
+    The function is called "fast" because takes the pre-computed sine, cosine, and other trigonometric functions of Euler's angles.
 
     :param sphi:            scalar, sine of phi
     :param cphi:            scalar, cosine of phi
@@ -211,8 +211,8 @@ def body_ang_vel_from_eulers(phi, theta, psi, phidot, thetadot, psidot):
 # ==========================
 def gen_heading_angle(lat, lon, alt):
     """
-    Function to generate heading angle to follow a set of points defined by latitude lat, longitude lon, and altitude alt.
-    The function returns the heading from North, positive clock-wise/
+    Function to generate heading angle to follow a set of points defined by latitude (lat), longitude (lon), and altitude (alt).
+    The function returns the heading from North, positive clock-wise
 
     :param lat:             rad, n x 1, latitude points
     :param lon:             rad, n x 1, longitude points
@@ -269,7 +269,7 @@ def heading_adjust_rotation(heading):
             continue
         else:
             next_p_2 = next_p - 2.0*np.pi   # angle in opposite direction
-            # Select angle based on mininum rotation necessary
+            # Select angle based on minimum rotation necessary
             if abs(curr_p - next_p) < abs(curr_p - next_p_2):   heading[idx+1] = next_p
             else:                                               heading[idx+1] = next_p_2
     return heading
@@ -344,14 +344,14 @@ class Coord():
 
     def ecef2enu(self, xecef, yecef, zecef):
         """
-        Coversion of Earth-centered, Earth-fixed (ECEF) reference frame coordinates to East-North-UP (ENU) reference frame.
+        Conversion of Earth-centered, Earth-fixed (ECEF) reference frame coordinates to East-North-UP (ENU) reference frame.
 
         :param xecef:           m, n x 1, X-coordinate in ECEF reference frame
         :param yecef:           m, n x 1, Y-coordinate in ECEF reference frame
         :param zecef:           m, n x 1, Z-coordinate in ECEF reference frame
-        :return x:              m, n x 1, x-coordainte (East) in ENU reference frame
-        :return y:              m, n x 1, y-coordainte (North) in ENU reference frame
[…]
