+++ src/prog_models/models/uav_model/vehicles/aero/aerodynamics.py
@@ -30,7 +30,6 @@ class DragModel():
         self.area = bodyarea
         self.cd = Cd
         self.rho = air_density
-        return
 
     def __call__(self, air_v):
         """
