+++ tests/test_uav_model.py
@@ -7,11 +7,11 @@ import numpy as np
 from scipy.interpolate import interp1d
 import pandas as pd
 
-from prog_models.aux_fcns.traj_gen import trajectory_gen_fcn as traj_gen
+from prog_models.utils.traj_gen import trajectory_gen as traj_gen
 from prog_models.models.uav_model import SmallRotorcraft 
-from prog_models.loading_fcns.controllers import LQR_I, LQR
+from prog_models.loading.controllers import LQR_I, LQR
 from prog_models.exceptions import ProgModelInputException, ProgModelException
-from prog_models.aux_fcns.traj_gen_utils import geometry as geom
+from prog_models.utils.traj_gen_utils import geometry as geom
 
 class TestUAVGen(unittest.TestCase):
     
