### 2023-05-31 ff66096
Adding test_uav_model to coverage assessment
