### 2023-05-31 2fdadf0
Update src/prog_models/sim_result.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
