+++ tests/test_battery.py
@@ -5,20 +5,13 @@ import sys
 import unittest
 
 from prog_models.models import BatteryCircuit, BatteryElectroChem, BatteryElectroChemEOL, BatteryElectroChemEOD, BatteryElectroChemEODEOL
+from prog_models.loading import Piecewise
 
-def future_loading(t, x=None):
-    # Variable (piece-wise) future loading scheme 
-    if (t < 600):
-        i = 2
-    elif (t < 900):
-        i = 1
-    elif (t < 1800):
-        i = 4
-    elif (t < 3000):
-        i = 2
-    else:
-        i = 3
-    return {'i': i}
+# Variable (piece-wise) future loading scheme 
+future_loading = Piecewise(
+    dict,
+    [600, 900, 1800, 3000, float('inf')],
+    {'i': [2, 1, 4, 2, 3]})
 
 
 class TestBattery(unittest.TestCase):
