### 2023-06-01 936e635
Apply suggestions from code review
Co-authored-by: Katy Jarvis <55932920+kjjarvis@users.noreply.github.com>
