### 2023-06-01 10096d8
Piecewise in events example
