### 2023-06-02 33d056a
Update src/prog_models/sim_result.py
Co-authored-by: Christopher Teubert <christopher.a.teubert@nasa.gov>
