### 2023-07-05 #5099 @kellyrowland
@teubert thanks! Reading through the paper, my only minor comments are that "Python" and "TensorFlow" should be capitalized in the text. Happy to open an issue on the repo if that's useful.
