### 2023-07-06 #5099 @kyleniemeyer
Hi @teubert, I'm doing some final checks before accepting.

It appears that the author/contributor list on Zenodo does not match the author list of the paper. These should usually match - is there a reason for this?

Also, I've made some typographic edits to the paper. Could you review and merge the PR? https://github.com/nasa/prog_models/pull/574

Lastly, the Goebel et al. (2017) reference seems to be missing some details. Is that a book, or something else?
