### 2023-07-06 #5099 @teubert
Thank you for the comments @kyleniemeyer

I just approved and merged your PR.

The author list for the paper are the individuals who contributed to the paper, which is a subset of the authors of the software itself. Should I add the other software authors to the paper?

Also, yes, the Goebel publication is a book. I'll see if I can find any other details (e.g., publisher) that I can add to the bibtex entry.
