### 2023-07-06 #5099 @teubert
I just added the publisher to the Goebel book- Looking through I didn't see any other information that could be added.
