### 2023-07-06 #5099 @kyleniemeyer
> The author list for the paper are the individuals who contributed to the paper, which is a subset of the authors of the software itself. Should I add the other software authors to the paper?
@teubert yes, the author lists should match - the paper is really meant as a stand-in for the software itself. Please add the missing authors to the paper.
